+++ joss_submission/paper.bib
@@ -102,7 +102,7 @@ volume = {336},
 pages = {120738},
 year = {2023},
 issn = {0306-2619},
-doi = {https://doi.org/10.1016/j.apenergy.2023.120738},
+doi = {10.1016/j.apenergy.2023.120738},
 url = {https://www.sciencedirect.com/science/article/pii/S0306261923001022},
 author = {Lukas Weimann and Guus Dubbink and Louis {van der Ham} and Matteo Gazzani},
 keywords = {Carbon capture, MEA, Energy system, Optimisation, MILP},
