### 2024-07-18 #6988 @JeanWi
Dear Kyle,
thanks for taking this up! Potentially suitable reviewers would be:
- Marta Victoria (Aarhus University, mvp@mpe.au.dk)
- Francesco Lombardi (Delft University, F.Lombardi@tudelft.nl)
- Stefan Pfenninger-Lee (Delft University, S.Pfenninger@tudelft.nl)
- Clara Büttner (Hochschule Flensburg, clara.buettner@hs-flensburg.de)
Best,
Jan
