### 2024-08-07 dda9815
bugfix with standby power and input ratios conv3
