+++ adopt_net0/components/technologies/genericTechnologies/conv3.py
@@ -196,7 +196,7 @@ class Conv3(Technology):
         standby_power = coeff_ti["standby_power"]
         phi = coeff_ti["phi"]
 
-        if standby_power == -1:
+        if self.component_options.performance_function_type == 1 or standby_power == -1:
 
             def init_input_input(const, t, car_input):
                 if car_input == self.component_options.main_input_carrier:
@@ -240,12 +240,15 @@ class Conv3(Technology):
                     dis.const_x_on = pyo.Constraint(expr=b_tec.var_x[t] == 1)
 
                     def init_input_on(const, car_input):
-                        if car_input == car_standby_power:
+                        if car_input == self.component_options.main_input_carrier:
                             return pyo.Constraint.Skip
                         else:
                             return (
                                 self.input[t, car_input]
-                                == phi[car_input] * self.input[t, car_standby_power]
+                                == phi[car_input]
+                                * self.input[
+                                    t, self.component_options.main_input_carrier
+                                ]
                             )
 
                     dis.const_input_on = pyo.Constraint(
