+++ adopt_net0/components/technologies/genericTechnologies/stor.py
@@ -962,7 +962,7 @@ class Stor(Technology):
                 if t > 1:
                     return -ramping_rate <= sum(
                         output_aux_rr[t, car_output] - output_aux_rr[t - 1, car_output]
-                        for car_output in b_tec.set_ouput_carriers
+                        for car_output in b_tec.set_output_carriers
                     )
                 else:
                     return pyo.Constraint.Skip
@@ -978,7 +978,7 @@ class Stor(Technology):
                         sum(
                             output_aux_rr[t, car_output]
                             - output_aux_rr[t - 1, car_output]
-                            for car_output in b_tec.set_ouput_carriers
+                            for car_output in b_tec.set_output_carriers
                         )
                         <= ramping_rate
                     )
