### 2024-08-07 6a7db67
extract h5 as dict
