### 2024-08-07 9d0cc03
change flexible energytopower ratio names in storage and add option for fixed capacity
