+++ adopt_net0/components/technologies/genericTechnologies/stor.py
@@ -144,11 +144,15 @@ class Stor(Technology):
 
         # For a flexibly optimized storage technology (i.e., not a fixed P-E ratio), an adapted CAPEX function is used
         # to account for charging and discharging capacity costs.
-        if self.flexibility_data["power_energy_ratio"] == "flex":
+        if self.flexibility_data["power_energy_ratio"] == "flexratio":
             self.economics.capex_model = 4
-        if self.flexibility_data["power_energy_ratio"] not in ["flex", "fixed"]:
+        if self.flexibility_data["power_energy_ratio"] not in [
+            "flexratio",
+            "fixedratio",
+            "fixedcapacity",
+        ]:
             raise Warning(
-                "power_energy_ratio should be either flexible ('flex') or fixed ('fixed')"
+                "power_energy_ratio should be either flexible ('flexratio') or fixed ('fixedratio') or as capacity ('fixedcapacity')"
             )
 
         # Coefficients
@@ -344,14 +348,21 @@ class Stor(Technology):
             # Cut according to Morales-Espana "LP Formulation for Optimal Investment and
             # Operation of Storage Including Reserves"
             def init_cut_bidirectional(const, t):
-                # output[t]/discharge_rate + input[t]/charge_rate <= storSize
-                return (
-                    self.output[t, self.component_options.main_input_carrier]
-                    / discharge_rate
-                    + self.input[t, self.component_options.main_input_carrier]
-                    / charge_rate
-                    <= b_tec.var_size
-                )
+                if self.flexibility_data["power_energy_ratio"] == "fixedcapacity":
+                    return (
+                        self.output[t, self.component_options.main_input_carrier]
+                        + self.input[t, self.component_options.main_input_carrier]
+                        <= b_tec.var_size
+                    )
+                else:
+                    # output[t]/discharge_rate + input[t]/charge_rate <= storSize
+                    return (
+                        self.output[t, self.component_options.main_input_carrier]
+                        / discharge_rate
+                        + self.input[t, self.component_options.main_input_carrier]
+                        / charge_rate
+                        <= b_tec.var_size
+                    )
 
             b_tec.const_cut_bidirectional = pyo.Constraint(
                 self.set_t_performance, rule=init_cut_bidirectional
@@ -395,43 +406,63 @@ class Stor(Technology):
 
         # Maximal charging and discharging rates
         def init_maximal_charge(const, t):
-            return (
-                self.input[t, self.component_options.main_input_carrier]
-                <= b_tec.var_capacity_charge
-            )
+            if self.flexibility_data["power_energy_ratio"] == "fixedcapacity":
+                return (
+                    self.input[t, self.component_options.main_input_carrier]
+                    <= charge_rate
+                )
+            else:
+                return (
+                    self.input[t, self.component_options.main_input_carrier]
+                    <= b_tec.var_capacity_charge
+                )
 
         b_tec.const_max_charge = pyo.Constraint(
             self.set_t_performance, rule=init_maximal_charge
         )
 
         def init_maximal_discharge(const, t):
-            return (
-                self.output[t, self.component_options.main_input_carrier]
-                <= b_tec.var_capacity_discharge
-            )
+            if self.flexibility_data["power_energy_ratio"] == "fixedcapacity":
+                return (
+                    self.output[t, self.component_options.main_input_carrier]
+                    <= discharge_rate
+                )
+            else:
+                return (
[…]
