### 2024-08-12 4f792be
Finished capex optimizations
