### 2024-08-15 4300704
Electrolyzer corrected for capex optimization
