+++ offshore_storage/postprocessing/process results.py
@@ -1,0 +1,187 @@
+import pandas as pd
+from pathlib import  Path
+import h5py
+
+dir = Path("//ad.geo.uu.nl/Users/StaffUsers/6574114/EhubResults/StorageOffshore"
+           "/CapexOptimization/Summary - Copy.xlsx")
+dir_processed = Path("//ad.geo.uu.nl/Users/StaffUsers/6574114/EhubResults/StorageOffshore"
+           "/CapexOptimization/Summary_processed.xlsx")
+
+all_technologies = {
+    "onshore Storage_Battery_CapexOptimization": ["onshore", "Battery"],
+    "onshore Storage_CAES_CapexOptimization": ["onshore", "CAES"],
+    "onshore Electrolyzer": ["onshore", "Electrolyzer"],
+    "offshore Storage_Battery_CapexOptimization": ["offshore", "Battery"],
+    "offshore Storage_CAES_CapexOptimization": ["offshore", "CAES"],
+    "offshore Storage_OceanBattery_CapexOptimization": ["offshore", "OceanBattery"],
+    "offshore Electrolyzer": ["offshore", "Electrolyzer"],
+}
+
+sizes = {
+    "Battery": 7,
+    "CAES": 3000,
+    "Electrolyzer": 1,
+    "OceanBattery": 7.5
+
+}
+
+def map_timestamp(timestamp, idx):
+    for key, value in all_technologies.items():
+        if key in timestamp:
+            return value[idx]
+    return None  # or some default value if no match is found
+
+def add_prefix_to_keys(dictionary, prefix):
+    new_dict = {}
+    for key, value in dictionary.items():
+        new_key = prefix + key
+        new_dict[new_key] = value
+    return new_dict
+
+
+df = pd.read_excel(dir)
+
+df['Node'] = df['time_stamp'].apply(lambda x: map_timestamp(x, 0))
+df['Technology'] = df['time_stamp'].apply(lambda x: map_timestamp(x, 1))
+df['SS'] = df['case'].str.extract(r'selfsufficiency_(\d+\.\d+)')
+df['OS'] = df['case'].str.extract(r'offshore_(\d+\.\d+)')
+df['SS'].fillna(1, inplace=True)
+df['OS'].fillna(1, inplace=True)
+df['max_specific_capex'] = df.apply(lambda row: row['cost_capex_tecs'] / sizes[row['Technology']], axis=1)
+
+
+
+
+#
+#
+#
+#
+#
+# # IMPORTS AND CURTAILMENT
+# max_re = pd.read_csv('C:/Users/6574114/PycharmProjects/PyHubProductive/mes_north_sea/clean_data/production_profiles_re/production_profiles_re.csv', index_col=0, header=[0,1])
+# max_re = max_re.loc[:, (slice(None), 'total')].sum()
+# max_re.reset_index(level=1, drop=True, inplace=True)
+#
+# # Get nodes and carriers
+# with h5py.File(
+#     df.loc[df['Case'] == 'Baseline', 'time_stamp'].values[0] + '/optimization_results.h5',
+#     'r') as hdf_file:
+#     nodes = extract_dataset_from_h5(hdf_file["topology/nodes"])
+#     carriers = extract_dataset_from_h5(hdf_file["topology/carriers"])
+#
+#
+# # Calculate Imports and Curtailment
+# imports_dict = {}
+# export_dict = {}
+# curtailment_dict = {}
+# generic_production_dict = {}
+# tec_output_dict = {}
+# demand_dict = {}
+# netw_dict = {}
+# tec_dict = {}
+# for idx, row in df.iterrows():
+#     case_path = df.loc[idx, "time_stamp"]
+#
+#     print(case_path)
+#     # Read data
[…]
