+++ main_offshore_storage_emission_optim.py
@@ -1,0 +1,344 @@
+import adopt_net0 as adopt
+from adopt_net0.case_offshore_storage.modelhub import ModelHubCapexOptimization, ModelHubEmissionOptimization
+import pandas as pd
+import numpy as np
+from pathlib import Path
+import json
+import pyomo.environ as pyo
+
+"""
+TODOS
+- Add loop for technologies
+- Add all technology datasheets
+"""
+
+#setup model (only required once)
+input_data_path = Path("./offshore_storage/model_input_emission_optim")
+# input_data_path = Path("./offshore_storage/model_input_test")
+# adopt.create_optimization_templates(input_data_path)
+# adopt.create_input_data_folder_template(input_data_path)
+# adopt.copy_technology_data(input_data_path)
+# adopt.copy_network_data(input_data_path)
+
+test = 1
+test_periods = 100
+climate_year = 2000
+# all_technologies = [
+#     ('offshore', "Storage_OceanBattery_CapexOptimization")
+# ]
+emission_target = 1
+all_technologies = [
+    ('onshore', "Storage_Battery_CapexOptimization"),
+    ('onshore', "Storage_CAES_CapexOptimization"),
+    ('onshore', "Electrolyzer"),
+    ('offshore', "Storage_Battery_CapexOptimization"),
+    ('offshore', "Storage_CAES_CapexOptimization"),
+    ('offshore', "Storage_OceanBattery_CapexOptimization"),
+    ('offshore', "Electrolyzer"),
+]
+# Write generic production
+def determine_time_series(f_demand, f_offshore, f_self_sufficiency, cy):
+    cap_pv = 215002
+    cap_wind_onshore = 115000
+    s_pv = cap_pv / (cap_pv + cap_wind_onshore)
+    s_wind = cap_wind_onshore / (cap_pv + cap_wind_onshore)
+
+    cf_pv = pd.read_csv("./offshore_storage/data/capacity_factors/DE_Solar2030.csv",
+                        sep=",")
+    cf_wind_onshore = pd.read_csv(
+        "./offshore_storage/data/capacity_factors/DE_Wind_onshore2030.csv",
+        sep=",")
+    cf_wind_offshore = pd.read_csv(
+        "./offshore_storage/data/capacity_factors/DE_Wind_offshore2030.csv",
+        sep=",")
+
+
+    demand = pd.read_csv("./offshore_storage/data/demand/DE_Demand2030.csv", sep=";")
+    demand_cy = demand[str(climate_year)] * f_demand
+    cf_pv = cf_pv[str(climate_year)]
+    cf_wind_onshore = cf_wind_onshore[str(climate_year)]
+    cf_wind_offshore = cf_wind_offshore[str(climate_year)]
+
+    annual_demand = sum(demand_cy)
+
+    e_offshore = sum(cf_wind_offshore)
+    e_onshore = sum(cf_wind_onshore) * s_wind + sum(cf_pv) * s_pv
+
+    # capacity required for 1MWh annual generation onshore/offshore
+    c_offshore = 1 / e_offshore * annual_demand * f_offshore * f_self_sufficiency
+    c_onshore = 1 / e_onshore * annual_demand * (1 - f_offshore) * f_self_sufficiency
+
+    # generation profiles
+    p_offshore = c_offshore * cf_wind_offshore
+    p_onshore = c_onshore * (cf_wind_onshore * s_wind + cf_pv * s_pv)
+
+    return demand_cy, p_onshore, p_offshore
+
+def set_data(climate_year, technology, f_demand, f_offshore,
+                               f_self_sufficiency,
+                               test):
+    gas_price = 26.78  # ERAA 2023
+    co2_price = 113  # ERAA 2023
+    hydrogen_price = gas_price + co2_price * 0.18
+
[…]
