+++ adopt_net0/case_offshore_storage/modelhub.py
@@ -53,27 +53,22 @@ class ModelHubEmissionOptimization(ModelHub):
 
         self._delete_objective()
 
-        if self.total_emission_limit:
-            try:
-                model.del_component(model.const_emission_limit)
-            except:
-                pass
+        try:
+            model.del_component(model.const_emission_limit)
+        except:
+            pass
 
-            model.const_emission_limit = Constraint(expr=model.var_emissions_net <=
-                                                     self.total_emission_limit * 1.001)
+        model.const_emission_limit = Constraint(expr=model.var_emissions_net <=
+                                                 self.total_emission_limit * 1.001)
 
-            model.const_emission_limit.pprint()
-            def init_min_size(obj):
-                return model.periods["period1"].node_blocks[self.technology_to_optimize[
-                    0]].tech_blocks_active[self.technology_to_optimize[1]].var_size
-            model.objective = Objective(rule=init_min_size, sense=minimize)
-            if config["solveroptions"]["solver"]["value"] == "gurobi_persistent":
-                self.solver.add_constraint(model.const_emission_limit)
-                self.solver.set_objective(model.objective)
-        else:
-            def init_emission_objective(obj):
-                return model.var_emissions_net
-            model.objective = Objective(rule=init_emission_objective, sense=minimize)
+        model.const_emission_limit.pprint()
+        def init_min_size(obj):
+            return model.periods["period1"].node_blocks[self.technology_to_optimize[
+                0]].tech_blocks_active[self.technology_to_optimize[1]].var_size
+        model.objective = Objective(rule=init_min_size, sense=minimize)
+        if config["solveroptions"]["solver"]["value"] == "gurobi_persistent":
+            self.solver.add_constraint(model.const_emission_limit)
+            self.solver.set_objective(model.objective)
 
         self._call_solver()
 
