### 2024-08-16 49eeaba
basline costs change after every run
