+++ main_offshore_storage_capex_optim.py
@@ -28,12 +28,12 @@ climate_year = 2000
 # ]
 
 all_technologies = [
-    ('onshore', "Storage_Battery_CapexOptimization"),
-    ('onshore', "Storage_CAES_CapexOptimization"),
+    # ('onshore', "Storage_Battery_CapexOptimization"),
+    # ('onshore', "Storage_CAES_CapexOptimization"),
     ('onshore', "Electrolyzer"),
-    ('offshore', "Storage_Battery_CapexOptimization"),
-    ('offshore', "Storage_CAES_CapexOptimization"),
-    ('offshore', "Storage_OceanBattery_CapexOptimization"),
+    # ('offshore', "Storage_Battery_CapexOptimization"),
+    # ('offshore', "Storage_CAES_CapexOptimization"),
+    # ('offshore', "Storage_OceanBattery_CapexOptimization"),
     ('offshore', "Electrolyzer"),
 ]
 # Write generic production
