+++ main_offshore_storage_emission_optim.py
@@ -21,7 +21,7 @@ input_data_path = Path("./offshore_storage/model_input_emission_optim")
 # adopt.copy_network_data(input_data_path)
 
 test = 1
-test_periods = 100
+test_periods = 10
 climate_year = 2000
 # all_technologies = [
 #     ('offshore', "Storage_OceanBattery_CapexOptimization")
@@ -29,12 +29,12 @@ climate_year = 2000
 emission_targets = [0.9, 0.8, 0.7, 0.6, 0.5, 0.4, 0.3, 0.2, 0.1, 0.0]
 all_technologies = [
     ('onshore', "Storage_Battery_CapexOptimization"),
-    # ('onshore', "Storage_CAES_CapexOptimization"),
-    # ('onshore', "Electrolyzer"),
-    # ('offshore', "Storage_Battery_CapexOptimization"),
-    # ('offshore', "Storage_CAES_CapexOptimization"),
-    # ('offshore', "Storage_OceanBattery_CapexOptimization"),
-    # ('offshore', "Electrolyzer"),
+    ('onshore', "Storage_CAES_CapexOptimization"),
+    ('onshore', "Electrolyzer"),
+    ('offshore', "Storage_Battery_CapexOptimization"),
+    ('offshore', "Storage_CAES_CapexOptimization"),
+    ('offshore', "Storage_OceanBattery_CapexOptimization"),
+    ('offshore', "Electrolyzer"),
 ]
 # Write generic production
 def determine_time_series(f_demand, f_offshore, f_self_sufficiency, cy):
@@ -295,6 +295,10 @@ for technology in all_technologies:
                 m_baseline.data.model_config["reporting"]["case_name"]["value"] = (
                         "baseline " + case_name)
                 m_baseline.solve()
+
+                set_data(climate_year, technology, factors['demand'], f_offshore,
+                         f_self_sufficiency, test)
+
                 total_emissions = m_baseline.model[m_baseline.info_solving_algorithms[
                     "aggregation_model"]].var_emissions_net.value
 
