+++ offshore_storage/postprocessing/process results.py
@@ -1,6 +1,7 @@
 import pandas as pd
 from pathlib import  Path
 import h5py
+from adopt_net0 import extract_datasets_from_h5group, extract_dataset_from_h5
 
 dir = Path("//ad.geo.uu.nl/Users/StaffUsers/6574114/EhubResults/StorageOffshore"
            "/CapexOptimization/Summary - Copy.xlsx")
@@ -8,13 +9,13 @@ dir_processed = Path("//ad.geo.uu.nl/Users/StaffUsers/6574114/EhubResults/Storag
            "/CapexOptimization/Summary_processed.xlsx")
 
 all_technologies = {
-    "onshore Storage_Battery_CapexOptimization": ["onshore", "Battery"],
-    "onshore Storage_CAES_CapexOptimization": ["onshore", "CAES"],
-    "onshore Electrolyzer": ["onshore", "Electrolyzer"],
-    "offshore Storage_Battery_CapexOptimization": ["offshore", "Battery"],
-    "offshore Storage_CAES_CapexOptimization": ["offshore", "CAES"],
-    "offshore Storage_OceanBattery_CapexOptimization": ["offshore", "OceanBattery"],
-    "offshore Electrolyzer": ["offshore", "Electrolyzer"],
+    "onshore_Storage_Battery_CapexOptimization": ["onshore", "Battery"],
+    "onshore_Storage_CAES_CapexOptimization": ["onshore", "CAES"],
+    "onshore_Electrolyzer": ["onshore", "Electrolyzer"],
+    "offshore_Storage_Battery_CapexOptimization": ["offshore", "Battery"],
+    "offshore_Storage_CAES_CapexOptimization": ["offshore", "CAES"],
+    "offshore_Storage_OceanBattery_CapexOptimization": ["offshore", "OceanBattery"],
+    "offshore_Electrolyzer": ["offshore", "Electrolyzer"],
 }
 
 sizes = {
@@ -43,25 +44,17 @@ df = pd.read_excel(dir)
 
 df['Node'] = df['time_stamp'].apply(lambda x: map_timestamp(x, 0))
 df['Technology'] = df['time_stamp'].apply(lambda x: map_timestamp(x, 1))
-df['SS'] = df['case'].str.extract(r'selfsufficiency_(\d+\.\d+)')
-df['OS'] = df['case'].str.extract(r'offshore_(\d+\.\d+)')
-df['SS'].fillna(1, inplace=True)
-df['OS'].fillna(1, inplace=True)
+df['SS'] = df['case'].str.extract(r'SS_(\d+\.?\d*)').astype(float)
+df['OS'] = df['case'].str.extract(r'OS_(\d+\.?\d*)').astype(float)
+df['Baseline'] = df['case'].str.contains('BL|baseline')
 df['max_specific_capex'] = df.apply(lambda row: row['cost_capex_tecs'] / sizes[row['Technology']], axis=1)
 
 
-
-
-#
-#
-#
-#
-#
-# # IMPORTS AND CURTAILMENT
+# IMPORTS AND CURTAILMENT
 # max_re = pd.read_csv('C:/Users/6574114/PycharmProjects/PyHubProductive/mes_north_sea/clean_data/production_profiles_re/production_profiles_re.csv', index_col=0, header=[0,1])
 # max_re = max_re.loc[:, (slice(None), 'total')].sum()
 # max_re.reset_index(level=1, drop=True, inplace=True)
-#
+
 # # Get nodes and carriers
 # with h5py.File(
 #     df.loc[df['Case'] == 'Baseline', 'time_stamp'].values[0] + '/optimization_results.h5',
[…]
