### 2024-08-21 4ca91b5
removed capex_optimization for emission in technology jsons
