### 2024-08-21 0a135ff
Updated network adaption
