### 2024-08-23 0fe9ddb
Changed technology sizes
