### 2024-08-23 cceff65
Strated coding CCPP from Siemens work
