+++ adopt_net0/components/technologies/specificTechnologies/combined_cycle.py
@@ -1,0 +1,746 @@
+import pyomo.environ as pyo
+import pyomo.gdp as gdp
+import copy
+import numpy as np
+import pandas as pd
+from pathlib import Path
+from scipy.interpolate import griddata
+
+from ..technology import Technology
+from ...utilities import link_full_resolution_to_clustered
+from ..utilities import fit_piecewise_function
+
+
+import logging
+
+log = logging.getLogger(__name__)
+
+
+class CCPP(Technology):
+    """
+    Combined Cycle Power Plant with Steam Production
+
+    Resembles Combined Cycle Power Plant with Steam Production.
+    The size of the power plant it fixed and it only possible to model the plant as
+    an existing technology (i.e. it cannot be sized). The model is based on Wiegner
+    et al. (2024). Optimizing the Use of Limited Amounts of Hydrogen in Existing
+    Combined Heat and Power Plants.
+
+    It is possible in the JSON file to specify the following options:
+
+    - Size of an Oxy-fuel hydrogen burner
+
+    - Size of a duct burner that can burn hydrogen or natural gas
+
+    - Max share (energy-based) of hydrogen combustion in the gas turbine ("max_input")
+
+    - How to treat steam production ("steam_production"):
+
+        - "ignore": technology has no additional output other than electricity
+
+        - "as heat": technology has heat output, equivalent to the sum of HP/MP
+          generation
+
+        - "as steam": technology has steam output, equivalent to the sum of HP/MP
+          generation
+
+        - "as hp/mp steam": technology has HP and MP steam output
+
+    - What additional components in the plant should be considered. Possible options
+      are:
+
+        - "None": only CCPP
+
+        - "DB": Duct burner
+
+        - "OHB": Oxyfuel-hydrogen burner
+
+    - Number of segments to use for performance function ("nr_segments").
+      Recommended: 2
+
+    **Variable declarations:**
+
+    - Total fuel input in :math:`t`: :math:`Input_{tot, t}`
+
+    - Fuel input to duct burner/ OHB :math:`Input_{OHB/DB, t}`
+
+    **Constraint declarations:**
+
+    The following constants are used:
+
+    - :math:`Input_{min}`: Minimal input per turbine
+
+    - :math:`Input_{max}`: Maximal input per turbine
+
+    - :math:`in_{H2max}`: Maximal H2 admixture to fuel (only for natural gas turbines, default is 0.05)
+
+    - :math:`{\\alpha}`: Performance parameter for electricity output
+
+    - :math:`{\\beta}`: Performance parameter for electricity output
+
+    - :math:`{\\epsilon}`: Performance parameter for heat output
+
+    - :math:`f({\\Theta})`: Ambient temperature correction factor
+
+    - Input calculation (For hydrogen turbines, :math:`Input_{NG, t}` is zero, and the second constraint is removed):
+
+      .. math::
+        Input_{H2, t} + Input_{NG, t} = Input_{tot, t}
+
+      .. math::
+        Input_{H2, t} \leq in_{H2max} Input_{tot, t}
+
+    - Turbines on:
+
+      .. math::
+        N_{on, t} \leq S
+
+    - If technology is on:
+
+      .. math::
+        Output_{el,t} = ({\\alpha} Input_{tot, t} + {\\beta} * N_{on, t}) *f({\\Theta})
+
+      .. math::
+        Output_{th,t} = {\\epsilon} Input_{tot, t} - Output_{el,t}
+
+      .. math::
+        Input_{min} * N_{on, t} \leq Input_{tot, t} \leq Input_{max} * N_{on, t}
+
+    - If the technology is off, input and output is set to 0:
+
+      .. math::
[…]
