+++ adopt_net0/components/technologies/specificTechnologies/__init__.py
@@ -2,3 +2,4 @@ from .dac_adsorption import DacAdsorption
 from .gas_turbine import GasTurbine
 from .heat_pump import HeatPump
 from .hydro_open import HydroOpen
+from .combined_cycle import CCPP
