### 2024-08-28 b9c44a9
Try upgrading setuptools
