+++ .github/workflows/testing.yml
@@ -27,6 +27,7 @@ jobs:
     - name: Install Dependencies
       run: |
         python -m pip install --upgrade pip
+        pip install --upgrade setuptools
         pip install coverage
         pip install -r requirements.txt
 
