### 2024-08-28 d9b9b9c
Try pinning setuptools
