### 2024-08-28 eec99c2
Try pinning setuptools in requirements.txt
