+++ requirements.txt
@@ -7,7 +7,7 @@ timezonefinder>=6.5.0
 statsmodels>=0.14.1
 pvlib>=0.10.4
 scipy>=1.13.0
-setuptools>=69.2.0
+setuptools==72.2.0
 matplotlib>=3.8.4
 openpyxl>=3.1.2
 pytest>=8.1.1
