+++ .github/workflows/testing.yml
@@ -27,7 +27,6 @@ jobs:
     - name: Install Dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install setuptools==72.2.0
         pip install coverage
         pip install -r requirements.txt
 
