+++ .github/workflows/testing.yml
@@ -27,8 +27,10 @@ jobs:
     - name: Install Dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install coverage
+        python -m venv venv
+        source venv/bin/activate
         pip install -r requirements.txt
+        pip install coverage
 
     - name: Check directory
       run: |
