+++ .github/workflows/testing.yml
@@ -26,7 +26,7 @@ jobs:
 
     - name: Install Dependencies
       env:
-        PIP_CONSTRAINT: ./constraints.txt
+        PIP_CONSTRAINT: .github/workflows/constraint.txt
       run: |
         python -m pip install --upgrade pip
         pip install coverage
