+++ .github/workflows/testing.yml
@@ -26,7 +26,7 @@ jobs:
 
     - name: Install Dependencies
       env:
-        PIP_CONSTRAINT: .github/workflows/constraint.txt
+        PIP_CONSTRAINT: .github/workflows/constraints.txt
       run: |
         python -m pip install --upgrade pip
         pip install coverage
