+++ adopt_net0/components/technologies/specificTechnologies/combined_cycle.py
@@ -200,15 +200,24 @@ class CCPP(Technology):
         )
         data_path["GT"] = Path("GT_fitting_data.csv")
         data_path["HP"] = Path("HP_fitting_data.csv")
-        # data_path["MP"] = Path("MP_fitting_data.csv")
+        data_path["MP"] = Path("MP_fitting_data.csv")
 
         # Fit GT performance
         perf_data = {}
         perf_data["GT"] = pd.read_csv(performance_data_path / data_path["GT"], sep=";")
-        perf_data["HP"] = pd.read_csv(
-            performance_data_path / data_path["HP"], sep=";", index_col=0, header=[0, 1]
-        )
-        # perf_data["MP"] = pd.read_csv(performance_data_path / data_path["GT"], sep=";")
+        for turbine in ["HP", "MP"]:
+            perf_data[turbine] = pd.read_csv(
+                performance_data_path / data_path[turbine],
+                sep=";",
+                index_col=0,
+                header=[0, 1],
+            )
+            perf_data[turbine] = pd.read_csv(
+                performance_data_path / data_path[turbine],
+                sep=";",
+                index_col=0,
+                header=[0, 1],
+            )
 
         # Fit to temperature
         igv_positions = list(perf_data["GT"]["IGV"].unique())
@@ -269,89 +278,85 @@ class CCPP(Technology):
         self.processed_coeff.time_dependent_full["GT"]["bp_el_y"] = bp_el_y
         self.processed_coeff.time_dependent_full["GT"]["alpha_th"] = alpha_th
 
-        # Fit HP performance
-        bp_hp = [0, 12.92, 100]
+        # Fit HP/MP performance
+        bp = {}
+        bp["HP"] = [0, 12.92, 100]
+        bp["MP"] = [0, 42.50, 100]
+        for p in ["HP", "MP"]:
 
-        alpha_gt = np.empty(shape=(len(T), nr_segments))
-        alpha_hp = np.empty(shape=(len(T), nr_segments))
-        alpha_mp = np.empty(shape=(len(T), nr_segments))
-        alpha_cst = np.empty(shape=(len(T), nr_segments))
-        if self.component_options.other["component"] == "DB":
-            alpha_db = np.empty(shape=(len(T), nr_segments))
-        elif self.component_options.other["component"] == "OHB":
-            alpha_ohb = np.empty(shape=(len(T), nr_segments))
-
-        for par in [1, 2]:
-            data = perf_data["HP"]["GT"]["alpha_" + str(par)]
-            alpha_gt[:, par - 1] = griddata(
-                np.array(data.index), np.array(data), T, method="linear"
-            )
-
-            data = perf_data["HP"]["HP"]["alpha_" + str(par)]
-            alpha_hp[:, par - 1] = griddata(
-                np.array(data.index), np.array(data), T, method="linear"
-            )
-
-            data = perf_data["HP"]["MP"]["alpha_" + str(par)]
-            alpha_mp[:, par - 1] = griddata(
-                np.array(data.index), np.array(data), T, method="linear"
-            )
-
-            data = perf_data["HP"]["cst"]["alpha_" + str(par)]
-            alpha_cst[:, par - 1] = griddata(
-                np.array(data.index), np.array(data), T, method="linear"
-            )
-
+            alpha_gt = np.empty(shape=(len(T), nr_segments))
+            alpha_hp = np.empty(shape=(len(T), nr_segments))
+            alpha_mp = np.empty(shape=(len(T), nr_segments))
+            alpha_cst = np.empty(shape=(len(T), nr_segments))
             if self.component_options.other["component"] == "DB":
-                data = perf_data["HP"]["DB"]["alpha_" + str(par)]
[…]
