+++ main_testCCPP.py
@@ -10,6 +10,7 @@ import pandas as pd
 path = Path("C:/Users/6574114/PycharmProjects/adopt_net0/test_ccpp")
 
 steam_demand = 0
+el_demand = 150
 
 # Create template files (comment these lines if already defined)
 # dp.create_optimization_templates(path)
@@ -33,6 +34,10 @@ if steam_demand:
     d = input_data["MP"]
 dp.fill_carrier_data(path, d, columns=["Demand"], carriers=["steam_mp"])
 
+# Electricity Demand
+if el_demand:
+    dp.fill_carrier_data(path, el_demand, columns=["Demand"], carriers=["electricity"])
+
 # Electricity Price & Export
 dp.fill_carrier_data(
     path, input_data["P2021"], columns=["Export price"], carriers=["electricity"]
@@ -42,6 +47,9 @@ dp.fill_carrier_data(path, 500, columns=["Export limit"], carriers=["electricity
 # Gas import
 dp.fill_carrier_data(path, 50, columns=["Import price"], carriers=["gas"])
 dp.fill_carrier_data(path, 1000, columns=["Import limit"], carriers=["gas"])
+# H2 import
+dp.fill_carrier_data(path, 40, columns=["Import price"], carriers=["hydrogen"])
+dp.fill_carrier_data(path, 1000, columns=["Import limit"], carriers=["hydrogen"])
 
 # CO2 price
 co2_price = pd.read_csv(
@@ -53,7 +61,7 @@ co2_price.to_csv(path / "period1/node_data/node1/CarbonCost.csv", sep=";")
 # Technologies
 with open(path / "period1/node_data/node1/Technologies.json", "r") as f:
     tecs = json.load(f)
-tecs["existing"] = {"CombinedCycle_fixed_size": 0}
+tecs["existing"] = {"CombinedCycle_fixed_size": 1}
 with open(path / "period1/node_data/node1/Technologies.json", "w") as f:
     json.dump(tecs, f)
 
@@ -62,13 +70,14 @@ dp.copy_technology_data(path)
 # dp.copy_network_data(path, "path to network data")
 #
 # # Read climate data and fill carried data (comment these lines if already defined)
-# dp.load_climate_data_from_api(path)
+dp.load_climate_data_from_api(path)
 # dp.fill_carrier_data(path, value=0)
 #
 # # Construct and solve the model
 pyhub = ModelHub()
 pyhub.read_data(path, start_period=0, end_period=2)
 pyhub.quick_solve()
+
 #
 # # Add values of (part of) the parameters and variables to the summary file
 # add_values_to_summary(Path("path to summary file"))
