+++ adopt_net0/components/technologies/specificTechnologies/combined_cycle.py
@@ -329,8 +329,9 @@ class CCPP(Technology):
             self.processed_coeff.time_independent[p] = {}
             self.processed_coeff.time_independent[p]["alpha_gt"] = alpha_gt
             self.processed_coeff.time_independent[p]["alpha_hp"] = alpha_hp
+            self.processed_coeff.time_independent[p]["alpha_mp"] = alpha_mp
             self.processed_coeff.time_independent[p]["alpha_cst"] = alpha_cst
-            self.processed_coeff.time_independent[p]["bp_hp"] = bp[p]
+            self.processed_coeff.time_independent[p]["bp_" + p.lower()] = bp[p]
             if self.component_options.other["component"] == "DB":
                 self.processed_coeff.time_independent[p]["alpha_db"] = alpha_db
             elif self.component_options.other["component"] == "OHB":
@@ -475,9 +476,14 @@ class CCPP(Technology):
         )
 
         # GT - P thermal
+        # TODO: calculate bounds
+        def init_gt_p_th_bounds(bds, t):
+            return tuple((0, 1000))
+
         b_tec.var_gt_p_th = pyo.Var(
             self.set_t_performance,
             within=pyo.NonNegativeReals,
+            bounds=init_gt_p_th_bounds,
         )
 
         # Duct Burner - input
@@ -548,6 +554,44 @@ class CCPP(Technology):
             bounds=init_mp_output_bounds,
         )
 
+        # Steam Output
+        # TODO: Calculate bounds
+        def init_mp_hp_bounds(bds, t):
+            return tuple((0, 1000))
+
+        b_tec.var_mp_p_hp = pyo.Var(
+            self.set_t_performance,
+            within=pyo.NonNegativeReals,
+            bounds=init_mp_hp_bounds,
+        )
+
+        def init_mp_mp_bounds(bds, t):
+            return tuple((0, 1000))
+
+        b_tec.var_mp_p_mp = pyo.Var(
+            self.set_t_performance,
+            within=pyo.NonNegativeReals,
+            bounds=init_mp_mp_bounds,
+        )
+
+        def init_hp_hp_bounds(bds, t):
+            return tuple((0, 1000))
+
+        b_tec.var_hp_p_hp = pyo.Var(
+            self.set_t_performance,
+            within=pyo.NonNegativeReals,
+            bounds=init_hp_hp_bounds,
+        )
+
+        def init_hp_mp_bounds(bds, t):
+            return tuple((0, 1000))
+
+        b_tec.var_hp_p_mp = pyo.Var(
+            self.set_t_performance,
+            within=pyo.NonNegativeReals,
+            bounds=init_hp_mp_bounds,
+        )
+
         if not b_tec.find_component("var_x"):
             b_tec.var_x = pyo.Var(
                 self.set_t_performance, domain=pyo.NonNegativeIntegers, bounds=(0, 1)
@@ -613,11 +657,11 @@ class CCPP(Technology):
     def _define_performance(self, b_tec):
         coeff_td = self.processed_coeff.time_dependent_used
         coeff_ti = self.processed_coeff.time_independent
-        alpha_th = coeff_td["GT"]["alpha_th"]
-        alpha = coeff_td["GT"]["alpha_el"]
-        beta = coeff_td["GT"]["beta_el"]
-        bp_x = coeff_td["GT"]["bp_el_x"]
         nr_segments = self.component_options.other["nr_segments"]
+        gt_alpha_th = coeff_td["GT"]["alpha_th"]
+        gt_alpha = coeff_td["GT"]["alpha_el"]
+        gt_beta = coeff_td["GT"]["beta_el"]
+        gt_bp_x = coeff_td["GT"]["bp_el_x"]
 
         # Total input to GT
[…]
