+++ adopt_net0/components/technologies/specificTechnologies/combined_cycle.py
@@ -1,3 +1,5 @@
+import warnings
+
 import pyomo.environ as pyo
 import pyomo.gdp as gdp
 import copy
@@ -24,15 +26,19 @@ class CCPP(Technology):
     The size of the power plant it fixed and it only possible to model the plant as
     an existing technology (i.e. it cannot be sized). The model is based on Wiegner
     et al. (2024). Optimizing the Use of Limited Amounts of Hydrogen in Existing
-    Combined Heat and Power Plants.
+    Combined Heat and Power Plants
+    https://www.sciencedirect.com/science/article/pii/S2667095X24000199.
 
     It is possible in the JSON file to specify the following options:
 
-    - Size of an Oxy-fuel hydrogen burner
+    - size_ohb: Size of an Oxy-fuel hydrogen burner
 
-    - Size of a duct burner that can burn hydrogen or natural gas
+    - size_db: Size of a duct burner that can burn hydrogen or natural gas
 
-    - Max share (energy-based) of hydrogen combustion in the gas turbine ("max_input")
+    - max_h2_in_gt: Max share (energy-based) of hydrogen combustion in the gas
+      turbine ("max_input")
+
+    - max_h2_in: Maximum of hydrogen input to plant
 
     - How to treat steam production ("steam_production"):
 
@@ -173,6 +179,14 @@ class CCPP(Technology):
             "nr_segments"
         ]
 
+        # All other options
+        self.component_options.other["steam_turbine_generator_efficiency"] = tec_data[
+            "Performance"
+        ]["steam_turbine_generator_efficiency"]
+        self.component_options.other["max_steam_extract_HP"] = 140
+        self.component_options.other["max_steam_extract_MP"] = 51
+        self.component_options.other["kappa_steam"] = 0.45
+
     def fit_technology_performance(self, climate_data: pd.DataFrame, location: dict):
         """
         Performs fitting for technology type CCPP
@@ -184,7 +198,6 @@ class CCPP(Technology):
         super(CCPP, self).fit_technology_performance(climate_data, location)
 
         # Climate data & Number of timesteps
-        time_steps = len(climate_data)
         T = copy.deepcopy(climate_data["temp_air"])
 
         # Remove outliers
@@ -193,11 +206,22 @@ class CCPP(Technology):
 
         # Determine correct reading paths
         data_path = {}
-        performance_data_path = Path(__file__).parent.parent.parent.parent
-        performance_data_path = (
-            performance_data_path
-            / "data/technology_data/PowerGeneration/CombinedCycle_fixed_size_data"
-        )
+        if "performance_data_path" in self.input_parameters.performance_data:
+            performance_data_path = self.input_parameters.performance_data[
+                "performance_data_path"
+            ]
+        else:
+            performance_data_path = Path(__file__).parent.parent.parent.parent
+            performance_data_path = (
+                performance_data_path
+                / "data/technology_data/PowerGeneration/CombinedCycle_fixed_size_data"
+            )
+            log.warning(
+                "Using performance data with noice (not the same as in "
+                "publication. Refer to the paper authors (Jan Wiegner) to get "
+                "the actual data."
+            )
+
         data_path["GT"] = Path("GT_fitting_data.csv")
         data_path["HP"] = Path("HP_fitting_data.csv")
         data_path["MP"] = Path("MP_fitting_data.csv")
@@ -205,6 +229,7 @@ class CCPP(Technology):
         # Fit GT performance
         perf_data = {}
[…]
