### 2024-09-12 4d1ad76
initialization of pyo.Parameter with dict keys not allowed, changed to list
