### 2024-09-27 #6988 @AdamRJensen
> Dear Kyle, thanks for taking this up! Potentially suitable reviewers would be:
>
> * Marta Victoria (Aarhus University, [mvp@mpe.au.dk](mailto:mvp@mpe.au.dk))
> * Francesco Lombardi (Delft University, [F.Lombardi@tudelft.nl](mailto:F.Lombardi@tudelft.nl))
> * Stefan Pfenninger-Lee (Delft University, [S.Pfenninger@tudelft.nl](mailto:S.Pfenninger@tudelft.nl))
> * Clara Büttner (Hochschule Flensburg, [clara.buettner@hs-flensburg.de](mailto:clara.buettner@hs-flensburg.de))

@JeanWi do you have the GitHub usernames for these people? If so, please write them here but leave out the "@" symbol so they don't get tagged just quite yet.
