### 2024-09-28 #6988 @JeanWi
Hi all,
Thanks Adam for editing this!
The GitHub user names are:
- martavp
- FLomb
- sjpfenninger
- ClaraBuettner
