### 2024-09-28 #6988 @AdamRJensen
:wave: @martavp, @ClaraBuettner, & @sjpfenninger, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Personally, I volunteer as an editor for JOSS because I think JOSS does a fantastic job at pushing for replicable open-source science!
