### 2024-09-30 #6988 @martavp
Thanks for the invitation @AdamRJensen and I'm sorry I cannot accept it. I don't have the bandwidth to review this during the next couple of months, but I do agree JOSS does a very good job and I'll be happy to support it as a reviewer in the future.
