### 2024-10-04 010e737
Minor bugfixe: result writing failed for scaled models if solution is infeasible (expected behavior: continue without writing results, actual problematic behavior: error when writing results)
