### 2024-10-04 f33d853
Add possibility to skip writing results
