### 2024-10-04 6a99102
Added escape characters to resolve SyntaxWarnings for Docstrings
