+++ adopt_net0/components/technologies/genericTechnologies/stor.py
@@ -47,15 +47,15 @@ class Stor(Technology):
     - Size constraint:
 
       .. math::
-        E_{t} \leq S
+        E_{t} \\leq S
 
     - Maximal charging and discharging:
 
       .. math::
-        Input_{t} \leq Input_{max}
+        Input_{t} \\leq Input_{max}
 
       .. math::
-        Output_{t} \leq Output_{max}
+        Output_{t} \\leq Output_{max}
 
     - Storage level calculation:
 
@@ -76,7 +76,7 @@ class Stor(Technology):
        Thus:
 
        .. math::
-         Input_{max} = \gamma_{charging} * S
+         Input_{max} = \\gamma_{charging} * S
 
     - If in 'Flexibility' the "power_energy_ratio == flexratio", then the
       capacity of the charging and discharging power is a variable in the
@@ -119,10 +119,10 @@ class Stor(Technology):
       output).
 
       .. math::
-         -rampingrate \leq Input_{t, maincar} - Input_{t-1, maincar} \leq rampingrate
+         -rampingrate \\leq Input_{t, maincar} - Input_{t-1, maincar} \\leq rampingrate
 
       .. math::
-         -rampingrate \leq \sum(Output_{t, car}) - \sum(Output_{t-1, car}) \leq
+         -rampingrate \\leq \sum(Output_{t, car}) - \sum(Output_{t-1, car}) \\leq
          rampingrate
 
 
