### 2024-10-04 2ebc173
Added escape characters to resolve SyntaxWarnings for Docstrings
