+++ adopt_net0/components/technologies/genericTechnologies/stor.py
@@ -122,7 +122,7 @@ class Stor(Technology):
          -rampingrate \\leq Input_{t, maincar} - Input_{t-1, maincar} \\leq rampingrate
 
       .. math::
-         -rampingrate \\leq \sum(Output_{t, car}) - \sum(Output_{t-1, car}) \\leq
+         -rampingrate \\leq \\sum(Output_{t, car}) - \\sum(Output_{t-1, car}) \\leq
          rampingrate
 
 
