### 2024-10-04 78f15fa
Changed DepreciationWarning to FutureWarning
