### 2024-10-04 545c21e
Added escape characters to resolve SyntaxWarnings for Docstrings
