+++ adopt_net0/components/technologies/specificTechnologies/heat_pump.py
@@ -42,7 +42,7 @@ class HeatPump(Technology):
     Ramping rates of the technology can be constraint.
 
     .. math::
-      -rampingrate \leq \sum(Input_{t, car}) - \sum(Input_{t-1, car}) \leq rampingrate
+      -rampingrate \leq \\sum(Input_{t, car}) - \\sum(Input_{t-1, car}) \leq rampingrate
 
     """
 
