+++ adopt_net0/components/technologies/genericTechnologies/conv2.py
@@ -26,30 +26,30 @@ class Conv2(Technology):
     - Size constraints are formulated on the input.
 
       .. math::
-         \sum(Input_{t, car}) \leq S
+         \\sum(Input_{t, car}) \leq S
 
     - It is possible to limit the maximum input of a carrier. This needs to be specified in the technology JSON files.
       Then it holds:
 
       .. math::
-        Input_{t, car} <= max_in_{car} * \sum(Input_{t, car})
+        Input_{t, car} <= max_in_{car} * \\sum(Input_{t, car})
 
     - ``performance_function_type == 1``: Linear through origin, i.e.:
 
       .. math::
-        Output_{t, car} == {\\alpha}_{1, car} \sum(Input_{t, car})
+        Output_{t, car} == {\\alpha}_{1, car} \\sum(Input_{t, car})
 
       .. math::
-        \min_part_load * S \leq {\\alpha}_1 \sum(Input_{t, car})
+        \min_part_load * S \leq {\\alpha}_1 \\sum(Input_{t, car})
 
     - ``performance_function_type == 2``: Linear with minimal partload (makes big-m transformation required). If the
       technology is in on, it holds:
 
       .. math::
-        Output_{t, car} = {\\alpha}_{1, car} \sum(Input_{t, car}) + {\\alpha}_{2, car}
+        Output_{t, car} = {\\alpha}_{1, car} \\sum(Input_{t, car}) + {\\alpha}_{2, car}
 
       .. math::
-        \sum(Input_{car}) \geq Input_{min} * S
+        \\sum(Input_{car}) \geq Input_{min} * S
 
       If the technology is off, input and output is set to 0:
 
@@ -57,7 +57,7 @@ class Conv2(Technology):
          Output_{t, car} = 0
 
       .. math::
-         \sum(Input_{t, car}) = 0
+         \\sum(Input_{t, car}) = 0
 
       If the technology has a standby-power, the input of the standy-by power carrier
       is:
@@ -80,7 +80,7 @@ class Conv2(Technology):
     - Additionally, ramping rates of the technology can be constrained.
 
       .. math::
-         -rampingrate \leq \sum(Input_{t, car}) - \sum(Input_{t-1, car}) \leq rampingrate
+         -rampingrate \leq \\sum(Input_{t, car}) - \\sum(Input_{t-1, car}) \leq rampingrate
 
     """
 
