### 2024-10-04 fb2f1cc
Added escape characters to resolve SyntaxWarnings for Docstrings
