+++ adopt_net0/components/technologies/genericTechnologies/conv2.py
@@ -18,7 +18,7 @@ class Conv2(Technology):
 
     This technology type resembles a technology with full input substitution,
     but different performance functions for the respective output carriers,
-    i.e. :math:`output_{car} = f_{car}(\sum(inputs))`. Three different performance
+    i.e. :math:`output_{car} = f_{car}(\\sum(inputs))`. Three different performance
     function fits are possible.
 
     **Constraint declarations:**
@@ -26,7 +26,7 @@ class Conv2(Technology):
     - Size constraints are formulated on the input.
 
       .. math::
-         \\sum(Input_{t, car}) \leq S
+         \\sum(Input_{t, car}) \\leq S
 
     - It is possible to limit the maximum input of a carrier. This needs to be specified in the technology JSON files.
       Then it holds:
@@ -40,7 +40,7 @@ class Conv2(Technology):
         Output_{t, car} == {\\alpha}_{1, car} \\sum(Input_{t, car})
 
       .. math::
-        \min_part_load * S \leq {\\alpha}_1 \\sum(Input_{t, car})
+        min_part_load * S \\leq {\\alpha}_1 \\sum(Input_{t, car})
 
     - ``performance_function_type == 2``: Linear with minimal partload (makes big-m transformation required). If the
       technology is in on, it holds:
@@ -49,7 +49,7 @@ class Conv2(Technology):
         Output_{t, car} = {\\alpha}_{1, car} \\sum(Input_{t, car}) + {\\alpha}_{2, car}
 
       .. math::
-        \\sum(Input_{car}) \geq Input_{min} * S
+        \\sum(Input_{car}) \\geq Input_{min} * S
 
       If the technology is off, input and output is set to 0:
 
@@ -80,7 +80,7 @@ class Conv2(Technology):
     - Additionally, ramping rates of the technology can be constrained.
 
       .. math::
-         -rampingrate \leq \\sum(Input_{t, car}) - \\sum(Input_{t-1, car}) \leq rampingrate
+         -rampingrate \\leq \\sum(Input_{t, car}) - \\sum(Input_{t-1, car}) \\leq rampingrate
 
     """
 
