### 2024-10-04 8b925db
Add possibility to skip writing results
