### 2024-10-14 #6988 @trevorb1
Hi @AdamRJensen; thanks so much for the offer to review this submission! I just read the paper and it looks really interesting!

In terms of the timeline, what is your expectation? I have a deliverable due the first week of November, and will be quite busy until then. However, I am available after that and can commit to having the review done by the weekend of Nov 23rd. I may be able to get it done a little earlier than this, but I don't want to over promise.

If this works with your timeline, please add me as a reviewer! Else, if you need the review done earlier, I will not be able to review.
