### 2024-10-15 #6988 @GregorBecker
Dear @AdamRJensen, due to an unfinished JOSS review I am involved in, I am unfortunately unable to participate in this review due to time constraints.
