### 2024-10-18 #6988 @JeanWi
@AdamRJensen. We thought about the following additional candidates:
- Elisabeth Zeyen (lisazeyen, TU Berlin)
- Fabian Neumann (fneum, TU Berlin)
- Diego Alejandro Tejada Arango (datejada, TNO Netherlands)
