### 2024-10-23 #6988 @JeanWi
Thanks everyone! @AdamRJensen: the editorial bot didn't reply, is this a bug or normal behavior?
