### 2024-10-23 #7402 @editorialbot
[REVIEW]: AdOpT-NET0: A technology-focused Python package for the optimization of multi-energy systems
**Submitting author:** @JeanWi (<a href="http://orcid.org/0000-0003-2993-2496">Jan F. Wiegner</a>)
**Repository:** https://github.com/UU-ER/AdOpT-NET0
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v0.1.7
