### 2024-11-06 #7402 @datejada
@AdamRJensen I have reviewed the checklist and the PDF. Overall, everything is in order regarding the checklist, except for the comparison with other similar packages. As for the PDF, I recommend that the authors clarify what they mean by the term "Flexible." For example, Table 1 states that the model offers flexibility in temporal resolution from 15 minutes to hours. However, does this flexibility apply to the entire model, or can different components of the model operate at varying resolutions (e.g., electricity every 15 minutes and gas every 6 hours)? More details on the term "Flexible" in the PDF would be appreciated.

Please let me know what the next steps are in the review process.

Thank you!
