+++ joss_submission/paper.bib
@@ -108,3 +108,20 @@ author = {Lukas Weimann and Guus Dubbink and Louis {van der Ham} and Matteo Gazz
 keywords = {Carbon capture, MEA, Energy system, Optimisation, MILP},
 abstract = {Assessing the role of carbon capture in energy systems dominated by non-dispatchable renewable energy sources requires a reliable and accurate model. However, carbon capture models used in complex systems optimisation are often very simplified. Therefore, we developed a mixed-integer linear model of post-combustion carbon capture starting from rigorous thermodynamic modelling in Aspen Plus. The final model decides the size and the operation of the capture process and returns the cost and energy requirements as a function of the CO2 concentration and the flow rate of the treated flue gas. Validation against actual plant data (Petra Nova) showed excellent accuracy with a deviation in total CO2 captured of just 2%. By applying the model to an exemplary case study, we show that it allows for co-optimising renewables deployment and carbon capture design and operation for a gas turbine, thus opening opportunities to explore new system designs of practical added value.}
 }
+
+
+@article{Hoffmann2024review,
+abstract = {Optimization-based frameworks for energy system modeling such as TIMES, ETHOS.FINE, or PyPSA have emerged as important tools to outline a cost-efficient energy transition. Consequently, numerous reviews have compared the capabilities and application cases of established energy system optimization frameworks with respect to their model features or adaptability but widely neglect the frameworks' underlying mathematical structure. This limits their added value for users who not only want to use models but also program them themselves. To address this issue, we follow a hybrid approach by not only reviewing 63 optimization-based frameworks for energy system modeling with a focus on their mathematical implementation but also conducting a meta-review of 68 existing literature reviews. Our work reveals that the basic concept of network-based energy flow optimization has remained the same since the earliest publications in the 1970s. Thereby, the number of open-source available optimization frameworks for energy system modeling has more than doubled in the last ten years, mainly driven by the uptake of energy transition and progress in computer-aided optimization. To go beyond a qualitative discussion, we also define the mathematical formulation for a mixed-integer optimization model comprising all the model features discussed in this work. We thereby aim to facilitate the implementation of future object-oriented frameworks and to increase the comprehensibility of existing ones for energy system modelers.},
+author = {Hoffmann, Maximilian and Schyska, Bruno U. and Bartels, Julian and Pelser, Tristan and Behrens, Johannes and Wetzel, Manuel and Gils, Hans Christian and Tang, Chuen-Fung and Tillmanns, Marius and Stock, Jan and Xhonneux, Andr{\'{e}} and Kotzur, Leander and Praktiknjo, Aaron and Vogt, Thomas and Jochem, Patrick and Lin{\ss}en, Jochen and Weinand, Jann M. and Stolten, Detlef},
+doi = {10.1016/j.adapen.2024.100190},
+issn = {26667924},
+journal = {Advances in Applied Energy},
+keywords = {Aggregation methods,Energy system optimization,Meta-review,Modeling frameworks,Multi-objective optimization},
+month = {dec},
+pages = {100190},
+publisher = {Elsevier},
+title = {{A review of mixed-integer linear formulations for framework-based energy system models}},
+url = {https://linkinghub.elsevier.com/retrieve/pii/S2666792424000283},
+volume = {16},
+year = {2024}
+}
