### 2024-11-08 3f82d05
Updated table + implemented reviewers comment about flexibility.
