+++ joss_submission/paper.md
@@ -54,47 +54,45 @@ highly complex but also offer synergies to reduce costs and environmental impact
 1 provides an overview of the covered dimensions of AdOpT-NET0, while
 Figure 1 and 2 show two examples of energy systems that can be modeled with the tool.
 
-| **Feature**                           | **AdOpT-NET0**                                                                                                                 |
-|---------------------------------------|--------------------------------------------------------------------------------------------------------------------------------|
-| **Model Dimensions**                  |                                                                                                                                |
-| Commodities                           | Energy and/or material commodities possible                                                                                    |
-| Space                                 | Single node or multi-node systems with network constraints                                                                     |
-| Time                                  | By default, hourly resolution (other resolutions possible)                                                                     |
-| Stochastic scenarios                  | Deterministic, Monte Carlo sampling possible                                                                                   |
-| Transformation pathways               | Multi-period possible, perfect forsight                                                                                        |
-| Components                            | Modelling of sources/sinks, converters, storage and networks possible. Linear or mixed-integer-linear                          |
-| **Component Extentions**              |                                                                                                                                |
-| Non-linear capacity expenditures      | Piece-wise investment cost function possible                                                                                   |
-| Technology dynamics                   | Constraining ramping, minimum-partload, minimum up-/down-time, maximum number of start-ups, slow start-ups/shut-downs possible |
-| Price elasticity of demand            | Not implemented                                                                                                                |
-| Demand response                       | Possible with defining a storage component                                                                                     |
-| Converter performance                 | Linear, piece-wise linear, technology specific                                                                                 |
-| Storage performance                   | Linear, piece-wise linear, technology specific                                                                                 |
-| Network performance                   | Linear, can be with an energy consumption of compression for gas networks                                                      |
-| **Boundary conditions**               |                                                                                                                                |
-| Technology potentials                 | Constraining maximum size of a technology possible                                                                             |
-| Regulations                           | Not implemented                                                                                                                |
-| System security and resource adequacy | Not implemented                                                                                                                |
-| **Multi criteria objectives**         | Pareto fronts                                                                                                                  |
-| **Complexity handling**               |                                                                                                                                |
-| Spatial aggregation                   | Not implemented                                                                                                                |
-| Technology aggregation                | Not implemented                                                                                                                |
-| Termporal aggragation                 | Typical periods via k-means clustering, averaging timesteps                                                                    |
-| Investment paths                      | One-time investment                                                                                                            |
-| **Model Implementation**              |                                                                                                                                |
-| Language                              | Python                                                                                                                         |
-| Translator                            | Pyomo                                                                                                                          |
-| Solver                                | Multiple (solvers compatible with Pyomo)                                                                                       |
+| **Feature**                                         | **AdOpT-NET0**                                                                                                                 |
+|-----------------------------------------------------|--------------------------------------------------------------------------------------------------------------------------------|
+| **Model Dimensions**                                |                                                                                                                                |
+| Commodities                                         | Energy and/or material commodities possible                                                                                    |
+| Space                                               | Single node or multi-node systems with network constraints                                                                     |
+| Time                                                | By default, hourly resolution (other resolutions possible)                                                                     |
+| Stochastic scenarios                                | Deterministic, Monte Carlo sampling possible                                                                                   |
+| Transformation pathways                             | Multi-period possible, perfect forsight                                                                                        |
[…]
