### 2024-11-08 ad1d010
Borders removed in table -> no horizontal lines in table possible
