+++ joss_submission/paper.md
@@ -70,7 +70,7 @@ Figure 1 and 2 show two examples of energy systems that can be modeled with the 
 | Transformation pathways                             | Multi-period possible, perfect forsight                                                                                        |
 +-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | Components                                          | Modelling of sources/sinks, converters, storage and networks possible. Linear or mixed-integer-linear                          |
-+=====================================================+================================================================================================================================+
++-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | **Component Extentions**                            |                                                                                                                                |
 +-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | Non-linear capacity expenditures                    | Piece-wise investment cost function possible                                                                                   |
@@ -86,7 +86,7 @@ Figure 1 and 2 show two examples of energy systems that can be modeled with the 
 | Storage performance                                 | Linear, piece-wise linear, technology specific                                                                                 |
 +-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | Network performance                                 | Linear, can be with an energy consumption of compression for gas networks                                                      |
-+=====================================================+================================================================================================================================+
++-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | **Boundary conditions**                             |                                                                                                                                |
 +-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | Technology potentials                               | Constraining maximum size of a technology possible                                                                             |
@@ -94,11 +94,11 @@ Figure 1 and 2 show two examples of energy systems that can be modeled with the 
 | Regulations                                         | Not implemented                                                                                                                |
 +-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | System security and resource adequacy               | Not implemented                                                                                                                |
-+=====================================================+================================================================================================================================+
++-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | **Multi criteria objectives**                       |                                                                                                                                |
 +-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | Pareto fronts                                       | Implemented                                                                                                                    |
-+=====================================================+================================================================================================================================+
++-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | **Complexity handling**                             |                                                                                                                                |
 +-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | Spatial aggregation                                 | Not implemented                                                                                                                |
@@ -108,7 +108,7 @@ Figure 1 and 2 show two examples of energy systems that can be modeled with the 
 | Termporal aggragation                               | Typical periods via k-means clustering, averaging timesteps                                                                    |
 +-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | Investment paths                                    | One-time investment                                                                                                            |
-+=====================================================+================================================================================================================================+
++-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | **Model Implementation**                            |                                                                                                                                |
 +-----------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------+
 | Language                                            | Python                                                                                                                         |
