### 2024-11-08 88a0324
Updated table + bibtex
