+++ joss_submission/paper.bib
@@ -22,11 +22,11 @@
 }
 
 @article{wiegner2024integration,
-  title={Energy system integration options for the North Sea region - Identifying key
-   energy infrastructure components},
+  title={Unleashing the full potential of the North Sea -- Identifying key energy infrastructure synergies for 2030 and 2040},
   author={Wiegner, Jan F and Gibescu, Madeleine and Gazzani, Matteo},
   year={2024},
-  journal={forthecoming}
+  journal={forthecoming},
+  doi={10.48550/arXiv.2411.00540}
 }
 
 @article{morales2017hidden,
@@ -111,7 +111,6 @@ abstract = {Assessing the role of carbon capture in energy systems dominated by 
 
 
 @article{Hoffmann2024review,
-abstract = {Optimization-based frameworks for energy system modeling such as TIMES, ETHOS.FINE, or PyPSA have emerged as important tools to outline a cost-efficient energy transition. Consequently, numerous reviews have compared the capabilities and application cases of established energy system optimization frameworks with respect to their model features or adaptability but widely neglect the frameworks' underlying mathematical structure. This limits their added value for users who not only want to use models but also program them themselves. To address this issue, we follow a hybrid approach by not only reviewing 63 optimization-based frameworks for energy system modeling with a focus on their mathematical implementation but also conducting a meta-review of 68 existing literature reviews. Our work reveals that the basic concept of network-based energy flow optimization has remained the same since the earliest publications in the 1970s. Thereby, the number of open-source available optimization frameworks for energy system modeling has more than doubled in the last ten years, mainly driven by the uptake of energy transition and progress in computer-aided optimization. To go beyond a qualitative discussion, we also define the mathematical formulation for a mixed-integer optimization model comprising all the model features discussed in this work. We thereby aim to facilitate the implementation of future object-oriented frameworks and to increase the comprehensibility of existing ones for energy system modelers.},
 author = {Hoffmann, Maximilian and Schyska, Bruno U. and Bartels, Julian and Pelser, Tristan and Behrens, Johannes and Wetzel, Manuel and Gils, Hans Christian and Tang, Chuen-Fung and Tillmanns, Marius and Stock, Jan and Xhonneux, Andr{\'{e}} and Kotzur, Leander and Praktiknjo, Aaron and Vogt, Thomas and Jochem, Patrick and Lin{\ss}en, Jochen and Weinand, Jann M. and Stolten, Detlef},
 doi = {10.1016/j.adapen.2024.100190},
 issn = {26667924},
