### 2024-11-08 f273ca3
Update paper.md
typos and final comments MG and IO
