### 2024-11-11 #7402 @datejada
@JeanWi Thank you for the changes, I like that you use the definitions in Hoffmann's paper, then it is easier to relate to other available models.

@AdamRJensen all my checklist is complete, I don't have further comments. Please let me know if there is something else I need to do. BR, Diego.
