### 2024-11-16 UU-ER/AdOpT-NET0#307 @trevorb1
[JOSS]: Python 3.13 Issue 
Installation on Python `3.13` does not work for me. If you confirm this problem, please pin the python version to `<3.13` (`3.12` worked fine for me) or correct this bug.

When installing in a fresh env with `3.13` I received this error:

```bash
$ pip install adopt_net0
Collecting adopt_net0
  Downloading adopt_net0-0.1.4-py3-none-any.whl.metadata (3.5 kB)
Collecting dill>=0.3.8 (from adopt_net0)
  Downloading dill-0.3.9-py3-none-any.whl.metadata (10 kB)
INFO: pip is looking at multiple versions of adopt-net0 to determine which version is compatible with other requirements. This could take a while.
Collecting adopt_net0
  Downloading adopt_net0-0.1.2-py3-none-any.whl.metadata (3.3 kB)
  Downloading adopt_net0-0.1.1-py3-none-any.whl.metadata (3.2 kB)
  Downloading adopt_net0-0.1.0-py3-none-any.whl.metadata (2.7 kB)
ERROR: Cannot install adopt-net0==0.1.0, adopt-net0==0.1.1, adopt-net0==0.1.2 and adopt-net0==0.1.4 because these package versions have conflicting dependencies.

The conflict is caused by:
    adopt-net0 0.1.4 depends on gurobipy>=11.0.1
    adopt-net0 0.1.2 depends on gurobipy>=11.0.1
    adopt-net0 0.1.1 depends on gurobipy>=11.0.1
    adopt-net0 0.1.0 depends on gurobipy>=11.0.1

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip to attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/topics/dependency-resolution/#dealing-with-dependency-conflicts
```
