### 2024-11-16 UU-ER/AdOpT-NET0#308 @trevorb1
[JOSS]: Docs Issue in Define Input Data when Specifying Networks 
In Step 2 under [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) on the documentation site, it says to specify the networks following this syntax:

```
"existing": [],
"new": ['electricityOffshore']
```

If I do this and run the command, `adopt.copy_network_data(input_data_path)` to copy the a json file I get the following `JSONDecodeError`

```
JSONDecodeError                           Traceback (most recent call last)
Cell In[6], line 1
----> 1 adopt.copy_network_data(p)

File ~/miniforge3/envs/adopt/lib/python3.12/site-packages/adopt_net0/data_preprocessing/data_loading.py:244, in copy_network_data(folder_path, ntw_data_path)
    242 json_ntw_file_path = folder_path / period / "Networks.json"
    243 with open(json_ntw_file_path, "r") as json_ntw_file:
--> 244     json_ntw = json.load(json_ntw_file)
    245 ntws_at_node = json_ntw["existing"] + json_ntw["new"]
    247 output_folder = folder_path / period / "network_data"

File ~/miniforge3/envs/adopt/lib/python3.12/json/__init__.py:293, in load(fp, cls, object_hook, parse_float, parse_int, parse_constant, object_pairs_hook, **kw)
    274 def load(fp, *, cls=None, object_hook=None, parse_float=None,
    275         parse_int=None, parse_constant=None, object_pairs_hook=None, **kw):
    276     """Deserialize ``fp`` (a ``.read()``-supporting file-like object containing
    277     a JSON document) to a Python object.
    278
   (...)
    291     kwarg; otherwise ``JSONDecoder`` is used.
    292     """
--> 293     return loads(fp.read(),
    294         cls=cls, object_hook=object_hook,
    295         parse_float=parse_float, parse_int=parse_int,
    296         parse_constant=parse_constant, object_pairs_hook=object_pairs_hook, **kw)
...
    354 except StopIteration as err:
--> 355     raise JSONDecodeError("Expecting value", s, err.value) from None
    356 return obj, end

JSONDecodeError: Expecting value: line 3 column 13 (char 34)
```

If I add double quotes, this solves the issue. ie.

```
"existing": [],
"new": ["electricityOffshore"]
```
