### 2024-11-16 UU-ER/AdOpT-NET0#309 @trevorb1
[JOSS]: Docs Issue in Define Input Data when Specifying Technology Data
In Step 3 under [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) on the documentation site, it says to specify the technologies following this syntax:

```
"existing": ['WindTurbine_Onshore_1500': 2, 'Photovoltaic': 2.4],
"new": ['Storage_Battery', 'Photovoltaic', 'Furnace_NG']
```

When I run the command `adopt.copy_technology_data(input_data_path)` after this, I get the following `JSONDecodeError`:

```
JSONDecodeError: Expecting ',' delimiter: line 2 column 44 (char 45)
```
