### 2024-11-16 #7402 @trevorb1
# Comments

Hi, @JeanWi thanks so much for submitting AdOpT-NET0 and suggesting me to review! This is a super interesting project that has been fun to test out, and seems like a great addition to the ESOM space! Please find below my comments and questions. If any of my comments are unclear, or you feel they are not valid, please let me know and I can clarify and/or discuss further!

## General Checks

All items complete!

## Functionality

- It would be good to include the link to the documentation site in the upper right hand side of the GitHub repository; under the "About" section.
- Installation on Python 3.13 does not work for me. Please see this [issue](https://github.com/UU-ER/AdOpT-NET0/issues/307).
- Functionality I have not been able to confirm yet, as I am experiencing [a bug](https://github.com/UU-ER/AdOpT-NET0/issues/309) working through the documentation that I can't solve. Once these are solved I will come back and assess functionality.

## Documentation

### General Comments

I appreciate the effort that has gone into documenting this project! This is an often overlooked step that can easily kill a project; so thank you for the effort! However, it feels like the `Getting Started` portion of docs have not been reviewed recently as I am experiencing several issues that are described below. [One of these issues](https://github.com/UU-ER/AdOpT-NET0/issues/309) is not letting me move forward with the review.

The current state of the documentation is not clear enough to guide the user through the configuration process. There is significant file management required to configure a model (which I fully understand, as this framework lends itself to user configuration), however, instructions on how to manage these configuration files needs to be improved. I would suggest either simplifying the configuration steps (ie. not having to copy files after running the template command), or being more explicit on what files the user is modifying/copying in each step (ie. using graphics such as a [tree diagram](https://tree.nathanfriend.com/) or something similar). Furthermore, directing users to locations that describe some of the core configuration options is needed (see specific issues below for examples.)

With that said, I think a more structured tutorial page may help answer many of my concerns. For example, if users are given a graphic showing a simple energy system to build (2 nodes, 2 networks, and just a couple techs), with an accompanying simple question to answer, and are walked through how to set up this model, this will help. Generalized documentation process can then have a specific example; allowing instructions like below (taken from [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) Step 2).

> Then, for each of the networks that you specify, an input data folder with that network name should be added in the corresponding folder (“existing” or “new”) in the network_topology folder.

to be replaced with something along the lines of:

> 1. Create the folder `data/period1/network_topology/existing/existingSimple`
> 2. Create the folder `data/period1/network_topology/new/existingSimple`

CaseStudy data and results are provided, however, I have not been able to generate this data myself through configuring a model following the documentation site.

### Specific Issues

- In [Defining System Topology](https://adopt-net0.readthedocs.io/en/latest/workflow/DefineTopology.html#defining-system-topology_) it is unclear where the `.\data\technology_data` folder is to reference for naming conventions. I see that in a later step this folder is created, but in the instructions its unclear what this folder is at this point in time.

- In [Defining System Topology](https://adopt-net0.readthedocs.io/en/latest/workflow/DefineTopology.html#defining-system-topology_) it states that investment periods are "a future feature, for now only single-period analysis is possible." This contradicts that paper that states that Multi-period studies are possible. Please clarify.

- In [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) the users are told to specify the networks in the model in step 2. While I can see the different network lists, its unclear what the differences are between the networks. For example, are `electricityOffshore` networks for only underwater interconnectors that may have different constraints/parameters applied to them? Then how does `electricityOnshore`, `electricityOffshore` and `electricitySimple` interact? Are all the constraints under the [`Network` class](https://adopt-net0.readthedocs.io/en/latest/src_code/model_components/networks.html#adopt_net0.components.networks.network.Network) available to all Network types? I would guess so based on the documentation note that `All networks are modelled using this class, and there are no subclasses of specific network models`, but then why are users only allowed to pick these networks?

- In [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) Step 2, the users are told to copy a `size_max_arcs.csv` to a folder. This csv does not exist in my templates.

- When defining my [Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) in Step 2, its unclear why I need distance? My node locations are already geolocated from step 1? Is this to represent scenarios where my node my be at a central location in an arbitrary region, but I want my connection to be at a demand centre? But if that is the case, we don't define these 'arbitrary shapes'? Can you please clarify?

- In [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) Step 2, the second example for creating a connection (`Connection: Example - allow_only_one_direction`) seems like this is a bidirectional link? But its labelled as only allowing one-direction? Please clarify.

- In [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) Step 2, the docs state "In order to read in the required data for our “electricityOffshore” network". It is still unclear the difference between these different networks. Does the electricityOffshore network contain data on sea depth, tides, ect.. that assist in assessing technology deployment decisions? Please clarify.

-  In [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) Step 3, the example json data is giving an error, and I have not been able to move past this point.  Please see [this](https://github.com/UU-ER/AdOpT-NET0/issues/308) and [this](https://github.com/UU-ER/AdOpT-NET0/issues/309) issue.

## Software Paper

### Overview

This paper describes the python package AdOpT-NET0; a modelling framework for performing energy system optimization modelling (ESOM) studies with a focus on the technology representation. Motivation for the need of multi-sector ESOM frameworks is first presented, followed by an overview of the features of AdOpT-NET0. Furthermore, a note on how AdOpT-NET0 builds on an existing MATLAB model to be open-source with feature improvements is given. Finally, a statement of need is presented highlighting how AdOpT-NET0 fits into the expanding field of ESOM frameworks.

Overall, the paper read well, was clear and concise, and stated clearly how AdOpT-NET0 fits into the literature (with the main emphasis placed on technology representation). Please find below minor clarification comments.

### Comments

- Line 21-22: You state that mulit-energy systems "are highly complex but also offer synergies to reduce costs and environmental impact." This reads like a definitive statement that deserves a reference. (Not that I disagree with the statement at all! Just a single reference here would be good to defend the need for multi-energy modelling).

- Table 1: Thanks for the comprehensive table! Can you please be more specific with the pareto front methods implemented. Similar to how you have given stochastic scenarios through "Monte Carlo" sampling, listing methods of pareto fronts would be good (ie. MGA, ɛ-constraint, ect.. following definitions from the Hoffmann paper)

- Line 30-31: (Apologies in advance for this rather pedantic wording comment!) You note that AdOpT-NET0 "allows for a highly realistic assessment of individual technologies and their integration into an energy system". I have trouble with this wording when no evidence is provided (ie. a ref to a research paper / study) to show that improved technology representation correlates to more realistic technology investment/integration decisions. AdOpT-NET0 can describe technology specific operational constraints well (and closer represents reality), but this does not guarantee highly realistic integration answers over multiple time horizons. Can you please clarify how AdOpT-NET0's detailed technology representation translates into realistic assessments of energy integration questions. Or can you please soften the wording to align with a prescriptive analytics tool (ie. guide decision making) rather than a predictive analytics tool (ie. predicting what will happen in the future). To clarify, I have no issue with the statement of representing technologies more realistically (which you give references for later), but rather how this translates to more realistic future energy integration questions (which is likely difficult to measure?).

- Figure 1: The figure is clear, I just have a question about the reference. I see it says adapted from Tiggeloven et al.; and it seems to be a cropped version of Figure 2 from that paper (specifically, box 2d)? I am unsure if this requires the license of the figure to also be included with caption? @JeanWi, if this figure is cropped from the Tiggeloven et al. reference, maybe @AdamRJensen can confirm if stating "adaptation" alone is okay, or if the figure license is also required?

- Line 32-33: (A manuscript suggestion that you are free to take or leave) You note that "several complexity reduction algorithms can be adopted to deal with infeasible computation times". I believe this is a super valuable point, and explicitly listing them for the reader is useful. ie. Something along the lines of "Complexity reduction algorithms for modelling seasonal storage (Gabrielli et al., 2018) and high VRE systems (Weimann & Gazzani, 2022) are implemented." Having these algorithms already integrated is a big plus of AdOpT-NET0 and would be good to highlight!

- Line 36: Its great that the closed MATLAB version has been ported to Python and made openly available. Is there any relevant publications or similar that use or describe the MATLAB version that require a referencing here? Or has it been used solely for in-house business applications?

- Statement of need paragraph: Listing the specific technologies you represent in greater detail compared to existing models is great, as it helps the reader understand the novelty of this framework! Where further clarification will help is with the target user. In the opening sentence you break down the field into complex NLP frameworks, and simplified LP frameworks. The users of this different frameworks may be different, though? (ie. policy focused research questions will use LP frameworks, while power flow research question (as you mention) may want the NLP detail). It is a little unclear to me if AdOpT-NET0 can be used both as a policy support tool and a power flow analysis tool, or just one or the other (or a category I am missing!).

- Has AdOpT-NET0 been used in any publications (other than the forthcoming one by Wiegner et al.)? If so, please list them, else please just disregard this comment.
