### 2024-11-18 #7402 @JeanWi
Hi @trevorb1, thanks a lot for taking the time to write this review, we really appreciate the time you put into this! Sorry for the bug in the documentation, that's indeed plain wrong. For the existing technologies it should have read (note the curly brackets):

`"existing": {'WindTurbine_Onshore_1500': 2, 'Photovoltaic': 2.4},`

We will start tackling your points now, but I thought I'd let you know already so you can continue in case you find the time.
Looking forward for your comments on trying out the tool!
