### 2024-11-18 #7402 @JeanWi
@AdamRJensen Regarding Figure 1 (and actually also Figure 2): This is an adaption of a figure from a [work by Julia Tiggeloven](https://pubs.acs.org/doi/10.1021/acs.iecr.3c02226) (referenced in the caption, she also coauthored this paper,  licensed under CC-BY-4.0). Also Figure 2 is an adaption of a figure in one of [my papers](https://arxiv.org/abs/2411.00540) (only available as a pre-print so far). So I think it shouldn't be a problem. Let me know, if I missed something.
