### 2024-11-18 b7a8986
bugfix overwriting bidirectional
