+++ joss_submission/paper.bib
@@ -124,3 +124,13 @@ url = {https://linkinghub.elsevier.com/retrieve/pii/S2666792424000283},
 volume = {16},
 year = {2024}
 }
+
+@article{mancarella2014mes,
+  title={MES (multi-energy systems): An overview of concepts and evaluation models},
+  author={Mancarella, Pierluigi},
+  journal={Energy},
+  volume={65},
+  pages={1--17},
+  year={2014},
+  publisher={Elsevier}
+}
