### 2024-11-18 ebc383e
add forthcoming reference
