+++ joss_submission/paper.bib
@@ -29,6 +29,13 @@
   doi={10.48550/arXiv.2411.00540}
 }
 
+@article{tiggeloven2024chemicalcluster,
+  title={Optimizing emissions reduction in ammonia-ethylene chemical clusters: synergistic integration of electrification, carbon capture, and hydrogen},
+  author={Tiggeloven, Julia L. and Faaij, Andre P.C. and Kramer, Gert Jan and Gazzani, Matteo},
+  year={2024},
+  journal={forthcoming},
+  doi={}
+
 @article{morales2017hidden,
   title={Hidden power system inflexibilities imposed by traditional unit commitment formulations},
   author={Morales-España, Germán and Ramírez-Elizondo, Laura and Hobbs, Benjamin F},
