### 2024-11-19 fa5695d
Check if packages are installed from pyproject.toml
