+++ .github/workflows/test_testing.yml
@@ -1,0 +1,34 @@
+name: Testing
+
+on:
+  workflow_call:
+
+jobs:
+  test:
+    runs-on: ${{ matrix.os }}
+
+    strategy:
+      matrix:
+        os: [windows-latest]
+
+    steps:
+    - name: Install GLPK Solver
+      run: | 
+        choco install glpk
+
+    - name: Checkout Code
+      uses: actions/checkout@v4
+
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: 3.12
+
+    - name: Install Dependencies
+      env:
+        PIP_CONSTRAINT: .github/workflows/constraints.txt
+      run: |
+        python -m pip install --upgrade pip
+        pip install coverage
+        pip install poetry
+        poetry install
