### 2024-11-19 c3b38c0
Check if packages are installed from pyproject.toml
