+++ .github/workflows/test_testing.yml
@@ -1,4 +1,4 @@
-name: Testing
+name: Test Testing
 
 on:
   workflow_call:
