### 2024-11-19 98b2897
Check if packages are installed from pyproject.toml
