+++ .github/workflows/test_00develop.yml
@@ -7,7 +7,6 @@ on:
 
 jobs:
   testing:
-    needs: autoformatting
     uses: ./.github/workflows/test_testing.yml
 
 
