+++ .github/workflows/test_testing.yml
@@ -28,3 +28,9 @@ jobs:
         pip install coverage
         pip install poetry
         poetry install
+
+    - name: Export requirements.txt
+      uses: actions/upload-artifact@v4
+      with:
+        name: poetry export --output requirements.txt
+        path: requirements.txt
