+++ .github/workflows/test_testing.yml
@@ -29,8 +29,23 @@ jobs:
         pip install poetry
         poetry install
 
-    - name: Export requirements.txt
-      uses: actions/upload-artifact@v4
-      with:
-        name: poetry export --output requirements.txt
-        path: requirements.txt
+          - name: Export requirements.txt
+        run: |
+          poetry export --without-hashes --output requirements.txt
+
+      # Step 5: Commit the changes
+    - name: Commit and push changes
+      run: |
+        git config user.name "GitHub Actions"
+        git config user.email "actions@github.com"
+        git add requirements.txt
+        git commit -m "Update requirements.txt generated by Poetry"
+        continue-on-error: true  # Prevents failure if no changes to commit
+
+      # Step 6: Push the changes back to the repository
+    - name: Push changes
+      run: |
+        git push
+      env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
