+++ .github/workflows/test_testing.yml
@@ -40,7 +40,6 @@ jobs:
         git config user.email "actions@github.com"
         git add requirements.txt
         git commit -m "Update requirements.txt generated by Poetry"
-        continue-on-error: true  # Prevents failure if no changes to commit
 
       # Step 6: Push the changes back to the repository
     - name: Push changes
