+++ pyproject.toml
@@ -17,9 +17,11 @@ documentation = "https://adopt-net0.readthedocs.io"
 
 [tool.poetry.dependencies]
 python = "^3.12"
+tsam = ">=2.3.1"
+
+[tool.poetry.group.test]
 Pyomo = ">=6.7.1"
 numpy = ">=1.26.4"
-dill = ">=0.3.8"
 pandas = ">=2.2.2"
 requests = ">=2.31.0"
 timezonefinder = ">=6.5.0"
@@ -28,12 +30,13 @@ pvlib = ">=0.10.4"
 scipy = ">=1.13.0"
 openpyxl = ">=3.1.2"
 pytest = ">=8.1.1"
-scikit-learn = ">=1.4.2"
 pwlf = ">=2.2.1"
 gurobipy = ">=11.0.1"
 scandir = ">=1.10.0"
 tables = ">=3.9.2"
-tsam = ">=2.3.1"
+
+[tool.poetry.group.dev.dependencies]
+
 
 [tool.poetry.urls]
 homepage = "https://github.com/UU-ER/AdOpT-NET0"
