+++ .github/workflows/test_testing.yml
@@ -33,18 +33,18 @@ jobs:
       run: |
         poetry export --without-hashes --without dev --output requirements.txt
 
-      # Step 5: Commit the changes
     - name: Commit and push changes
       run: |
-        git config user.name "GitHub Actions"
-        git config user.email "actions@github.com"
+        git config --global user.name "GitHub Actions"
+        git config --global user.email "actions@github.com"
         git add requirements.txt
         git commit -m "Update requirements.txt generated by Poetry"
-
-      # Step 6: Push the changes back to the repository
-    - name: Push changes
-      run: |
         git push
-      env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+#
+#      # Step 6: Push the changes back to the repository
+#    - name: Push changes
+#      run: |
+#        git push
+#      env:
+#          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
