+++ .github/workflows/test_testing.yml
@@ -14,6 +14,8 @@ jobs:
     steps:
     - name: Checkout Code
       uses: actions/checkout@v4
+      with:
+        ref: ${{ github.head_ref }}
 
     - name: Set up Python
       uses: actions/setup-python@v4
@@ -35,8 +37,8 @@ jobs:
 
     - name: Commit and push changes
       run: |
-        git config --global user.name "GitHub Actions"
-        git config --global user.email "actions@github.com"
+        git config user.name "GitHub Actions"
+        git config user.email "actions@github.com"
         git add requirements.txt
         git commit -m "Update requirements.txt generated by Poetry"
         git push
