### 2024-11-19 731a053
added CI workflow to pin packages in requirements.txt
