+++ .github/workflows/pin_requirements.yml
@@ -1,0 +1,42 @@
+name: Pin requirements
+
+on:
+  workflow_call:
+
+jobs:
+  test:
+    runs-on: ${{ matrix.os }}
+
+    strategy:
+      matrix:
+        os: [windows-latest]
+
+    steps:
+    - name: Checkout Code
+      uses: actions/checkout@v4
+      with:
+        ref: ${{ github.head_ref }}
+
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: 3.12
+
+    - name: Install Dependencies
+      env:
+        PIP_CONSTRAINT: .github/workflows/constraints.txt
+      run: |
+        python -m pip install --upgrade pip
+        pip install poetry
+
+    - name: Write requirements.txt
+      run: |
+        poetry export --without-hashes --without dev --output requirements.txt
+
+    - name: Commit and push changes
+      run: |
+        git config user.name "GitHub Actions"
+        git config user.email "actions@github.com"
+        git add requirements.txt
+        git commit -m "Update requirements.txt generated by Poetry"
+        git push
