+++ pyproject.toml
@@ -18,10 +18,6 @@ documentation = "https://adopt-net0.readthedocs.io"
 [tool.poetry.dependencies]
 python = "^3.12"
 tsam = ">=2.3.1"
-
-[tool.poetry.group.test.dependencies]
-
-[tool.poetry.group.dev.dependencies]
 Pyomo = ">=6.7.1"
 numpy = ">=1.26.4"
 pandas = ">=2.2.2"
@@ -36,6 +32,16 @@ pwlf = ">=2.2.1"
 gurobipy = ">=11.0.1"
 scandir = ">=1.10.0"
 tables = ">=3.9.2"
+
+[tool.poetry.group.test.dependencies]
+pytest = ">=8.1.1"
+coverage = ">=7.6.7"
+
+[tool.poetry.group.dev.dependencies]
+pre-commit = ">=3.7.0"
+black = ">=24.4.0"
+sphinx = ">=7.3.7"
+sphinx_rtd_theme = ">=2.0.0"
 
 [tool.poetry.urls]
 homepage = "https://github.com/UU-ER/AdOpT-NET0"
