### 2024-11-19 ba520ec
Remove setuptools constraints
