+++ .github/workflows/testing.yml
@@ -25,8 +25,6 @@ jobs:
         python-version: 3.12
 
     - name: Install Dependencies
-      env:
-        PIP_CONSTRAINT: .github/workflows/constraints.txt
       run: |
         python -m pip install --upgrade pip
         pip install coverage
