### 2024-11-19 f60eefd
Remove setuptools constraints
