### 2024-11-19 4f6798c
include test group in testing.yml
