+++ .github/workflows/testing.yml
@@ -27,8 +27,8 @@ jobs:
     - name: Install Dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install coverage
-        pip install -r requirements.txt
+        pip install poetry
+        poetry install --with test
 
     - name: Check directory
       run: |
