### 2024-11-19 7ed14f3
- test on python 3.12 and 3.13 - install coverage
