+++ .github/workflows/testing.yml
@@ -10,6 +10,7 @@ jobs:
     strategy:
       matrix:
         os: [windows-latest]
+        python-version: [3.12, 3.13]
 
     steps:
     - name: Install GLPK Solver
@@ -22,13 +23,14 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v4
       with:
-        python-version: 3.12
+        python-version: ${{ matrix.python-version }}
 
     - name: Install Dependencies
       run: |
         python -m pip install --upgrade pip
         pip install poetry
-        poetry install --with test
+        pip install coverage
+        poetry install
 
     - name: Check directory
       run: |
