### 2024-11-19 aff77df
install pytest
