+++ .github/workflows/testing.yml
@@ -29,7 +29,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install poetry --without test,dev
-        pip install coverage, pytest
+        pip install coverage pytest
         poetry install
 
     - name: Check directory
