+++ .github/workflows/testing.yml
@@ -28,9 +28,9 @@ jobs:
     - name: Install Dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install poetry --without test,dev
+        pip install poetry
         pip install coverage pytest
-        poetry install
+        poetry install --without test,dev
 
     - name: Check directory
       run: |
