### 2024-11-19 eed8164
adapted testing to run coverage in virtual environment created by poetry
