+++ .github/workflows/testing.yml
@@ -29,7 +29,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install poetry
-        pip install coverage pytest
         poetry install --without test,dev
 
     - name: Check directory
@@ -38,11 +37,11 @@ jobs:
 
     - name: Test
       run: |
-        coverage run -m pytest
+        poetry run coverage run -m pytest
 
     - name: Build coverage report
       run: |
-        coverage report -m >coverage.txt
+        poetry run coverage report -m >coverage.txt
 
     - name: Upload Coverage Report
       uses: actions/upload-artifact@v4
