### 2024-11-19 c184020
adapted testing to run coverage in virtual environment created by poetry
