+++ .github/workflows/testing.yml
@@ -29,7 +29,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install poetry
-        poetry install --without test,dev
+        poetry install --without dev
 
     - name: Check directory
       run: |
