### 2024-11-19 9a7bcd8
Moved gurobipy to dev.dependencies
