### 2024-11-19 a286b6d
Removed gurobipy from dependencies
