### 2024-11-19 d66d8fc
- Dynamic versioning with poetry - moved tables to dev.dependencies
