+++ .github/workflows/uploadpypi.yml
@@ -26,31 +26,32 @@ jobs:
     - name: Install Dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install build
+        pip install poetry
+        poetry install --without test,dev
 
     - name: Build a binary wheel and a source tarball
-      run: python -m build
+      run: poetry build
     - name: Store the distribution packages
       uses: actions/upload-artifact@v4
       with:
         name: python-package-distributions
         path: dist/
-
-  publish-to-testpypi:
-    name: Upload release to PyPI
-    needs:
-      - build
-    runs-on: ubuntu-latest
-    environment:
-      name: pypi
-      url: https://pypi.org/p//adopt-net0
-    permissions:
-      id-token: write  # IMPORTANT: this permission is mandatory for trusted publishing
-    steps:
-    - name: Download all the dists
-      uses: actions/download-artifact@v4
-      with:
-        name: python-package-distributions
-        path: dist/
-    - name: Publish package distributions to PyPI
-      uses: pypa/gh-action-pypi-publish@release/v1
+#
+#  publish-to-testpypi:
+#    name: Upload release to PyPI
+#    needs:
+#      - build
+#    runs-on: ubuntu-latest
+#    environment:
+#      name: pypi
+#      url: https://pypi.org/p//adopt-net0
+#    permissions:
+#      id-token: write
+#    steps:
+#    - name: Download all the dists
+#      uses: actions/download-artifact@v4
+#      with:
+#        name: python-package-distributions
+#        path: dist/
+#    - name: Publish package distributions to PyPI
+#      uses: pypa/gh-action-pypi-publish@release/v1
