### 2024-11-19 b5274e0
Dynamic versioning with poetry
