### 2024-11-19 2665926
- coverage reports uploaded for each python version
