+++ pyproject.toml
@@ -18,6 +18,9 @@ documentation = "https://adopt-net0.readthedocs.io"
 [tool.poetry-dynamic-versioning]
 enable = true
 
+[tool.poetry-dynamic-versioning.from-file]
+source = "version.py"
+
 [tool.poetry.dependencies]
 python = "^3.12"
 tsam = ">=2.3.1"
