+++ .github/workflows/testing.yml
@@ -47,4 +47,4 @@ jobs:
       uses: actions/upload-artifact@v4
       with:
         name: coverage-report
-        path: coverage.txt
+        path: "coverage-${{ matrix.python-version }}.txt"
