+++ requirements.txt
@@ -1,20 +1,15 @@
-blosc2==2.7.1 ; python_version >= "3.12" and python_version < "4"
 certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
 cffi==1.17.1 ; python_version >= "3.12" and python_version < "4"
 charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
 colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
 et-xmlfile==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
-gurobipy==12.0.0 ; python_version >= "3.12" and python_version < "4.0"
 h3==4.1.2 ; python_version >= "3.12" and python_version < "4"
 h5py==3.12.1 ; python_version >= "3.12" and python_version < "4.0"
 highspy==1.8.1 ; python_version >= "3.12" and python_version < "4.0"
 idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
 iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
 joblib==1.4.2 ; python_version >= "3.12" and python_version < "4.0"
-msgpack==1.1.0 ; python_version >= "3.12" and python_version < "4"
-ndindex==1.9.2 ; python_version >= "3.12" and python_version < "4"
 networkx==3.4.2 ; python_version >= "3.12" and python_version < "4.0"
-numexpr==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
 numpy==2.1.3 ; python_version >= "3.12" and python_version < "4.0"
 openpyxl==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
 packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
@@ -24,7 +19,6 @@ pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
 ply==3.11 ; python_version >= "3.12" and python_version < "4.0"
 pvlib==0.11.1 ; python_version >= "3.12" and python_version < "4.0"
 pwlf==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
-py-cpuinfo==9.0.0 ; python_version >= "3.12" and python_version < "4.0"
 pycparser==2.22 ; python_version >= "3.12" and python_version < "4"
 pydoe==0.3.8 ; python_version >= "3.12" and python_version < "4.0"
 pyomo==6.8.2 ; python_version >= "3.12" and python_version < "4.0"
@@ -38,11 +32,9 @@ scipy==1.14.1 ; python_version >= "3.12" and python_version < "4.0"
 setuptools==75.5.0 ; python_version >= "3.12" and python_version < "4"
 six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
 statsmodels==0.14.4 ; python_version >= "3.12" and python_version < "4.0"
-tables==3.10.1 ; python_version >= "3.12" and python_version < "4.0"
 threadpoolctl==3.5.0 ; python_version >= "3.12" and python_version < "4.0"
 timezonefinder==6.5.4 ; python_version >= "3.12" and python_version < "4"
 tqdm==4.67.0 ; python_version >= "3.12" and python_version < "4.0"
 tsam==2.3.5 ; python_version >= "3.12" and python_version < "4.0"
-typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
 tzdata==2024.2 ; python_version >= "3.12" and python_version < "4.0"
 urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
