### 2024-11-19 3cadf16
- coverage reports uploaded for each python version
