+++ .github/workflows/pin_requirements.yml
@@ -4,7 +4,7 @@ on:
   workflow_call:
 
 jobs:
-  test:
+  pin_requirements:
     runs-on: ${{ matrix.os }}
 
     strategy:
