+++ .github/workflows/uploadpypi.yml
@@ -36,22 +36,22 @@ jobs:
       with:
         name: python-package-distributions
         path: dist/
-#
-#  publish-to-testpypi:
-#    name: Upload release to PyPI
-#    needs:
-#      - build
-#    runs-on: ubuntu-latest
-#    environment:
-#      name: pypi
-#      url: https://pypi.org/p//adopt-net0
-#    permissions:
-#      id-token: write
-#    steps:
-#    - name: Download all the dists
-#      uses: actions/download-artifact@v4
-#      with:
-#        name: python-package-distributions
-#        path: dist/
-#    - name: Publish package distributions to PyPI
-#      uses: pypa/gh-action-pypi-publish@release/v1
+
+  publish-to-pypi:
+    name: Upload release to PyPI
+    needs:
+      - build
+    runs-on: ubuntu-latest
+    environment:
+      name: pypi
+      url: https://pypi.org/p//adopt-net0
+    permissions:
+      id-token: write
+    steps:
+    - name: Download all the dists
+      uses: actions/download-artifact@v4
+      with:
+        name: python-package-distributions
+        path: dist/
+    - name: Publish package distributions to PyPI
+      uses: pypa/gh-action-pypi-publish@release/v1
