+++ .github/workflows/00develop.yml
@@ -6,16 +6,12 @@ on:
       - 'develop'
 
 jobs:
+  pin_requirements:
+    uses: ./.github/workflows/pin_requirements.yml
   autoformatting:
+    needs: pin_requirements
     uses: ./.github/workflows/autoformatting.yml
   testing:
     needs: autoformatting
     uses: ./.github/workflows/testing.yml
-  pin_requirements:
-    needs: testing
-    uses: ./.github/workflows/pin_requirements.yml
-  build:
-    uses: ./.github/workflows/uploadpypi.yml
-
-
 
