+++ .github/workflows/uploadpypi.yml
@@ -14,7 +14,7 @@ jobs:
     - name: Create VERSION file
       run: |
         version=${GITHUB_REF#refs/*/}
-        echo "__version__ = \"${version//v/}\"" > version.py
+        echo "${version//v/}" > version.py
     - name: Check VERSION file
       run: cat version.py
 
@@ -27,6 +27,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install poetry
+        poetry self add "poetry-dynamic-versioning[plugin]"
         poetry install --without test,dev
 
     - name: Build a binary wheel and a source tarball
