### 2024-11-20 413bd62
Zurich case study OB
