+++ OceanBattery_Zurich.py
@@ -1,0 +1,139 @@
+# TODO: Implement option for complete linearization
+# TODO: Implement length of time step
+# TODO: Implement all technologies
+import pandas as pd
+import numpy as np
+from pyomo.environ import *
+from pathlib import Path
+
+from src.model_construction.construct_balances import add_system_costs
+from src.model_configuration import ModelConfiguration
+import src.data_management as dm
+from src.energyhub import EnergyHub
+
+
+
+# Save Data File to file
+data_save_path = Path('./user_data/data_handle_test')
+
+# TOPOLOGY
+topology = dm.SystemTopology()
+topology.define_time_horizon(year=2001,start_date='05-01 00:00', end_date='05-30 '
+                                                                          '23:00', resolution=1)
+topology.define_carriers(['electricity'])
+topology.define_nodes(['offshore'])
+topology.define_new_technologies('offshore', ['Storage_OceanBattery_specific_3'])
+
+# Initialize instance of DataHandle
+data = dm.DataHandle(topology)
+
+SD_list = [8]
+CAPEX_list = [0.3]
+# SD_list = np.arange(1, 2, 1)
+# CAPEX_list = np.arange(0.1, 0.3, 0.1)
+
+
+# CLIMATE DATA
+from_file = 0
+if from_file == 0:
+    data.read_climate_data_from_file('offshore', './data/climate_data_offshore.p')
+    # data.read_climate_data_from_file('offshore', './data/climate_data_offshore.txt')
+else:
+    # lat = 52
+    # lon = 5.16
+    # data.read_climate_data_from_api('offshore', lon, lat,save_path='./data/climate_data_offshore.txt')
+    lat = 52.2
+    lon = 4.4
+    data.read_climate_data_from_api('offshore', lon, lat,
+                                    save_path='./data/climate_data_offshore.p')
+
+
+# PRODUCTION
+# electricity_production = np.ones(len(topology.timesteps)) * 1000
+# data.read_production_profile('offshore', 'electricity', electricity_production, 1)
+
+# DEMAND
+# electricity_demand_low = 800
+# electricity_demand_high = 1000
+# electricity_demand = np.zeros(len(topology.timesteps))
+# for i in range(len(topology.timesteps)):
+#     if i % 2 == 0:
+#         electricity_demand[i] = electricity_demand_low
+#     else:
+#         electricity_demand[i] = electricity_demand_high
+# data.read_demand_data('offshore', 'electricity', electricity_demand)
+
+# IMPORT
+el_import = np.ones(len(topology.timesteps)) * 1000
+data.read_import_limit_data('offshore', 'electricity', el_import)
+# EXPORT
+el_export = np.ones(len(topology.timesteps)) * 1000
+data.read_export_limit_data('offshore', 'electricity', el_export)
+
+# Electricity Price
+loadpath_electricityprice = './data/ob_input_data/day_ahead_2019.csv'
+el_import_price = pd.read_csv(loadpath_electricityprice)
+el_import_price = el_import_price['Day-ahead Price [EUR/MWh]'][0:8760]
+el_import_price = el_import_price.interpolate()
+el_import_price = np.array(el_import_price)
+
+data.read_import_price_data('offshore', 'electricity', el_import_price)
+data.read_export_price_data('offshore', 'electricity', el_import_price)
+
+# Read data
+data.read_technology_data()
+data.read_network_data()
+
+# Orignial values
+capex_original = data.technology_data['offshore']['Storage_OceanBattery_specific_3'].economics.capex_data['unit_capex']
+
+
+# SAVING/LOADING DATA FILE
+configuration = ModelConfiguration()
+# configuration.reporting.case_name = 'MAY19_SD' + str(round(sd, 2)) + '_CAPEX' + str(round(capex_share, 2))
+
+# # Read data
[…]
