### 2024-11-22 UU-ER/AdOpT-NET0#324 @trevorb1
[JOSS]: Unable to get `adopt.load_climate_data_from_api()` function to work 
Following the instructions in [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data), I am told to use the [`adopt.load_climate_data_from_api()`](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#climate-data-from-api) to retrieve JRC data. However, I receive the following error.

```
import adopt_net0 as adopt
p = "data"
adopt.load_climate_data_from_api(p)
```

```
Importing Climate Data...
<Response [400]>
Traceback (most recent call last):
  File "/home/trevor/repos/AdOpT-NET0/review/joss.py", line 5, in <module>
    adopt.load_climate_data_from_api(p)
  File "/home/trevor/miniforge3/envs/adopt/lib/python3.12/site-packages/adopt_net0/data_preprocessing/data_loading.py", line 66, in load_climate_data_from_api
    data = import_jrc_climate_data(lon, lat, year, alt)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/trevor/miniforge3/envs/adopt/lib/python3.12/site-packages/adopt_net0/data_preprocessing/data_loading.py", line 310, in import_jrc_climate_data
    climate_data = data["outputs"]["tmy_hourly"]
                   ~~~~^^^^^^^^^^^
KeyError: 'outputs'
```
