### 2024-11-22 UU-ER/AdOpT-NET0#325 @trevorb1
[JOSS]: Broken Table Link in Documentation 
In [Define Model Configuration](https://adopt-net0.readthedocs.io/en/latest/workflow/DefineModelConfiguration.html#define-model-configuration), there is a broken table link. The paragraph reads:

> In the table :ref:here<model_configuration> here you can find all settings that can be specified, and a description, the options that can be chosen and the default value from the template for each setting.
