### 2024-11-22 UU-ER/AdOpT-NET0#324 @trevorb1
On this note, when I follow the [example 1](https://adopt-net0.readthedocs.io/en/latest/case_studies/CaseStudy_DAC.html#optimizing-the-design-and-operation-of-a-dac-system-under-varying-ambient-conditions) in the docs this code runs fine. So Im not sure if its me missing a file I was supposed to copy over somewhere, of a file that got missed creating in the instructions?
