### 2024-11-23 UU-ER/AdOpT-NET0#326 @trevorb1
[JOSS]: Case Study 1 Docs
Please find below comments relating to the [case study 1](https://adopt-net0.readthedocs.io/en/latest/case_studies/CaseStudy_DAC.html#optimizing-the-design-and-operation-of-a-dac-system-under-varying-ambient-conditions) documentation.

- A note adding that users must create the folder `./caseStudies/dac` themselves, or having the template function create the folder for them would be good. Else you run into a `FileNotFoundError: [Errno 2] No such file or directory:`
- Missing bracket in the line
  > ... file itself (Topology.json) or, as we do it here) via some lines ...
- Please link to the explanation of "Method 1" when it notes below. I believe [this](https://adopt-net0.readthedocs.io/en/latest/advanced_topics/time_aggregation.html#clustering-into-typical-days) is the correct location?
  > As the DAC model is rather complex, we also cluster the full resolution into 50 typical days (method 1).
- When I run the code under [`Run Model - Infeasability`](https://adopt-net0.readthedocs.io/en/latest/case_studies/CaseStudy_DAC.html#run-model-infeasibility), I get the following error: `FileNotFoundError: The folder you want to save your results to ('userData') does not exist. Create the folder or change the folder name in the ModelConfig`. Its easy to solve, but should be a step in the instructions.
- The visualization dashboard is quite nice! It would be good to add in a step just letting the user know where the result is to drag into the platform (ie. `userData/Summary.xlsx` or the `*.h5` file)
