### 2024-11-23 UU-ER/AdOpT-NET0#327 @trevorb1
[JOSS]: Case Study 3 docs
Please find below comments relating to the [case study 3](https://adopt-net0.readthedocs.io/en/latest/case_studies/CaseStudy_Networks.html) documentation.
- The first image is not loading for me. I tried on two different browsers.
- Similar to case study 1, would be good to link to your docs when you discuss setting [Method 1](https://adopt-net0.readthedocs.io/en/latest/case_studies/CaseStudy_Networks.html#adapt-model-configurations) for the model configuration
