### 2024-11-23 UU-ER/AdOpT-NET0#328 @trevorb1
[JOSS]: Pytest execution instructions 
To run tests, the docs say to run the following command:

```bash
coverage run -m pytest
```

However, I received the following error and had to `pip install coverage`.

```bash
$ coverage run -m pytest
coverage: command not found
```

Please either create a desperate developer install version of your package (with optional dependencies) or update instructions to include the `coverage` requirement (if this issue can be replicated by your development team).
