### 2024-11-23 UU-ER/AdOpT-NET0#330 @trevorb1
[JOSS]: Tests create many new files 
When I run tests via `coverage run -m pytest`. many new files are created for me. I would suggest either using temporary files, or removing these files at the end of the test, if possible?

![image](https://github.com/user-attachments/assets/e7dcf35f-b091-409d-8fa1-70de9081270d)
