### 2024-11-23 UU-ER/AdOpT-NET0#331 @trevorb1
[JOSS]: Issue Templates 
In the docs for [reporting issues](https://adopt-net0.readthedocs.io/en/latest/ForDevelopers.html#reporting-issues), it states to:

> If you encounter an issue or a missing feature, you can report it on [github](https://github.com/UU-ER/AdOpT-NET0/issues). Please attach a detailed description and a tag (bug, enhancement,…)

It would be good to provide [issue template](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository) for the users so you can get the exact information you want for an enhancement, bug, doc issue, ect. They are quite easy to set up and very helpful!
