### 2024-11-23 UU-ER/AdOpT-NET0#332 @trevorb1
[JOSS]: Issue with path in `In Short` sections
The first command in the [In Short](https://adopt-net0.readthedocs.io/en/latest/ModellingWorkflow.html#in-short) section states to create the input path by:

```
import adopt_net0 as adopt
import json
from pathlib import Path

input_data_path = "path_to_your_input_data_folder"
adopt.create_optimization_templates(input_data_path)
```

I believe the input path file should be specified using the `pathlib` library

```
input_data_path = Path("path_to_your_input_data_folder")
```

Else you run into the following error:

```python
with open(input_data_path / "Topology.json", "r") as json_file:
    topology = json.load(json_file)
```
```bash
TypeError: unsupported operand type(s) for /: 'str' and 'str'
```
