### 2024-11-23 UU-ER/AdOpT-NET0#333 @trevorb1
[JOSS]: Issue with `adopt.fill_carrier_data()` call in `In Short` section of Docs
In the [In Short](https://adopt-net0.readthedocs.io/en/latest/ModellingWorkflow.html#in-short) section of the documentation, we are told to define the demand data with the following commands:

```python
adopt.fill_carrier_data(input_data_path, value=0.01, columns=['Demand'], carriers=['electricity'], nodes=['node1'])
adopt.fill_carrier_data(input_data_path, value=100, columns=['Import price'], carriers=['electricity'], nodes=['node1'])
adopt.fill_carrier_data(input_data_path, value=1, columns=['Import limit'], carriers=['electricity'], nodes=['node1'])
```

However, I receive the following error.

```bash
TypeError: fill_carrier_data() got an unexpected keyword argument 'value'
```
