### 2024-11-23 #7402 @trevorb1
Hi @JeanWi; thanks so much for providing the fix here so I can move on with the review! I appreciate it. Regarding the figures, thank you for clarifying the license on them. If you and @AdamRJensen are okay with the current citing of the figures, thats works great for me :)

Also, apologies for this note of mine from the previous comment:

> With that said, I think a more structured tutorial page may help answer many of my concerns.

I completely missed seeing that three different case studies have already been set up and documented [here](https://adopt-net0.readthedocs.io/en/latest/CaseStudies.html). I only saw the `case_study_results` folder in the repo during my first look through. Totally my fault for missing that. My comments below include now walking through these case studies.

## Documentation

Please find below the remainder of my items. This list can be seen as an extension to `Documentation` - `Specific Issues` list from my previous comment:

### Specific Issues
- In [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) Step 4, it notes "For the carriers, whether or not curtailment of generic production is possible in `EnergybalanceOptions.JSON`". I found it difficult to location this file and would recommend putting exact location (`<data_path>/<investment_period>/node_data/<node>/carrier_data/EnergybalacneOptions.json`) or some other way to guide the user there.
-  In In [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data), for specifying time dependent data, I could not get the `adopt.load_climate_data_from_api(...)` function to run. Please see [this issue](https://github.com/UU-ER/AdOpT-NET0/issues/324). (Once running, Im excited to try this function, though! Having AdOpT-NETO handle climate data for users is a nice feature!)
- In In [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data), the instruction numbers do not carry over to time-dependent data (ie. it goes from `1.`, `2.`, `...` to just a paragraph. Its unclear if these steps are required.
- In [Define Model Configuration](https://adopt-net0.readthedocs.io/en/latest/workflow/DefineModelConfiguration.html#define-model-configuration), there is a broken table link. See this [issue](https://github.com/UU-ER/AdOpT-NET0/issues/325).
- Please see some general comments on Case Study 1 in [this issue ](https://github.com/UU-ER/AdOpT-NET0/issues/326)
- Case study 2 worked great for me!
- Please see a minor issue on case study 3 [here](https://github.com/UU-ER/AdOpT-NET0/issues/327). (Sidenote; this case study was very well presented and descriptive!)
- In the [Advanced Topics: Model Configuration](https://adopt-net0.readthedocs.io/en/latest/advanced_topics/model_configuration.html) section, it would be good to add what the different objective options are (ie. what are the differences between `[‘costs’, ‘emissions_pos’, ‘emissions_net’, ‘emissions_minC’, ‘costs_emissionlimit’, ‘pareto’]`. I do see that under some sections, they are mentioned (ie. for paerto studies is says to change the option to pareto). But its not necessarily clear what the difference between the emission scenarios are and how the interact with the `emission_limit` option (unless I just missed seeing where this is in the docs!)
- In the [In Short](https://adopt-net0.readthedocs.io/en/latest/ModellingWorkflow.html#in-short) section, there is an issue when specifying the path of the input data. Please see [this issue](https://github.com/UU-ER/AdOpT-NET0/issues/332)
- In the [In Short](https://adopt-net0.readthedocs.io/en/latest/ModellingWorkflow.html#in-short) section, there is an issue with the arguments in the `fill_carrier_data(...)` function. Please see [this issue](https://github.com/UU-ER/AdOpT-NET0/issues/333).

#### General Notes

- The case studies are great! Exactly what I was looking for to learn AdOpT-NET0. Just a suggestion (ie. not a required change), but highlighting these better in the docs would be beneficial. My first instinct was to go to the `getting started` section. However, having these examples to accompany the instructions in that section would have been great.
- The visualization dashboard is quite nice!
- Again, apologies for missing the case studies originally. I hope this didn't cause too much additional work on your side.

### Automated Tests
- Tests exist with automated CI workflow!
- One test fails for me; please see [this issue](https://github.com/UU-ER/AdOpT-NET0/issues/329).
- Please add that users must `pip install coverage` to run tests; please see [this issue](https://github.com/UU-ER/AdOpT-NET0/issues/328)
- The tests create a whole bunch of new files in the directory I run from? Do you experience this as well? If so, I would suggest either using temporary files, or deleting these files at the end of the tests. Please see [this issue](https://github.com/UU-ER/AdOpT-NET0/issues/330).
- If a significant portion of the code base is tested, I would suggest added a code coverage badge to showcase that!. However, I do understand that testing both optimization models and file management functions can be tricky, so converge may not necessarily be reflective of the work that has gone into this project. Therefore, this is just a suggestion (not a requirement) as I do see there is tests for main functionality, and the claimed functionality works for me.

### Community Guidelines
- A contributing guide exists that includes detailed instructions on how to contribute, code style, where to report bugs, ect.
- The only thing I would add is including at least one [issue template](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository). Issue template are quite easy to set up and can force contributors to give certain information you are looking for (ie. OS, package version, ect...). This way users will know exactly what you mean by `detailed description`. Please see [this issue](https://github.com/UU-ER/AdOpT-NET0/issues/331).

## Final Notes
- Well done to all the authors! This is a super cool project!
- If any of my comments require further explanation, please just let me know and I will elaborate.
- If you do not agree with any of my comments, please just let me know your concerns and we can discuss further.
