+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,29 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is, including possible error outputs.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Environment (please specify the environment you work in):**
+ - OS: [e.g. Windows 11]
+ - Python version: [e.g. Python 3.12]
+ - AdOpT-NET0 version [e.g. v0.1.5]
+- Python environment [produced by e.g. "pip freeze"] 
+
+**Additional context**
+Add any other context about the problem here.
