+++ .github/workflows/pin_requirements.yml
@@ -16,6 +16,7 @@ jobs:
       uses: actions/checkout@v4
       with:
         ref: ${{ github.head_ref }}
+        token: ${{ github.MY_TOKEN }}
 
     - name: Set up Python
       uses: actions/setup-python@v4
