+++ requirements.txt
@@ -7,14 +7,6 @@ h3==4.1.2 ; python_version >= "3.12" and python_version < "4"
 h5py==3.12.1 ; python_version >= "3.12" and python_version < "4.0"
 highspy==1.8.1 ; python_version >= "3.12" and python_version < "4.0"
 idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
-iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
-joblib==1.4.2 ; python_version >= "3.12" and python_version < "4.0"
-networkx==3.4.2 ; python_version >= "3.12" and python_version < "4.0"
-numpy==2.1.3 ; python_version >= "3.12" and python_version < "4.0"
-openpyxl==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
-packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
-pandas==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
-patsy==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
 pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
 ply==3.11 ; python_version >= "3.12" and python_version < "4.0"
 pvlib==0.11.1 ; python_version >= "3.12" and python_version < "4.0"
