### 2024-11-23 UU-ER/AdOpT-NET0#329 @JeanWi
Thanks for reporting! Indeed there was a typo in the technology name. Interestingly it never came to surface for us, as apparently the systems we use are case-insensitive. I assume you use a linux-based system?
(just learned about case-sensitive/insensitive systems with this :)).

https://stackoverflow.com/questions/6710511/case-sensitive-path-comparison-in-python
