### 2024-11-23 UU-ER/AdOpT-NET0#328 @JeanWi
We have changed the building and dependency management to poetry in #320, now also including test and development dependency groups - does this make sense to you?
