### 2024-11-23 c6a6741
Update paper.bib
