+++ joss_submission/paper.bib
@@ -32,9 +32,10 @@
 @article{tiggeloven2024chemicalcluster,
   title={Optimizing emissions reduction in ammonia-ethylene chemical clusters: synergistic integration of electrification, carbon capture, and hydrogen},
   author={Tiggeloven, Julia L. and Faaij, Andre P.C. and Kramer, Gert Jan and Gazzani, Matteo},
-  year={2024},
-  journal={forthcoming},
+  year={forthcoming},
+  journal={To be published},
   doi={}
+}
 
 @article{morales2017hidden,
   title={Hidden power system inflexibilities imposed by traditional unit commitment formulations},
