+++ .github/workflows/testing.yml
@@ -40,6 +40,6 @@ jobs:
         poetry run pytest --cov --cov-report=xml
 
     - name: Upload results to Codecov
-      uses: codecov/codecov-action@v4
+      uses: codecov/codecov-action@v5
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
