+++ tests/test_technologies.py
@@ -1,1151 +1,1151 @@
-import warnings
-
-import pytest
-from pathlib import Path
-from pyomo.environ import ConcreteModel, Set, Constraint, TerminationCondition
-import json
-import numpy as np
-
-from tests.utilities import (
-    make_climate_data,
-    make_data_for_testing,
-    run_model,
-)
-from adopt_net0.components.technologies.technology import Technology
-from adopt_net0.data_management.utilities import open_json, select_technology
-from adopt_net0.components.utilities import annualize
-from adopt_net0.components.utilities import perform_disjunct_relaxation
-
-
-def define_technology(
-    tec_name: str,
-    nr_timesteps: int,
-    load_path: Path,
-    perf_type: int = None,
-    CAPEX_model: int = None,
-):
-    """
-    Reads technology data and fits it
-
-    :param str tec_name: name of the technology.
-    :param int nr_timesteps: Number of timesteps to create climate data for
-    :param Path load_path: Path to load from
-    :param int perf_type: performance function type (for generic conversion tecs)
-    :param int CAPEX_model: capex model (1,2,3,4)
-    :return: Technology class
-    """
-    # Technology Class Creation
-    with open(load_path / (tec_name + ".json")) as json_file:
-        tec = json.load(json_file)
-    tec["name"] = tec_name
-
-    if perf_type:
-        tec["Performance"]["performance_function_type"] = perf_type
-    if CAPEX_model:
-        tec["Economics"]["CAPEX_model"] = CAPEX_model
-
-    tec = select_technology(tec)
-
-    # Technology fitting
-    climate_data = make_climate_data("2022-01-01 12:00", nr_timesteps)
-    location = {}
-    location["lon"] = 5.5
-    location["lat"] = 52.5
-    location["alt"] = 0
-    if tec.component_options.ccs_possible:
-        tec.ccs_data = open_json(tec.component_options.ccs_type, load_path)
-    tec.fit_technology_performance(climate_data, location)
-
-    return tec
-
-
-def construct_tec_model(tec, nr_timesteps: int, dynamics: int = None):
-    """
-    Construct a mock technology model for testing
-
-    :param Technology tec: Technology object.
-    :param int nr_timesteps: Number of timesteps to create climate data for
-    :param int dynamics: if dynamics should be used in mock model
-    :return ConcreteModel m: Pyomo Concrete Model
-    """
-
-    m = ConcreteModel()
-    m.set_t = Set(initialize=list(range(1, nr_timesteps + 1)))
-    m.set_t_full = Set(initialize=list(range(1, nr_timesteps + 1)))
-    data = make_data_for_testing(nr_timesteps)
-
-    if dynamics:
-        data["config"]["performance"]["dynamics"]["value"] = dynamics
-
-    m = tec.construct_tech_model(m, data, m.set_t, m.set_t_full)
-    if tec.big_m_transformation_required:
-        m = perform_disjunct_relaxation(m)
-
-    return m
-
-
-def generate_output_constraint(model, demand: list, output_ratios: dict = None):
-    """
-    Generate an output constraint of a technology model
-
-    :param model: pyomo model
-    :param list demand: list of demand values to use
-    :param output_ratios: output ratios to use
-    :return: pyomo model
-    """
-
-    def init_output_constraint(const, t, car):
-        if output_ratios:
-            if isinstance(output_ratios.get(car), dict):
-                alpha = output_ratios[car]["alpha1"]
-            else:
-                alpha = output_ratios[car]
-            if isinstance(alpha, list):
-                return model.var_output_tot[t, car] >= demand[t - 1] * alpha[0]
-            else:
-                return model.var_output_tot[t, car] == demand[t - 1] * alpha
-        else:
[…]
