+++ .github/workflows/testing.yml
@@ -42,4 +42,4 @@ jobs:
     - name: Upload results to Codecov
       uses: codecov/codecov-action@v5
       with:
-        token: ${{ secrets.CODECOV_TOKE }}
+        token: ${{ secrets.CODECOV_TOKEN }}
