+++ README.md
@@ -43,4 +43,5 @@ The package relies heavily on other python packages. Among others this package u
 ## Credits
 This tool was developed at Utrecht University.
 
-This is some new text
+I also changed some text here
+This is even more text
