### 2024-11-24 51a7575
Develop (#341)
* Fixed bug in network consumption

* initialization of pyo.Parameter with dict keys not allowed, changed to list

* Minor bugfixe:
result writing failed for scaled models if solution is infeasible (expected behavior: continue without writing results, actual problematic behavior: error when writing results)

* Minor bugfix:
create_carbonprice_parameter() in construct_nodes.py threw error for different pandas versions.

* Add possibility to skip writing results

* Add possibility to skip writing results

* Added escape characters to resolve SyntaxWarnings for Docstrings

* Added 
[…]
