### 2024-11-25 4374fa3
Raise exception if climate data import was unsuccessful
