### 2024-11-25 UU-ER/AdOpT-NET0#324 @JeanWi
I believe on your first attempt there was a server error on the JRC side. Our code would just respond then with a key error (not very user friendly). We just changed that. Thanks for spotting!
