### 2024-11-25 690d106
moved requirements back to test CI
