+++ requirements.txt
@@ -33,7 +33,7 @@ six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
 statsmodels==0.14.4 ; python_version >= "3.12" and python_version < "4.0"
 threadpoolctl==3.5.0 ; python_version >= "3.12" and python_version < "4.0"
 timezonefinder==6.5.5 ; python_version >= "3.12" and python_version < "4"
-tqdm==4.67.1 ; python_version >= "3.12" and python_version < "4.0"
+tqdm==4.67.0 ; python_version >= "3.12" and python_version < "4.0"
 tsam==2.3.5 ; python_version >= "3.12" and python_version < "4.0"
 tzdata==2024.2 ; python_version >= "3.12" and python_version < "4.0"
 urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
