+++ .github/workflows/pin_requirements.yml
@@ -15,7 +15,6 @@ jobs:
     - name: Checkout Code
       uses: actions/checkout@v4
       with:
-        ref: ${{ github.head_ref }}
         token: ${{ secrets.PAT_TOKEN }}
 
     - name: Set up Python
