+++ .github/workflows/pin_requirements.yml
@@ -15,6 +15,7 @@ jobs:
     - name: Checkout Code
       uses: actions/checkout@v4
       with:
+        ref: ${{ github.head_ref }}
         token: ${{ secrets.PAT_TOKEN }}
 
     - name: Set up Python
