+++ .github/workflows/testing.yml
@@ -40,6 +40,6 @@ jobs:
         poetry run pytest --cov --cov-report=xml
 
     - name: Upload coverage reports to Codecov with GitHub Action
-      uses: codecov/codecov-action@v5
-      env:
-        CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+      uses: codecov/codecov-action@v4
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
