+++ .github/workflows/testing.yml
@@ -39,7 +39,7 @@ jobs:
       run: |
         poetry run pytest --cov --cov-report=xml
 
-    - name: Upload coverage reports to Codecov with GitHub Action
+    - name: Upload results to Codecov
       uses: codecov/codecov-action@v4
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
