+++ .github/workflows/00develop.yml
@@ -1,14 +1,45 @@
-name: Develop
+name: Testing
 
-on: [push, pull_request]
+on:
+  push
 
 jobs:
-#  pin_requirements:
-#    uses: ./.github/workflows/pin_requirements.yml
-#  autoformatting:
-#    needs: pin_requirements
-#    uses: ./.github/workflows/autoformatting.yml
-  testing:
-#    needs: autoformatting
-    uses: ./.github/workflows/testing.yml
+  test:
+    runs-on: ${{ matrix.os }}
 
+    strategy:
+      matrix:
+        os: [windows-latest]
+        python-version: [3.12, 3.13]
+
+    steps:
+    - name: Install GLPK Solver
+      run: | 
+        choco install glpk
+
+    - name: Checkout Code
+      uses: actions/checkout@v4
+
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+
+    - name: Install Dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install poetry
+        poetry install --without dev
+
+    - name: Check directory
+      run: |
+        python -c "import os; print(os.listdir()); print(os.getcwd())"
+
+    - name: Test
+      run: |
+        poetry run pytest --cov --cov-report=xml
+
+    - name: Upload results to Codecov
+      uses: codecov/codecov-action@v4
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
