### 2024-11-26 cc44a0d
Fix CI being stuck after push to pin requirements
