+++ .github/workflows/pin_requirements.yml
@@ -16,6 +16,7 @@ jobs:
       uses: actions/checkout@v4
       with:
         ref: ${{ github.head_ref }}
+        token: ${{ secrets.PAT_TOKEN }}
 
     - name: Set up Python
       uses: actions/setup-python@v4
@@ -33,8 +34,7 @@ jobs:
 
     - name: Commit and push changes
       run: |
-        git config user.name "GitHub Actions"
-        git config user.email "actions@github.com"
+        git config user.email "AdOpT-Development@solisservices.onmicrosoft.com"
         git add requirements.txt
         git commit -m "Update requirements.txt generated by Poetry"
         git push
