+++ .github/workflows/00develop.yml
@@ -8,6 +8,7 @@ on:
 jobs:
   pin_requirements:
     uses: ./.github/workflows/pin_requirements.yml
+    secrets: inherit
   autoformatting:
     needs: pin_requirements
     uses: ./.github/workflows/autoformatting.yml
