### 2024-11-26 738dcfb
Fix CI being stuck after push to pin requirements
