+++ .github/workflows/autoformatting.yml
@@ -3,6 +3,11 @@ name: Autoformatting
 on:
   workflow_call:
 
+# Cancel any in-progress runs when a new run is triggered
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 jobs:
   lint:
     runs-on: ubuntu-latest
