+++ .github/workflows/00develop.yml
@@ -4,7 +4,6 @@ on:
   pull_request:
     branches:
       - 'develop'
-  push:
 
 jobs:
   pin_requirements:
