### 2024-11-26 5ac4e23
Improve network documentation
