+++ docs/source/index.rst
@@ -7,8 +7,9 @@ Welcome to AdOpT-NET0 documentation!
 =====================================
 Energyhub is a Python Library for bottom-up multi energy system modelling. It can model conversion technologies and
 networks for any carrier and optimize the design and operation of your energy system.
-We recommend to start with the :ref:`installation<installation>` and the
-:ref:`workflow for (first-time) users<workflow>`.
+We recommend to start with the :ref:`installation<installation>` the
+:ref:`workflow for (first-time) users<workflow>` and the
+:ref:`exemplary case studies <case-studies>`.
 
 Table of Content
 ==================
@@ -17,12 +18,11 @@ Table of Content
    :maxdepth: 2
 
    Installation
+   CaseStudies
    ModellingWorkflow
    AdvancedTopics
    SourceCode
    ForDevelopers
-   CaseStudies
-
 
 Indices and tables
 ==================
