### 2024-11-27 e55037f
Added case study link to the get started page+ fixed pytest
