+++ pyproject.toml
@@ -33,7 +33,6 @@ statsmodels = ">=0.14.1"
 pvlib = ">=0.10.4"
 scipy = ">=1.13.0"
 openpyxl = ">=3.1.2"
-pytest = ">=8.1.1"
 pwlf = ">=2.2.1"
 scandir = ">=1.10.0"
 
