+++ docs/source/index.rst
@@ -7,7 +7,7 @@ Welcome to AdOpT-NET0 documentation!
 =====================================
 Energyhub is a Python Library for bottom-up multi energy system modelling. It can model conversion technologies and
 networks for any carrier and optimize the design and operation of your energy system.
-We recommend to start with the :ref:`installation<installation>` the
+We recommend to start with the :ref:`installation<installation>`, the
 :ref:`workflow for (first-time) users<workflow>`, and the
 :ref:`exemplary case studies <case-studies>`.
 
