### 2024-11-27 e3d67cb
Fixed ccs_captured is smaller than var_ccs_size in technology.py
