+++ adopt_net0/components/technologies/technology.py
@@ -1388,7 +1388,25 @@ class Technology(ModelComponent):
         emissions_based_on = self.component_options.emissions_based_on
         capture_rate = coeff_ti["capture_rate"]
 
-        # TODO: maybe make the full set of all carriers as a intersection between this set and the others?
+        # Initialize the size of CCS as in _define_size (size given in mass flow of CO2 entering the CCS object)
+        b_tec.para_size_min_ccs = pyo.Param(
+            domain=pyo.NonNegativeReals,
+            initialize=self.ccs_component.input_parameters.size_min,
+            mutable=True,
+        )
+        b_tec.para_size_max_ccs = pyo.Param(
+            domain=pyo.NonNegativeReals,
+            initialize=self.ccs_component.input_parameters.size_max,
+            mutable=True,
+        )
+
+        # Size CCS
+        b_tec.var_size_ccs = pyo.Var(
+            within=pyo.NonNegativeReals,
+            bounds=(b_tec.para_size_min_ccs, b_tec.para_size_max_ccs),
+        )
+
+        # TODO: maybe make the full set of all carriers as an intersection between this set and the others?
         # Emission Factor
         b_tec.para_tec_emissionfactor = pyo.Param(
             domain=pyo.Reals,
@@ -1447,6 +1465,17 @@ class Technology(ModelComponent):
         b_tec.const_input_output_ccs = pyo.Constraint(
             self.set_t_global, rule=init_input_output_ccs
         )
+
+        def init_size_output_ccs(const, t):
+            return (
+                b_tec.var_output_ccs[t, "CO2captured"]
+                <= b_tec.var_size_ccs
+            )
+
+        b_tec.const_size_output_ccs = pyo.Constraint(
+            self.set_t_global, rule=init_size_output_ccs
+        )
+
 
         # Electricity and heat demand CCS
         def init_input_ccs(const, t, car):
