### 2024-11-28 f2f8e1e
Added json to test ccs copy_techonology_data, implemented copy_data function in data_loading.py for techs and networks
