+++ joss_submission/paper.bib
@@ -24,7 +24,7 @@
 @article{wiegner2024integration,
   title={Unleashing the full potential of the North Sea -- Identifying key energy infrastructure synergies for 2030 and 2040},
   author={Wiegner, Jan F and Gibescu, Madeleine and Gazzani, Matteo},
-  year={2024},
+  year={},
   journal={forthecoming},
   doi={10.48550/arXiv.2411.00540}
 }
@@ -32,8 +32,8 @@
 @article{tiggeloven2024chemicalcluster,
   title={Optimizing emissions reduction in ammonia-ethylene chemical clusters: synergistic integration of electrification, carbon capture, and hydrogen},
   author={Tiggeloven, Julia L. and Faaij, Andre P.C. and Kramer, Gert Jan and Gazzani, Matteo},
-  year={forthcoming},
-  journal={To be published},
+  year={},
+  journal={forthecoming},
   doi={}
 }
 
