### 2024-11-28 c353709
Added tree diagram for input data directory
