### 2024-11-28 5c448d1
Fixed figure
