### 2024-11-28 UU-ER/AdOpT-NET0#373 @JeanWi
@Luca: Check Rebuttal
Hi @trevorb1, thanks for this very thorough review. We are extremely grateful that someone outside of our research team took the time to try and test the package. Below, I reply to your comments that were not directly tackled with an issue.

***
> It would be good to include the link to the documentation site in the upper right hand side of the GitHub repository; under the "About" section.

Done 👍

***
> General Comments in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138

We worked on the structure of the documentation, highlighting the exemplary case studies and adding a tree diagram for clarity. We hope that this helps in understanding how to set up a user defined case study.

***
> Specific Comment in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138:
In [Defining System Topology](https://adopt-net0.readthedocs.io/en/latest/workflow/DefineTopology.html#defining-system-topology_) it is unclear where the .\data\technology_data folder is to reference for naming conventions. I see that in a later step this folder is created, but in the instructions its unclear what this folder is at this point in time.

We have updated and clarified the documentation respectively.

***
> Specific Comment in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138:
In [Defining System Topology](https://adopt-net0.readthedocs.io/en/latest/workflow/DefineTopology.html#defining-system-topology_) it states that investment periods are "a future feature, for now only single-period analysis is possible." This contradicts that paper that states that Multi-period studies are possible. Please clarify.

We have planned multi-period functionality in the future, but have only partly implemented it. Sorry for the confusion. We made a clarification in the paper (see https://github.com/UU-ER/AdOpT-NET0/pull/370)

***
> Specific Comments on networks in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138

We tried to tackle your comments on the networks in the following PRs: https://github.com/UU-ER/AdOpT-NET0/pull/359

***
> Comment on the software paper in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138:
Line 21-22: You state that mulit-energy systems "are highly complex but also offer synergies to reduce costs and environmental impact." This reads like a definitive statement that deserves a reference. (Not that I disagree with the statement at all! Just a single reference here would be good to defend the need for multi-energy modelling).

We have added a reference 👍

***
> Comment on the software paper in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138:
Table 1: Thanks for the comprehensive table! Can you please be more specific with the pareto front methods implemented. Similar to how you have given stochastic scenarios through "Monte Carlo" sampling, listing methods of pareto fronts would be good (ie. MGA, ɛ-constraint, ect.. following definitions from the Hoffmann paper)

Done 👍

***
> Comment on the software paper in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138:
Line 30-31: (Apologies in advance for this rather pedantic wording comment!) You note that AdOpT-NET0 "allows for a highly realistic assessment of individual technologies and their integration into an energy system". I have trouble with this wording when no evidence is provided (ie. a ref to a research paper / study) to show that improved technology representation correlates to more realistic technology investment/integration decisions. AdOpT-NET0 can describe technology specific operational constraints well (and closer represents reality), but this does not guarantee highly realistic integration answers over multiple time horizons. Can you please clarify how AdOpT-NET0's detailed technology representation translates into realistic assessments of energy integration questions. Or can you please soften the wording to align with a prescriptive analytics tool (ie. guide decision making) rather than a predictive analytics tool (ie. predicting what will happen in the future). To clarify, I have no issue with the statement of representing technologies more realistically (which you give references for later), but rather how this translates to more realistic future energy integration questions (which is likely difficult to measure?).

No worries at all, we agree that it the sentence didn't reflect the trade-off between detailed technology modeling and computational complexity. The section now reads: "A key feature of AdOpT-NET0 is its high level of technological detail, allowing for a comprehensive representation of individual technologies and their operational constraints. This detailed representation supports the exploration of technology integration into energy systems, enabling informed decision-making without limiting the scope of the analysis."

***
> Comment on the software paper in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138:
Line 32-33: (A manuscript suggestion that you are free to take or leave) You note that "several complexity reduction algorithms can be adopted to deal with infeasible computation times". I believe this is a super valuable point, and explicitly listing them for the reader is useful. ie. Something along the lines of "Complexity reduction algorithms for modelling seasonal storage (Gabrielli et al., 2018) and high VRE systems (Weimann & Gazzani, 2022) are implemented." Having these algorithms already integrated is a big plus of AdOpT-NET0 and would be good to highlight!

Good suggestion, thanks! 👍

***
> Comment on the software paper in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138:
Line 36: Its great that the closed MATLAB version has been ported to Python and made openly available. Is there any relevant publications or similar that use or describe the MATLAB version that require a referencing here? Or has it been used solely for in-house business applications?

Gabrielli et al. (2018), Weimann et al. (2023), Weimann et al. (2019), Weimann et al. (2022) and Wiegner et al. (2022) all used the Matlab version and we cited them at different places in the paper (but not explicitly in the section on MATLAB). The features and technologies were then migrated to Python. We didn's see any particular benefit of mentioning them specifically though (we also have a reference to P. Gabrielli and L. Weimann in the acknowledgements). If you have a strong preference to mention them explicitly, we can of course do so.

***
> Comment on the software paper in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138:
Statement of need paragraph: Listing the specific technologies you represent in greater detail compared to existing models is great, as it helps the reader understand the novelty of this framework! Where further clarification will help is with the target user. In the opening sentence you break down the field into complex NLP frameworks, and simplified LP frameworks. The users of this different frameworks may be different, though? (ie. policy focused research questions will use LP frameworks, while power flow research question (as you mention) may want the NLP detail). It is a little unclear to me if AdOpT-NET0 can be used both as a policy support tool and a power flow analysis tool, or just one or the other (or a category I am missing!).

We have added a sentence elaborating on the target audience in that section. See https://github.com/UU-ER/AdOpT-NET0/commit/0e9489d6e79578e44011c9049471b0b630b0c3ea

***
> Comment on the software paper in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2480848138:
Has AdOpT-NET0 been used in any publications (other than the forthcoming one by Wiegner et al.)? If so, please list them, else please just disregard this comment.

We have added a second forthcoming publication (Tiggeloven et al.). There are currently two other papers in the making, but they are not ready for submission yet.

***
>  Comment in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2495339140:
In [Define Input Data](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html#define-input-data) Step 4, it notes "For the carriers, whether or not curtailment of generic production is possible in EnergybalanceOptions.JSON". I found it difficult to location this file and would recommend putting exact location (<data_path>/<investment_period>/node_data/<node>/carrier_data/EnergybalacneOptions.json) or some other way to guide the user there.

We have added a structure of the [tree for the input data](https://adopt-net0.readthedocs.io/en/latest/src_code/DataPreparation.html), where the EnergybalanceOptions.json file also appears. We hope this makes it more clear.

***
> Comment in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2495339140:
In the [Advanced Topics: Model Configuration](https://adopt-net0.readthedocs.io/en/latest/advanced_topics/model_configuration.html) section, it would be good to add what the different objective options are (ie. what are the differences between [‘costs’, ‘emissions_pos’, ‘emissions_net’, ‘emissions_minC’, ‘costs_emissionlimit’, ‘pareto’]. I do see that under some sections, they are mentioned (ie. for paerto studies is says to change the option to pareto). But its not necessarily clear what the difference between the emission scenarios are and how the interact with the emission_limit option (unless I just missed seeing where this is in the docs!)

Good point, thanks! We have included the documentation on the objective functions.

***
> Comment in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2495339140:
The case studies are great! Exactly what I was looking for to learn AdOpT-NET0. Just a suggestion (ie. not a required change), but highlighting these better in the docs would be beneficial. My first instinct was to go to the getting started section. However, having these examples to accompany the instructions in that section would have been great.

Good point, done 👍

***
> Comment in https://github.com/openjournals/joss-reviews/issues/7402#issuecomment-2495339140:
If a significant portion of the code base is tested, I would suggest added a code coverage badge to showcase that!. However, I do understand that testing both optimization models and file management functions can be tricky, so converge may not necessarily be reflective of the work that has gone into this project. Therefore, this is just a suggestion (not a requirement) as I do see there is tests for main functionality, and the claimed functionality works for me.

We have added a codecov integration, that was a long-due change anyways, thanks for triggering it! 👍
