### 2024-11-28 ddc9557
Updated publish_tests
