+++ .github/workflows/00publish_tests.yml
@@ -13,3 +13,4 @@ jobs:
   publish:
     needs: testing
     uses: ./.github/workflows/uploadpypi.yml
+    secrets: inherit
