+++ .github/workflows/00publish_tests.yml
@@ -10,7 +10,56 @@ jobs:
   testing:
     needs: autoformatting
     uses: ./.github/workflows/testing.yml
-  publish:
+  build:
+    name: Build distribution
     needs: testing
-    uses: ./.github/workflows/uploadpypi.yml
-    secrets: inherit
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Create VERSION file
+        run: |
+          version=${GITHUB_REF#refs/*/}
+          echo "${version//v/}" > version.py
+      - name: Check VERSION file
+        run: cat version.py
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
+
+      - name: Install Dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install poetry
+          poetry self add "poetry-dynamic-versioning[plugin]"
+          poetry install --without test,dev
+
+      - name: Build a binary wheel and a source tarball
+        run: poetry build
+      - name: Store the distribution packages
+        uses: actions/upload-artifact@v4
+        with:
+          name: python-package-distributions
+          path: dist/
+
+  publish-to-pypi:
+    name: Upload release to PyPI
+    needs:
+      - build
+    runs-on: ubuntu-latest
+    environment:
+      name: pypi
+      url: https://pypi.org/p//adopt-net0
+    permissions:
+      id-token: write
+    steps:
+      - name: Download all the dists
+        uses: actions/download-artifact@v4
+        with:
+          name: python-package-distributions
+          path: dist/
+      - name: Publish package distributions to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
