+++ .github/workflows/00publish_tests.yml
@@ -18,6 +18,8 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v4
+        with:
+          python-version: 3.12
       - name: Install Dependencies
         run: |
           python -m pip install --upgrade pip
