+++ .github/workflows/00publish_tests.yml
@@ -1,8 +1,8 @@
 name: PublishTests
 
-on: [push]
-#  release:
-#    types: [published]
+on:
+  release:
+    types: [published]
 
 jobs:
 #  autoformatting:
