+++ .github/workflows/00publish_tests.yml
@@ -10,23 +10,7 @@ jobs:
 #  testing:
 #    needs: autoformatting
 #    uses: ./.github/workflows/testing.yml
-  build:
-    name: Build distribution
+  publish:
 #    needs: testing
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-python@v4
-        with:
-          python-version: 3.12
-      - name: Install Dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install poetry
-          poetry install --without dev
-      - name: Publish package
-        run: |
-          poetry config repositories.testpypi https://test.pypi.org/legacy/
-          poetry config pypi-token.testpypi ${{ secrets.TEST_PYPI_API_TOKEN }}
-          poetry publish --build -r testpypi
+    uses: ./.github/workflows/uploadpypi.yml
+    secrets: inherit
