+++ .github/workflows/uploadpypi.yml
@@ -1,7 +1,24 @@
-#name: Upload Python Package
-#
-#on:
-#  workflow_call:
-#
-#jobs:
-#
+name: Upload Python Package
+
+on:
+  workflow_call:
+
+jobs:
+  publish-to-pypi:
+    name: Upload release to PyPI
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v4
+        with:
+          python-version: 3.12
+      - name: Install Dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install poetry
+          poetry install --without dev
+      - name: Publish package
+        run: |
+          poetry config repositories.testpypi https://test.pypi.org/legacy/
+          poetry config pypi-token.testpypi ${{ secrets.TEST_PYPI_API_TOKEN }}
+          poetry publish --build -r testpypi
