+++ .github/workflows/uploadpypi.yml
@@ -9,14 +9,25 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - uses: actions/setup-python@v4
+
+      - name: Create VERSION file
+        run: |
+          version=${GITHUB_REF#refs/*/}
+          echo "${version//v/}" > version.py
+      - name: Check VERSION file
+        run: cat version.py
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
         with:
-          python-version: 3.12
+          python-version: "3.12"
+
       - name: Install Dependencies
         run: |
           python -m pip install --upgrade pip
           pip install poetry
-          poetry install --without dev
+          poetry self add "poetry-dynamic-versioning[plugin]"
+          poetry install --without test,dev
       - name: Publish package
         run: |
           poetry config repositories.testpypi https://test.pypi.org/legacy/
