+++ .github/workflows/uploadpypi.yml
@@ -31,5 +31,5 @@ jobs:
       - name: Publish package
         run: |
           poetry config repositories.testpypi https://test.pypi.org/legacy/
-          poetry config pypi-token.testpypi ${{ secrets.TEST_PYPI_API_TOKEN }}
+          poetry config pypi-token.testpypi ${{ secrets.PYPI_API_TOKEN }}
           poetry publish --build -r testpypi
