+++ .github/workflows/00publish_tests.yml
@@ -5,12 +5,12 @@ on:
     types: [published]
 
 jobs:
-#  autoformatting:
-#    uses: ./.github/workflows/autoformatting.yml
-#  testing:
-#    needs: autoformatting
-#    uses: ./.github/workflows/testing.yml
+  autoformatting:
+    uses: ./.github/workflows/autoformatting.yml
+  testing:
+    needs: autoformatting
+    uses: ./.github/workflows/testing.yml
   publish:
-#    needs: testing
+    needs: testing
     uses: ./.github/workflows/uploadpypi.yml
     secrets: inherit
