+++ .github/workflows/00deploy.yml
@@ -1,4 +1,4 @@
-name: Deploy
+name: Testing
 
 on:
   pull_request:
