+++ joss_submission/paper.bib
@@ -136,9 +136,21 @@ year = {2024}
 @article{mancarella2014mes,
   title={MES (multi-energy systems): An overview of concepts and evaluation models},
   author={Mancarella, Pierluigi},
+  doi = {10.1016/j.energy.2013.10.041},
   journal={Energy},
   volume={65},
   pages={1--17},
   year={2014},
   publisher={Elsevier}
+}
+
+@article{Gabrielli2020a,
+author = {Gabrielli, Paolo and Poluzzi, Alessandro and Kramer, Gert Jan and Spiers, Christopher and Mazzotti, Marco and Gazzani, Matteo},
+doi = {10.1016/j.rser.2019.109629},
+journal = {Renewable and Sustainable Energy Reviews},
+month = {apr},
+publisher = {Elsevier Ltd},
+title = {{Seasonal energy storage for zero-emissions multi-energy systems via underground hydrogen storage}},
+volume = {121},
+year = {2020}
 }
