### 2024-12-02 UU-ER/AdOpT-NET0#328 @trevorb1
Makes sense to me! Thanks @JeanWi!
