### 2024-12-02 UU-ER/AdOpT-NET0#379 @trevorb1
[JOSS] Gurobipy not installed 
**Describe the bug**
When I solve the model, I get an `ApplicationError` with Gurobipy not installed.

**To Reproduce**
Follow the getting started section until you arrive at the `Model construction and solving` section.

**Expected behavior**
The model should solve.

**Screenshots**
```bash
(adopt) $ pip show adopt_net0
Name: adopt_net0
Version: 0.1.5
Summary: A python package for multi energy system modelling
Home-page: https://github.com/UU-ER/AdOpT-NET0
Author: Jan F. Wiegner
Author-email: j.f.wiegner@uu.nl
License: MIT
Location: /home/trevor/miniforge3/envs/adopt/lib/python3.13/site-packages
Requires: numpy, openpyxl, pandas, pvlib, pwlf, Pyomo, requests, scandir, scipy, statsmodels, timezonefinder, tsam
Required-by:
```

```python
m = adopt.ModelHub()
m.read_data("<path_to_data>", start_period=None, end_period=None)
m.quick_solve()
```

```bash
--- Reading in data ---
Input data folder has been checked successfully - no errors occurred.
Reading data from data
Topology read successfully
Model Configuration read successfully
Time series read successfully
Node Locations read successfully
Energy balance options read successfully
/home/trevor/miniforge3/envs/adopt/lib/python3.13/site-packages/adopt_net0/components/technologies/genericTechnologies/res.py:202: UserWarning: TurbineName not in csv, standard WindTurbine_Onshore_1500 selected.
  warnings.warn(
/home/trevor/miniforge3/envs/adopt/lib/python3.13/site-packages/adopt_net0/components/technologies/genericTechnologies/res.py:202: UserWarning: TurbineName not in csv, standard WindTurbine_Onshore_1500 selected.
  warnings.warn(
Technology data read successfully
Network data read successfully
The gurobipy module (an optional Pyomo dependency) failed to import: ModuleNotFoundError: No module named 'gurobipy'
```

**Environment (please specify the environment you work in):**
 - OS: Linux
 - Python version: Python 3.13
 - AdOpT-NET0 version: 0.1.5
- Python environment:

```
Package                   Version
------------------------- --------------
adopt_net0                0.1.5
anyio                     4.6.2.post1
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 2.4.1
async-lru                 2.0.4
attrs                     24.2.0
babel                     2.16.0
beautifulsoup4            4.12.3
bleach                    6.2.0
Brotli                    1.1.0
cached-property           1.5.2
certifi                   2024.8.30
cffi                      1.17.1
charset-normalizer        3.4.0
comm                      0.2.2
debugpy                   1.8.9
decorator                 5.1.1
defusedxml                0.7.1
entrypoints               0.4
et_xmlfile                2.0.0
exceptiongroup            1.2.2
executing                 2.1.0
fastjsonschema            2.21.0
fqdn                      1.5.1
h11                       0.14.0
h2                        4.1.0
h3                        4.1.2
h5py                      3.12.1
highspy                   1.8.1
hpack                     4.0.0
httpcore                  1.0.7
httpx                     0.28.0
hyperframe                6.0.1
idna                      3.10
importlib_metadata        8.5.0
importlib_resources       6.4.5
ipykernel                 6.29.5
ipython                   8.30.0
ipython_genutils          0.2.0
ipywidgets                8.1.5
isoduration               20.11.0
jedi                      0.19.2
Jinja2                    3.1.4
joblib                    1.4.2
json5                     0.10.0
jsonpointer               3.0.0
jsonschema                4.23.0
jsonschema-specifications 2024.10.1
jupyter                   1.1.1
jupyter_client            7.4.9
jupyter-console           6.6.3
jupyter_core              5.7.2
jupyter-events            0.10.0
jupyter-lsp               2.2.5
jupyter_server            2.14.2
jupyter_server_terminals  0.5.3
jupyterlab                4.3.1
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
jupyterlab_widgets        3.0.13
MarkupSafe                3.0.2
matplotlib-inline         0.1.7
mistune                   3.0.2
nbclassic                 1.1.0
nbclient                  0.10.1
nbconvert                 7.16.4
nbformat                  5.10.4
nest_asyncio              1.6.0
networkx                  3.4.2
notebook                  6.5.7
notebook_shim             0.2.4
numpy                     2.1.3
openpyxl                  3.1.5
overrides                 7.7.0
packaging                 24.2
pandas                    2.2.3
pandocfilters             1.5.0
parso                     0.8.4
patsy                     1.0.1
pexpect                   4.9.0
pickleshare               0.7.5
pip                       24.3.1
pkgutil_resolve_name      1.3.10
platformdirs              4.3.6
ply                       3.11
prometheus_client         0.21.0
prompt_toolkit            3.0.48
psutil                    6.1.0
ptyprocess                0.7.0
pure_eval                 0.2.3
pvlib                     0.11.1
pwlf                      2.3.0
pycparser                 2.22
pyDOE                     0.3.8
Pygments                  2.18.0
Pyomo                     6.8.2
PySocks                   1.7.1
python-dateutil           2.9.0.post0
python-json-logger        2.0.7
pytz                      2024.2
PyYAML                    6.0.2
pyzmq                     26.2.0
referencing               0.35.1
requests                  2.32.3
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.21.0
scandir                   1.10.0
scikit-learn              1.5.2
scipy                     1.14.1
Send2Trash                1.8.3
setuptools                75.6.0
six                       1.16.0
sniffio                   1.3.1
soupsieve                 2.5
stack-data                0.6.2
statsmodels               0.14.4
terminado                 0.18.1
threadpoolctl             3.5.0
timezonefinder            6.5.5
tinycss2                  1.4.0
tomli                     2.1.0
tornado                   6.4.2
tqdm                      4.67.1
traitlets                 5.14.3
tsam                      2.3.5
types-python-dateutil     2.9.0.20241003
typing_extensions         4.12.2
typing-utils              0.1.0
tzdata                    2024.2
uri-template              1.3.0
urllib3                   2.2.3
wcwidth                   0.2.13
webcolors                 24.8.0
webencodings              0.5.1
websocket-client          1.8.0
widgetsnbextension        4.0.13
zipp                      3.21.0
zstandard                 0.23.0

```

**Additional context**
Add any other context about the problem here.
