### 2024-12-02 UU-ER/AdOpT-NET0#379 @trevorb1
I also get this error if I try to install it:

```
$ pip install gurobipy
ERROR: Could not find a version that satisfies the requirement gurobipy (from versions: none)
ERROR: No matching distribution found for gurobipy
```
