### 2024-12-03 #7402 @trevorb1
Hi @JeanWi; thanks for the detailed reply! Please find below my remaining notes.

@AdamRJensen I have updated my checklist! Just one box is left regarding my comments below.

## Functionality
- Installation now works for me!
- All functional claims have been defended!

## Documentation
- Thank you for addressing the majority of my comments! Sorry to be a pain, but I have a few minor points left.
- Please see [this issue](https://github.com/UU-ER/AdOpT-NET0/issues/379) with the gurobipy install
- I still think clarifying what data is used by the geolocated points vs. manual distances is required. I see the note of "the distance needs to be added manually, even if the node locations were specified previously" but this does not clarify what data is being pulled from where? Geolocations are only used for renewable potentials?
- Clarification on the differences between different networks still needs to be improved, I feel. For example, if you look in the docs it reads for Networks:

> Networks that can be used in two directions (e.g. electricity cables), are called bidirectional and are treated respectively with their size and costs. Other networks, e.g. pipelines, require two installations to be able to transport in two directions. As such their CAPEX is double and their size in both directions can be different.

Therefore, I am assuming there is some underlying functional/constraint difference between the networks, but I can't find these differences? Moreover, we also set connections between nodes in the `connections.csv`?

I guess I main point on this is that Im still confused on the differences between the networks. For example, I don't know where to look to see the differences between `electricityOffshore`, `electrcityOnshore`, and `electricitySimple` and in what cases I should use which? (Unless I have just missed seeing this in the docs!). Happy to further discuss this point though.

- Other than that, all points are addressed!

## Software Paper
- Missing DOI on the Mancarella, P. (2014) reference
- Missing Year on second Wiegner reference
- On the note of the MATLAB reference, thanks for the clarification! imo the sentence "It combines 5+ years of research and is inspired by a closed-source MATLAB version of the model." reads like the reference should be included. However, maybe @AdamRJensen can weigh in on if its required or not (since its cited in your other references) as Im not sure on the full scope of citation rules for software.
- All items in "Software Paper" are now complete!
