### 2024-12-03 d9a8630
Updated DOIs/ years
