+++ joss_submission/paper.bib
@@ -24,16 +24,16 @@
 @article{wiegner2024integration,
   title={Unleashing the full potential of the North Sea -- Identifying key energy infrastructure synergies for 2030 and 2040},
   author={Wiegner, Jan F and Gibescu, Madeleine and Gazzani, Matteo},
-  year={},
-  journal={forthcoming},
+  year={forthcoming},
+  journal={},
   doi={10.48550/arXiv.2411.00540}
 }
 
 @article{tiggeloven2024chemicalcluster,
   title={Optimizing emissions reduction in ammonia-ethylene chemical clusters: synergistic integration of electrification, carbon capture, and hydrogen},
   author={Tiggeloven, Julia L. and Faaij, Andre P.C. and Kramer, Gert Jan and Gazzani, Matteo},
-  year={},
-  journal={forthcoming},
+  year={forthcoming},
+  journal={},
   doi={}
 }
 
