### 2024-12-03 eaef50a
Throw error if NodeLocations are not specified
