### 2024-12-03 63ec5aa
Updated network docu
