### 2024-12-03 ce92952
Update test dependencies
