### 2024-12-03 cac3dc5
Update test dependencies (nbconvert)
