+++ pyproject.toml
@@ -40,6 +40,7 @@ scandir = ">=1.10.0"
 pytest = ">=8.1.1"
 pytest-cov = ">= 6.0.0"
 nbformat = ">= 5.10.4"
+nbconvert = ">= 5.4.0"
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = ">=3.7.0"
