### 2024-12-03 fd520f3
Update NodeLocations for tests
