### 2024-12-03 ed1529e
Renamed allow_one_direction to bidirectional for networks
