+++ adopt_net0/components/networks/network.py
@@ -135,7 +135,7 @@ class Network(ModelComponent):
     This part calculates variables for all respective nodes.
 
     - CAPEX calculation of the whole network as a sum of CAPEX of all arcs. If
-      ``allow_only_one_direction`` is set to 1 for this network, the capex and fixed
+      ``bidirectional_network`` is set to 1 for this network, the capex and fixed
       opex for an arc is only counted once.
 
     - OPEX fix, as fraction of total CAPEX
@@ -154,9 +154,9 @@ class Network(ModelComponent):
 
     - Energy consumption of other carriers at each node.
 
-    - If  ``allow_only_one_direction`` is set to 1 for this network only additional
+    - If  ``bidirectional_network`` is set to 1 for this network only additional
       constraints are enforced to ensure that at each time step a flow can only be in
-      one direction. For ``allow_only_one_direction_precise = 0``, only a cut and a
+      one direction. For ``bidirectional_network_precise = 0``, only a cut and a
       cosntraint on the sizes of the two directions of an arc are formulated:
 
       .. math::
@@ -165,7 +165,7 @@ class Network(ModelComponent):
       .. math::
         flow_{nodeFrom, nodeTo} + flow_{nodeTo, nodeFrom} \\lor S_{nodeTo, nodeFrom}
 
-      For ``allow_only_one_direction_precise = 1`` additional disjunctions are
+      For ``bidirectional_network_precise = 1`` additional disjunctions are
       formulated, thus adding binaries and complexity to the model:
 
       .. math::
@@ -279,7 +279,7 @@ class Network(ModelComponent):
 
         b_netw = self._define_possible_arcs(b_netw)
 
-        if self.component_options.allow_only_one_direction:
+        if self.component_options.bidirectional_network:
             b_netw = self._define_unique_arcs(b_netw)
 
         b_netw = self._define_size(b_netw)
@@ -321,7 +321,7 @@ class Network(ModelComponent):
         b_netw.arc_block = pyo.Block(b_netw.set_arcs, rule=arc_block_init)
 
         # CONSTRAINTS FOR BIDIRECTIONAL NETWORKS
-        if self.component_options.allow_only_one_direction:
+        if self.component_options.bidirectional_network:
             b_netw = self._define_bidirectional_constraints(b_netw)
 
         b_netw = self._define_capex_total(b_netw)
@@ -416,7 +416,7 @@ class Network(ModelComponent):
                 initialize=init_size_initial,
             )
             # Check if sizes in both direction are the same for bidirectional existing networks
-            if self.component_options.allow_only_one_direction:
+            if self.component_options.bidirectional_network:
                 for from_node in coeff_ti["size_initial"]:
                     for to_node in coeff_ti["size_initial"][from_node].index:
                         assert (
@@ -966,7 +966,7 @@ class Network(ModelComponent):
         )
 
         # Disjunction
-        if self.component_options.allow_only_one_direction_precise:
+        if self.component_options.bidirectional_network_precise:
             self.big_m_transformation_required = 1
 
             def init_bidirectional(dis, t, node_from, node_to, ind):
@@ -1011,7 +1011,7 @@ class Network(ModelComponent):
         :param b_netw: pyomo network block
         :return: pyomo network block
         """
-        if self.component_options.allow_only_one_direction:
+        if self.component_options.bidirectional_network:
             arc_set = b_netw.set_arcs_unique
         else:
             arc_set = b_netw.set_arcs
@@ -1033,7 +1033,7 @@ class Network(ModelComponent):
         :param b_netw: pyomo network block
         :return: pyomo network block
         """
-        if self.component_options.allow_only_one_direction:
+        if self.component_options.bidirectional_network:
             arc_set = b_netw.set_arcs_unique
         else:
             arc_set = b_netw.set_arcs
