### 2024-12-03 #7402 @JeanWi
Hi @trevorb1, thanks checking everything! We really appreciate all the effort you went through and spot all the unclarities and inconsistencies that come with being a bit lazy with the documentation sometimes...

All changes have been implemented in [this PR](https://github.com/UU-ER/AdOpT-NET0/pull/380) and are part of v0.1.6 of AdOpT. Find below the answers to your remaining comments.

> Please see https://github.com/UU-ER/AdOpT-NET0/issues/379 with the gurobipy install

This is a very annoying problem, we answered in the issue itself. Thanks for highlighting! We hope it was solved to your satisfaction.

> I still think clarifying what data is used by the geolocated points vs. manual distances is required. I see the note of "the distance needs to be added manually, even if the node locations were specified previously" but this does not clarify what data is being pulled from where? Geolocations are only used for renewable potentials?

To tackle this, we have (1) made the specification of node locations a requirement and (2) added the following description in the [documentation](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html):  "Climate data is loaded from this data. Additionally it is used to calculate the position of the sun for PV    modelling (using pvlib). Note that the distance between nodes is not based on the provided locations."

> Clarification on the differences between different networks still needs to be improved, I feel.

We also added to the documentation the following information to the [documentation](https://adopt-net0.readthedocs.io/en/latest/workflow/LoadData.html): "The technologies and networks shipped with AdOpT-NET0 can be seen as templates and the performance and cost parameters as well as technology specific options can be further modified by the user. Therefore, you can modify the json files in the input data folder of your case study after they have been copied. Also it is possible to specify technologies or networks that are not provided in AdOpT by using the models defined in the network class or the technology classes."

Additionally we now provide more information on bi-directional networks in the [network documentation](https://adopt-net0.readthedocs.io/en/latest/src_code/model_components/networks.html#networks).

> Software Paper

We have fixed the issues in the latest commit to the joss_submission branch. The two papers mentioned are forthcoming, so we did not specify a year. @AdamRJensen: could you produce a new article proof? Also let us know if you would like to see the MATLAB papers cited in the respective section.
