### 2024-12-03 UU-ER/AdOpT-NET0#379 @trevorb1
Thanks for the clarification, @JeanWi! Apologies for missing the note on the pyomo solver. I do understand not wanting gurobipy to be a dependency on the project.

Il close this issue for now, as to not hold up the review any longer! A suggestion would to be clearer on the [solving page](https://adopt-net0.readthedocs.io/en/latest/workflow/SolveModel.html#model-construction-and-solving) on how to download the pyomo supported solvers (could even be just linking to pyomos documentation on this).
