### 2024-12-03 #7402 @trevorb1
Thanks for the quick fixes, @JeanWi!
- `gurobipy` issue is solved.
- Node location issue solved.
- The documentation for networks has improved. However, it still remains a little vague to me on the purpose of pre-defined networks when parameters are the same between them and users are free to add new networks. With that said, I don't want to hold up the review on this single point that may just be a wording issue and would sort itself out if using AdOpT-NET0 for a large model (plus the other reviewer did not have an issue with wording). So I am fine to mark this point as complete :)
