### 2024-12-03 #7402 @trevorb1
@AdamRJensen; all items in my checklist are now complete, and I recommend to accept AdOpT-NET0!

Well done to @JeanWi and the other co-authors! Thank you for your detailed feedback to all of my comments!

Thank you to @AdamRJensen for facilitating the review process!
