### 2024-12-04 #7402 @JeanWi
> * The documentation for networks has improved. However, it still remains a little vague to me on the purpose of pre-defined networks when parameters are the same between them and users are free to add new networks. With that said, I don't want to hold up the review on this single point that may just be a wording issue and would sort itself out if using AdOpT-NET0 for a large model (plus the other reviewer did not have an issue with wording). So I am fine to mark this point as complete :)

I agree with you that this can be a bit confusing for new users and I have to admit that we also did not update the data very well. When we use the model, typically we specify the data per case study. But our long-term vision is to have a database of technology data connected to AdOpT-NET0 with respective financial and technical data.
We discussed this issue internally and came up with two options: (1) deleting the network json files, keeping just one to use as a template or (2) keeping it as it is and allowing for a wider range of network templates but requiring the user to specify the data for each case study individually. Both options have their pros and cons (of course we have a preference on leaving it as it is though :D).
