### 2024-12-09 #7402 @JeanWi
@AdamRJensen: Do you need anything from our side still? To our knowledge we addressed all comments, it only remains how you prefer previous work has been cited. Maybe you are on vacation, so I do not want to pressure you, but just wanted to make sure that we didn't miss anything.
