### 2024-12-10 3bafe28
added dependencies citation
