### 2024-12-10 f683b58
Updated dois and removed urls
