+++ joss_submission/paper.bib
@@ -111,10 +111,8 @@ pages = {120738},
 year = {2023},
 issn = {0306-2619},
 doi = {10.1016/j.apenergy.2023.120738},
-url = {https://www.sciencedirect.com/science/article/pii/S0306261923001022},
 author = {Lukas Weimann and Guus Dubbink and Louis {van der Ham} and Matteo Gazzani},
 keywords = {Carbon capture, MEA, Energy system, Optimisation, MILP},
-abstract = {Assessing the role of carbon capture in energy systems dominated by non-dispatchable renewable energy sources requires a reliable and accurate model. However, carbon capture models used in complex systems optimisation are often very simplified. Therefore, we developed a mixed-integer linear model of post-combustion carbon capture starting from rigorous thermodynamic modelling in Aspen Plus. The final model decides the size and the operation of the capture process and returns the cost and energy requirements as a function of the CO2 concentration and the flow rate of the treated flue gas. Validation against actual plant data (Petra Nova) showed excellent accuracy with a deviation in total CO2 captured of just 2%. By applying the model to an exemplary case study, we show that it allows for co-optimising renewables deployment and carbon capture design and operation for a gas turbine, thus opening opportunities to explore new system designs of practical added value.}
 }
 
 
@@ -128,7 +126,6 @@ month = {dec},
 pages = {100190},
 publisher = {Elsevier},
 title = {{A review of mixed-integer linear formulations for framework-based energy system models}},
-url = {https://linkinghub.elsevier.com/retrieve/pii/S2666792424000283},
 volume = {16},
 year = {2024}
 }
@@ -163,11 +160,11 @@ edition={Third},
 volume={67},
 year={2021},
 publisher={Springer Science \& Business Media}
+doi = {10.1007/978-3-030-68928-5}
 }
 
 @article{Anderson2023pvlib,
 doi = {10.21105/joss.05994},
-url = {https://doi.org/10.21105/joss.05994},
 year = {2023},
 publisher = {The Open Journal},
 volume = {8},
@@ -184,9 +181,7 @@ volume = {315},
 pages = {119029},
 year = {2022},
 issn = {0306-2619},
-doi = {https://doi.org/10.1016/j.apenergy.2022.119029},
-url = {https://www.sciencedirect.com/science/article/pii/S0306261922004342},
+doi = {10.1016/j.apenergy.2022.119029},
 author = {Maximilian Hoffmann and Leander Kotzur and Detlef Stolten},
 keywords = {Typical days, Segmentation, Time slices, Energy system models, Time series aggregation, Temporal aggregation, Temporal resolution, Renewable energy systems, Clustering, Computation time, Pareto optimality},
[…]
