### 2024-12-10 a9a6b04
Updated dois and removed urls
