+++ joss_submission/paper.bib
@@ -159,7 +159,7 @@ author={Bynum, Michael L. and Hackebeil, Gabriel A. and Hart, William E. and Lai
 edition={Third},
 volume={67},
 year={2021},
-publisher={Springer Science \& Business Media}
+publisher={Springer Science \& Business Media},
 doi = {10.1007/978-3-030-68928-5}
 }
 
