### 2024-12-10 7474989
Fixing Tiggeloven et al. citation
