+++ joss_submission/paper.bib
@@ -31,7 +31,7 @@
 
 @article{tiggeloven2024chemicalcluster,
   title={Optimizing emissions reduction in ammonia-ethylene chemical clusters: synergistic integration of electrification, carbon capture, and hydrogen},
-  author={Tiggeloven, Julia L. and Faaij, Andre P.C. and Kramer, Gert Jan and Gazzani, Matteo},
+  author={Tiggeloven, Julia L and Faaij, Andre P.C. and Kramer, Gert Jan and Gazzani, Matteo},
   year={forthcoming},
   journal={},
   doi={}
@@ -92,7 +92,7 @@
 
 @article{tiggeloven2023optimization,
   title={Optimization of Electric Ethylene Production: Exploring the Role of Cracker Flexibility, Batteries, and Renewable Energy Integration},
-  author={Tiggeloven, J. L. and Faaij, A. P. C. and Kramer, G. J. and Gazzani, M.},
+  author={Tiggeloven, Julia L and Faaij, A. P. C. and Kramer, G. J. and Gazzani, M.},
   journal={Industrial and Engineering Chemistry Research},
   volume={62},
   number={40},
