### 2024-12-10 7e74188
Updated calculation function for dni (+ adding a functional test)
