### 2024-12-16 #7402 @JeanWi
> @JeanWi At this point could you:
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [ ]  Please list the DOI of the archived version here.
>
> I can then move forward with recommending acceptance of the submission.

@AdamRJensen: Thanks!
I did not make a new tag, as there were no new changes. The version is 0.1.7 and is archived [here](https://zenodo.org/records/14361112), including the respective meta data (authors and title). The doi of that version is: 10.5281/zenodo.14361112.

Is that ok, or would you like to have a new release/tag?

Currently the branch with the joss paper is out of date with the main branch, is this ok? We can also merge the main branch into the joss_submission branch, so it is up to date.
