### 2024-12-16 #7402 @AdamRJensen
> Is that ok, or would you like to have a new release/tag?
>
> Currently the branch with the joss paper is out of date with the main branch, is this ok? We can also merge the main branch into the joss_submission branch, so it is up to date.

I think all those things are fine.
