### 2024-12-17 #7402 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
