### 2025-01-02 #7402 @kyleniemeyer
Hello @JeanWi, before accepting, I just noticed one minor issue in the paper: you have two citations for things with the years given as "forthcoming". This isn't really a standard way of citing in-press publications; normally, the current year (or year of the status) is given, then a note of "in press" and ideally with a DOI for the upcoming article, plus a preprint/eprint link if possible.

If these are not yet accepted / in press, then preprints/eprints should be cited instead.

Can you correct those two citations with this in mind? Tiggeloven et al. and Wiegner et al.
