### 2025-01-06 #7402 @JeanWi
Hi @kyleniemeyer, thanks for pointing this out.
For Wiegner et al., this is not a problem as there is a preprint available and we will change it respectively.
Tiggeloven et al. is conditionally accepted and far in the review process (we expect an answer in the next two weeks). Since it does not have a preprint available, we would suggest to wait until it is officially published so that we can cite it properly, including a DOI.

Let me know if this is acceptable or if you'd suggest an alternative approach.
