+++ docs/source/FAQ.rst
@@ -1,0 +1,9 @@
+.. _FAQ:
+
+FAQ
+=====================================
+
+**How to write a specific capex function for a new technology?**
+
+**How curtailment can be allowed for certain technology?**
+
