+++ docs/source/FAQ.rst
@@ -6,4 +6,3 @@ FAQ
 **How to write a specific capex function for a new technology?**
 
 **How curtailment can be allowed for certain technology?**
-
