+++ docs/source/FAQ.rst
@@ -1,7 +1,9 @@
 .. _FAQ:
 
+=====================================
 FAQ
 =====================================
+
 
 **How to write a specific capex function for a new technology?**
 
