### 2025-01-10 6f59d6c
"created new network classes for type"
