### 2025-01-10 8c35dcc
"read_tech_data back where it changed"
