### 2025-01-10 5b1631a
"restored create tech where needed"
