+++ adopt_net0/modelhub.py
@@ -10,7 +10,7 @@ import sys
 import datetime
 
 from .utilities import get_set_t
-from .data_management import DataHandle, read_tec_data
+from .data_management import DataHandle, create_technology_class
 from .model_construction import *
 from .result_management.read_results import add_values_to_summary
 from .utilities import get_glpk_parameters, get_gurobi_parameters
@@ -473,7 +473,7 @@ class ModelHub:
         }
         for technology in technologies:
             # read in technology data
-            tec_data = read_tec_data(
+            tec_data = create_technology_class(
                 technology,
                 self.data.data_path
                 / investment_period
