+++ .github/workflows/pin_requirements.yml
@@ -27,6 +27,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install poetry
+        pip install poetry-plugin-export
 
     - name: Write requirements.txt
       run: |
