### 2025-01-12 4653dcc
"glpk back as config.solver"
