### 2025-01-13 454ac09
Created mock function
