### 2025-01-14 #7402 @kyleniemeyer
@JeanWi that sounds fine - we can hold off until you have a DOI to cite for the second paper. Please do add the preprint info / DOI for the first.
