### 2025-01-16 13b4013
"creation of network subclasses"
