+++ adopt_net0/components/networks/genericNetworks/electricity.py
@@ -11,5 +11,13 @@ from ...utilities import link_full_resolution_to_clustered
 
 class Electricity(Network):
 
+    def __init__(self, netw_data: dict):
+        """
+        Constructor
+
+        :param dict netw_data: network data
+        """
+        super().__init__(netw_data)
+
     def fit_network_performance_performance(self):
         super(Electricity, self).fit_network_performance(self)
