+++ adopt_net0/result_management/save_results.py
@@ -3,14 +3,14 @@ from pathlib import Path
 import os
 
 from pyomo.environ import ConcreteModel
-from ..utilities import get_set_t
+from ..utilities import get_set_t, get_hour_factors, get_nr_timesteps_averaged
 
 import logging
 
 log = logging.getLogger(__name__)
 
 
-def get_summary(model, solution, folder_path: Path, model_info: dict) -> dict:
+def get_summary(model, solution, folder_path: Path, model_info: dict, data) -> dict:
     """
     Retrieves all variable values relevant for the summary of an optimization run.
 
@@ -23,6 +23,8 @@ def get_summary(model, solution, folder_path: Path, model_info: dict) -> dict:
     :return: a dictionary containing the most important model results (i.e., summary_dict)
     :rtype: dict
     """
+    config = model_info["config"]
+
     # SUMMARY: create dictionary
     summary_dict = {}
 
@@ -74,6 +76,102 @@ def get_summary(model, solution, folder_path: Path, model_info: dict) -> dict:
     )
     summary_dict["emissions_net"] = sum(
         model.periods[period].var_emissions_net.value for period in model.set_periods
+    )
+
+    # Emission balance
+    from_technologies = {}
+    from_technologies_neg = {}
+    from_carriers = {}
+    from_carriers_neg = {}
+    from_networks = {}
+    for period in model.set_periods:
+        b_period = model.periods[period]
+        set_t = get_set_t(config, b_period)
+        hour_factors = get_hour_factors(config, data, period)
+        nr_timesteps_averaged = get_nr_timesteps_averaged(config)
+
+        from_technologies[period] = sum(
+            sum(
+                sum(
+                    b_period.node_blocks[node]
+                    .tech_blocks_active[tec]
+                    .var_tec_emissions_pos[t]
+                    .value
+                    * nr_timesteps_averaged
+                    * hour_factors[t - 1]
+                    for t in set_t
+                )
+                for tec in b_period.node_blocks[node].set_technologies
+            )
+            for node in model.set_nodes
+        )
+
+        from_technologies_neg[period] = sum(
+            sum(
+                sum(
+                    b_period.node_blocks[node]
+                    .tech_blocks_active[tec]
+                    .var_tec_emissions_neg[t]
+                    .value
+                    * nr_timesteps_averaged
+                    * hour_factors[t - 1]
+                    for t in set_t
+                )
+                for tec in b_period.node_blocks[node].set_technologies
+            )
+            for node in model.set_nodes
+        )
+
+        from_carriers[period] = sum(
+            sum(
+                b_period.node_blocks[node].var_import_emissions_pos[t].value
+                * nr_timesteps_averaged
+                * hour_factors[t - 1]
+                for t in set_t
+            )
+            for node in model.set_nodes
+        )
+
+        from_carriers_neg[period] = sum(
+            sum(
+                b_period.node_blocks[node].var_import_emissions_neg[t].value
+                * nr_timesteps_averaged
+                * hour_factors[t - 1]
+                for t in set_t
+            )
+            for node in model.set_nodes
+        )
+
+        if not config["energybalance"]["copperplate"]["value"]:
+            from_networks[period] = sum(
+                sum(
+                    sum(
+                        b_period.network_block[netw].var_netw_emissions_pos[t, node]
+                        * nr_timesteps_averaged
+                        * hour_factors[t - 1]
+                        for t in set_t
+                    )
+                    for node in model.set_nodes
+                )
+                for netw in b_period.set_networks
+            )
+        else:
+            from_networks[period] = 0
+
+    summary_dict["technology_emissions_pos"] = sum(
[…]
