+++ requirements.txt
@@ -9,7 +9,7 @@ highspy==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
 idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
 joblib==1.4.2 ; python_version >= "3.12" and python_version < "4.0"
 networkx==3.4.2 ; python_version >= "3.12" and python_version < "4.0"
-numpy==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
+numpy==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
 openpyxl==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
 packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
 pandas==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
