### 2025-01-28 774c969
"New classes for network, removed emission from electricity network"
