+++ adopt_net0/components/networks/genericNetworks/fluid.py
@@ -19,5 +19,309 @@ class Fluid(Network):
         """
         super().__init__(netw_data)
 
-    def fit_network_performance_performance(self):
-        super(Fluid, self).fit_network_performance(self)
+    def fit_network_performance(self):
+        super(Fluid, self).fit_network_performance()
+
+        input_parameters = self.input_parameters
+        time_independent = {}
+
+        # Emissions
+        time_independent["loss2emissions"] = input_parameters.performance_data[
+            "loss2emissions"
+        ]
+        time_independent["emissionfactor"] = input_parameters.performance_data[
+            "emissionfactor"
+        ]
+
+    def construct_netw_model(
+        self, b_netw, data: dict, set_nodes, set_t_full, set_t_clustered
+    ):
+        super(Fluid, self).construct_netw_model(
+            b_netw, data, set_nodes, set_t_full, set_t_clustered
+        )
+
+        b_netw = self._define_emission_vars(b_netw)
+
+        """if self.component_options.energyconsumption:
+            b_netw = self._define_energyconsumption_parameters(b_netw)"""
+
+        # CONSTRAINTS FOR BIDIRECTIONAL NETWORKS
+        if self.component_options.bidirectional_network:
+            b_netw = self._define_bidirectional_constraints(b_netw)
+
+        b_netw = self._define_emission_constraints(b_netw)
+
+        """if self.component_options.energyconsumption:
+            b_netw = self._define_energyconsumption_total(b_netw)"""
+
+    '''def _define_emission_vars(self, b_netw):
+        """
+        Defines network emissions
+
+        :param b_netw: pyomo network block
+        :return: pyomo network block
+        """
+        b_netw.var_netw_emissions_pos = pyo.Var(
+            self.set_t, self.set_nodes, domain=pyo.NonNegativeReals
+        )
+
+        return b_netw'''
+
+    '''def _define_energyconsumption_parameters(self, b_netw):
+        """
+        Constructs constraints for network energy consumption
+
+        :param b_netw: pyomo network block
+        :return: pyomo network block
+        """
+        # Set of consumed carriers
+        b_netw.set_consumed_carriers = pyo.Set(
+            initialize=list(self.energy_consumption.keys())
+        )
+
+        # Parameters
+        def init_cons_send1(para, car):
+            return self.energy_consumption[car]["send"]["k_flow"]
+
+        b_netw.para_send_kflow = pyo.Param(
+            b_netw.set_consumed_carriers, domain=pyo.Reals, initialize=init_cons_send1
+        )
+
+        def init_cons_send2(para, car):
+            return self.energy_consumption[car]["send"]["k_flowDistance"]
+
+        b_netw.para_send_kflowDistance = pyo.Param(
+            b_netw.set_consumed_carriers, domain=pyo.Reals, initialize=init_cons_send2
+        )
+
+        def init_cons_receive1(para, car):
+            return self.energy_consumption[car]["receive"]["k_flow"]
+
+        b_netw.para_receive_kflow = pyo.Param(
+            b_netw.set_consumed_carriers,
+            domain=pyo.Reals,
+            initialize=init_cons_receive1,
+        )
+
+        def init_cons_receive2(para, car):
+            return self.energy_consumption[car]["receive"]["k_flowDistance"]
+
+        b_netw.para_receive_kflowDistance = pyo.Param(
+            b_netw.set_consumed_carriers,
+            domain=pyo.Reals,
+            initialize=init_cons_receive2,
+        )
+
+        # Consumption at each node
+        b_netw.var_consumption = pyo.Var(
+            self.set_t,
+            b_netw.set_consumed_carriers,
+            self.set_nodes,
+            domain=pyo.NonNegativeReals,
+        )
+
+        return b_netw'''
+
+    '''def _calculate_energy_consumption(self):
+        """
[…]

[…]
