### 2025-01-29 d4a552c
"Removed energy consumption from electricity network"
