+++ adopt_net0/components/networks/genericNetworks/fluid.py
@@ -19,65 +19,33 @@ class Fluid(Network):
         """
         super().__init__(netw_data)
 
+        self._calculate_energy_consumption()
+
     def fit_network_performance(self):
+        """
+        Fits network performance for fluid network (bounds and coefficients).
+        """
         super(Fluid, self).fit_network_performance()
 
         input_parameters = self.input_parameters
-        time_independent = {}
+        # time_independent = {}
 
         # Emissions
-        time_independent["loss2emissions"] = input_parameters.performance_data[
-            "loss2emissions"
-        ]
-        time_independent["emissionfactor"] = input_parameters.performance_data[
-            "emissionfactor"
-        ]
-
-    def construct_netw_model(
-        self, b_netw, data: dict, set_nodes, set_t_full, set_t_clustered
-    ):
-        super(Fluid, self).construct_netw_model(
-            b_netw, data, set_nodes, set_t_full, set_t_clustered
+        self.processed_coeff.time_independent["loss2emissions"] = (
+            input_parameters.performance_data["loss2emissions"]
+        )
+        self.processed_coeff.time_independent["emissionfactor"] = (
+            input_parameters.performance_data["emissionfactor"]
         )
 
-        b_netw = self._define_emission_vars(b_netw)
-
-        """if self.component_options.energyconsumption:
-            b_netw = self._define_energyconsumption_parameters(b_netw)"""
-
-        # CONSTRAINTS FOR BIDIRECTIONAL NETWORKS
-        if self.component_options.bidirectional_network:
-            b_netw = self._define_bidirectional_constraints(b_netw)
-
-        b_netw = self._define_emission_constraints(b_netw)
-
-        """if self.component_options.energyconsumption:
-            b_netw = self._define_energyconsumption_total(b_netw)"""
-
-    '''def _define_emission_vars(self, b_netw):
+    def _define_energyconsumption_parameters(self, b_netw):
         """
-        Defines network emissions
+        Constructs constraints for fluid network energy consumption
 
         :param b_netw: pyomo network block
         :return: pyomo network block
         """
-        b_netw.var_netw_emissions_pos = pyo.Var(
-            self.set_t, self.set_nodes, domain=pyo.NonNegativeReals
-        )
-
-        return b_netw'''
-
-    '''def _define_energyconsumption_parameters(self, b_netw):
-        """
-        Constructs constraints for network energy consumption
-
-        :param b_netw: pyomo network block
-        :return: pyomo network block
-        """
-        # Set of consumed carriers
-        b_netw.set_consumed_carriers = pyo.Set(
-            initialize=list(self.energy_consumption.keys())
-        )
+        super(Fluid, self)._define_energyconsumption_parameters(b_netw)
 
         # Parameters
         def init_cons_send1(para, car):
@@ -120,9 +88,9 @@ class Fluid(Network):
             domain=pyo.NonNegativeReals,
         )
 
-        return b_netw'''
+        return b_netw
 
-    '''def _calculate_energy_consumption(self):
+    def _calculate_energy_consumption(self):
         """
         Fits the performance parameters for a network, i.e. the consumption at each node.
         """
@@ -152,9 +120,9 @@ class Fluid(Network):
                 self.energy_consumption[car]["send"]["k_flowDistance"] = 0
                 self.energy_consumption[car]["receive"] = {}
                 self.energy_consumption[car]["receive"]["k_flow"] = 0
-                self.energy_consumption[car]["receive"]["k_flowDistance"] = 0'''
+                self.energy_consumption[car]["receive"]["k_flowDistance"] = 0
 
-    '''def _define_energyconsumption_arc(self, b_arc, b_netw):
+    def _define_energyconsumption_arc(self, b_arc, b_netw):
         """
         Defines the energy consumption for an arc
 
[…]
