+++ adopt_net0/model_construction/construct_nodes.py
@@ -215,8 +215,8 @@ def construct_node_block(b_node, data: dict, set_t_full, set_t_clustered):
         network_energy_consumption = determine_network_energy_consumption(
             data["network_data"]
         )
-        if network_energy_consumption:
-            b_node.var_netw_consumption = pyo.Var(set_t, b_node.set_carriers)
+
+        b_node.var_netw_consumption = pyo.Var(set_t, b_node.set_carriers)
 
     # Generic production profile
     b_node.var_generic_production = pyo.Var(
