+++ tests/test_networks.py
@@ -6,6 +6,7 @@ from adopt_net0.components.networks import Network
 from tests.utilities import make_data_for_testing, run_model
 from adopt_net0.data_preprocessing.template_creation import create_empty_network_matrix
 from adopt_net0.components.utilities import perform_disjunct_relaxation
+from adopt_net0.data_management.utilities import select_network
 
 
 def define_network(
@@ -35,7 +36,7 @@ def define_network(
     if not energyconsumption:
         netw_data["Performance"]["energyconsumption"] = {}
 
-    netw_data = Network(netw_data)
+    netw_data = select_network(netw_data)
 
     return netw_data
 

[…]
