### 2025-01-29 8dd6d6d
Cleaning and: - electricity networks are always bi-directional - removed printing in handle_input_data.py
