+++ adopt_net0/components/networks/genericNetworks/electricity.py
@@ -18,26 +18,11 @@ class Electricity(Network):
         :param dict netw_data: network data
         """
         super().__init__(netw_data)
+        # electricity networks are always bi-directional
+        self.component_options.bidirectional_network = 1
 
     def fit_network_performance(self):
         super(Electricity, self).fit_network_performance()
-
-    def _define_emissions_arc(self, b_arc, b_netw):
-        """
-        Defines emission per arc equal to zero for Electricity networks
-
-        :param b_arc: pyomo arc block
-        :param b_netw: pyomo network block
-        :return: pyomo arc block
-        """
-        super(Electricity, self)._define_emissions_arc(b_arc, b_netw)
-
-        def init_arc_emissions(const, t):
-            return b_arc.var_emissions[t] == 0
-
-        b_arc.const_arc_emissions = pyo.Constraint(self.set_t, rule=init_arc_emissions)
-
-        return b_arc
 
     def _define_emission_constraints(self, b_netw):
         """
@@ -55,63 +40,6 @@ class Electricity(Network):
             self.set_t, self.set_nodes, rule=init_netw_emissions
         )
         return b_netw
-
-    def _calculate_energy_consumption(self):
-        """
-        Fits the performance parameters for a network, i.e. the consumption at each node.
-        Overwritten in child class
-        """
-        # Get energy consumption at nodes form file
-
-        self.energy_consumption = {}
-
-    def _define_energyconsumption_arc(self, b_arc, b_netw):
-        """
-        Defines the energy consumption for an electricity arc
-
-        :param b_arc: pyomo arc block
-        :param b_netw: pyomo network block
-        :return: pyomo arc block
-        """
-        super(Electricity, self)._define_energyconsumption_arc(b_arc, b_netw)
-        rated_capacity = self.input_parameters.rated_power
-
-        b_arc.var_consumption_send = pyo.Var(
-            self.set_t,
-            b_netw.set_consumed_carriers,
-            domain=pyo.NonNegativeReals,
-            bounds=(
-                b_netw.para_size_min * rated_capacity,
-                b_arc.para_size_max * rated_capacity,
-            ),
-        )
-        b_arc.var_consumption_receive = pyo.Var(
-            self.set_t,
-            b_netw.set_consumed_carriers,
-            domain=pyo.NonNegativeReals,
-            bounds=(
-                b_netw.para_size_min * rated_capacity,
-                b_arc.para_size_max * rated_capacity,
-            ),
-        )
-
-        # Sending node
-        def init_consumption_send(const, t, car):
-            return b_arc.var_consumption_send[t, car] == 0
-
-        b_arc.const_consumption_send = pyo.Constraint(
-            self.set_t, b_netw.set_consumed_carriers, rule=init_consumption_send
-        )
-
-        # Receiving node
-        def init_consumption_receive(const, t, car):
-            return b_arc.var_consumption_receive[t, car] == 0
-
-        b_arc.const_consumption_receive = pyo.Constraint(
-            self.set_t, b_netw.set_consumed_carriers, rule=init_consumption_receive
-        )
-
-        return b_arc
 
     def _define_energyconsumption_total(self, b_netw):
         """
@@ -134,20 +62,3 @@ class Electricity(Network):
         )
 
         return b_netw
-
-    def write_results_netw_design(self, h5_group, model_block):
-        """
-        Function to report electricity network design
-
-        :param model_block: pyomo network block
-        :param h5_group: h5 group to write to
-        """
-        super(Electricity, self).write_results_netw_design(h5_group, model_block)
-
-        for arc_name in model_block.set_arcs:
-            arc = model_block.arc_block[arc_name]
-            str = "".join(arc_name)
[…]
