+++ requirements.txt
@@ -3,7 +3,7 @@ cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0"
 charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
 colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
 et-xmlfile==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
-h3==4.1.2 ; python_version >= "3.12" and python_version < "4.0"
+h3==4.2.0 ; python_version >= "3.12" and python_version < "4.0"
 h5py==3.12.1 ; python_version >= "3.12" and python_version < "4.0"
 highspy==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
 idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
@@ -29,8 +29,8 @@ scipy==1.15.1 ; python_version >= "3.12" and python_version < "4.0"
 six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
 statsmodels==0.14.4 ; python_version >= "3.12" and python_version < "4.0"
 threadpoolctl==3.5.0 ; python_version >= "3.12" and python_version < "4.0"
-timezonefinder==6.5.7 ; python_version >= "3.12" and python_version < "4.0"
+timezonefinder==6.5.8 ; python_version >= "3.12" and python_version < "4.0"
 tqdm==4.67.1 ; python_version >= "3.12" and python_version < "4.0"
 tsam==2.3.5 ; python_version >= "3.12" and python_version < "4.0"
-tzdata==2024.2 ; python_version >= "3.12" and python_version < "4.0"
+tzdata==2025.1 ; python_version >= "3.12" and python_version < "4.0"
 urllib3==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
