### 2025-02-04 a0dc254
Updated b_tec.set_output_carriers_ccs.[1] with .at(1) in technology.py and removed var_size_ccs definition that was double in _define_ccs_emissions
