### 2025-02-05 00f41d9
Started coding a new cement_hybrid_ccs.py class
