+++ main.py
@@ -1,31 +1,124 @@
 # from adopt_net0.model_configuration import ModelConfiguration
+import adopt_net0 as adopt
+import json
+import pandas as pd
 from pathlib import Path
-import adopt_net0.data_preprocessing as dp
-from adopt_net0.modelhub import ModelHub
-from adopt_net0.result_management.read_results import add_values_to_summary
+import numpy as np
+
 
 # Specify the path to your input data
-path = "specify path to input data"
+path = Path("./testCCS")
 
 # Create template files (comment these lines if already defined)
-dp.create_optimization_templates(path)
-dp.create_montecarlo_template_csv(path)
+adopt.create_optimization_templates(path)
 
-# Create folder structure (comment these lines if already defined)
-dp.create_input_data_folder_template(path)
+# Load json template
+with open(path / "Topology.json", "r") as json_file:
+    topology = json.load(json_file)
+# Nodes
+topology["nodes"] = ["industrial_cluster"]
+# Carriers:
+topology["carriers"] = ["electricity", "CO2captured", "heat", "hydrogen", "gas"]
+# Investment periods:
+topology["investment_periods"] = ["period1"]
+# Save json template
+with open(path / "Topology.json", "w") as json_file:
+    json.dump(topology, json_file, indent=4)
 
-# Copy technology and network data into folder (comment these lines if already defined)
-dp.copy_technology_data(path, "path to tec data")
-dp.copy_network_data(path, "path to network data")
+end_period = 10 * 180
 
-# Read climate data and fill carried data (comment these lines if already defined)
-dp.load_climate_data_from_api(path)
-dp.fill_carrier_data(path, value=0)
+
+# Load json template
+with open(path / "ConfigModel.json", "r") as json_file:
+    configuration = json.load(json_file)
+# Change objective
+configuration["optimization"]["objective"]["value"] = "costs"
+# Set MILP gap
+configuration["solveroptions"]["mipgap"]["value"] = 0.02
+# Save json template
+with open(path / "ConfigModel.json", "w") as json_file:
+    json.dump(configuration, json_file, indent=4)
+
+adopt.create_input_data_folder_template(path)
+
+node_location = pd.read_csv(path / "NodeLocations.csv", sep=";", index_col=0, header=0)
+node_location.at["industrial_cluster", "lon"] = 10
+node_location.at["industrial_cluster", "lat"] = 10
+node_location.at["industrial_cluster", "alt"] = 10
+node_location = node_location.reset_index()
+node_location.to_csv(path / "NodeLocations.csv", sep=";", index=False)
+
+with open(
+    path / "period1" / "node_data" / "industrial_cluster" / "Technologies.json", "r"
+) as json_file:
+    technologies = json.load(json_file)
+technologies["new"] = ["GasTurbine_simple_CCS"]
+
+with open(
+    path / "period1" / "node_data" / "industrial_cluster" / "Technologies.json", "w"
+) as json_file:
+    json.dump(technologies, json_file, indent=4)
+
+# Copy over technology files
+adopt.copy_technology_data(path)
+
+
+# Set import limits/cost
+adopt.fill_carrier_data(
+    path,
+    value_or_data=5000,
+    columns=["Import limit"],
+    carriers=["electricity"],
+    nodes=["industrial_cluster"],
+)
+adopt.fill_carrier_data(
+    path,
+    value_or_data=20000,
+    columns=["Import limit"],
+    carriers=["heat"],
+    nodes=["industrial_cluster"],
+)
+adopt.fill_carrier_data(
+    path,
+    value_or_data=80,
[…]
