+++ adopt_net0/components/technologies/specificTechnologies/cement_hybrid_ccs.py
@@ -29,3 +29,143 @@ class CementHybridCCS(Technology):
         :param dict tec_data: technology data
         """
         super().__init__(tec_data)
+        self.component_options.emissions_based_on = "output"
+        self.component_options.size_based_on = "output"
+        self.component_options.size_cement = "size_cement"
+        self.component_options.main_input_carrier = tec_data["Performance"][
+            "main_input_carrier"
+        ]
+
+    def fit_technology_performance(self, climate_data: pd.DataFrame, location: dict):
+        """
+        Fits the technology performance
+
+        :param pd.Dataframe climate_data: dataframe containing climate data
+        :param dict location: dict containing location details
+        """
+        super(CementHybridCCS, self).fit_technology_performance(climate_data, location)
+
+        # Number of segments
+        nr_segments = self.input_parameters.performance_data["nr_segments"]
+
+        # Read performance data from file
+        performance_data_path = Path(__file__).parent.parent.parent.parent
+        performance_data_path = (
+            performance_data_path
+            / "data/technology_data/Industrial/CementHybridCCS_data/dac_adsorption_performance.txt"
+        )
+
+        performance_data = pd.read_csv(performance_data_path, sep=",")
+        performance_data = performance_data.rename(
+            columns={"T": "temp_air", "RH": "humidity"}
+        )
+
+        # Unit Conversion of input data
+        performance_data.E_tot = performance_data.E_tot.multiply(
+            performance_data.CO2_Out / 3600
+        )  # in MWh / h
+        performance_data.E_el = performance_data.E_el.multiply(
+            performance_data.CO2_Out / 3600
+        )  # in MWh / h
+        performance_data.E_th = performance_data.E_th.multiply(
+            performance_data.CO2_Out / 3600
+        )  # in MWh / h
+        performance_data.CO2_Out = performance_data.CO2_Out / 1000  # in t / h
+
+        # Get humidity and temperature
+        RH = copy.deepcopy(climate_data["rh"])
+        T = copy.deepcopy(climate_data["temp_air"])
+
+        # Set minimum temperature
+        T.loc[T < min(performance_data.temp_air)] = min(performance_data.temp_air)
+
+        # Derive performance points for each timestep
+        def interpolate_performance_point(t, rh, point_data, var):
+            zi = griddata(
+                (point_data.temp_air, point_data.humidity),
+                point_data[var],
+                (T, RH),
+                method="linear",
+            )
+            return zi
+
+        CO2_Out = np.empty(shape=(len(T), len(performance_data.Point.unique())))
+        E_tot = np.empty(shape=(len(T), len(performance_data.Point.unique())))
+        E_el = np.empty(shape=(len(T), len(performance_data.Point.unique())))
+        for point in performance_data.Point.unique():
+            CO2_Out[:, point - 1] = interpolate_performance_point(
+                T, RH, performance_data.loc[performance_data.Point == point], "CO2_Out"
+            )
+            E_tot[:, point - 1] = interpolate_performance_point(
+                T, RH, performance_data.loc[performance_data.Point == point], "E_tot"
+            )
+            E_el[:, point - 1] = interpolate_performance_point(
+                T, RH, performance_data.loc[performance_data.Point == point], "E_el"
+            )
+
+        # Derive piecewise definition
+        alpha = np.empty(shape=(len(T), nr_segments))
+        beta = np.empty(shape=(len(T), nr_segments))
[…]
