### 2025-02-07 276032a
"creation of new connection class, new documentation"
