+++ adopt_net0/data_management/utilities.py
@@ -70,6 +70,8 @@ def select_technology(tec_data: dict):
         return HydroOpen(tec_data)
     elif tec_data["tec_type"] == "CCPP":
         return CCPP(tec_data)
+    elif tec_data["tec_type"] == "CementHybridCCS":
+        return CementHybridCCS(tec_data)
 
 
 def read_tec_data(tec_name: str, load_path: Path):
