+++ adopt_net0/components/technologies/specificTechnologies/cement_hybrid_ccs.py
@@ -29,11 +29,12 @@ class CementHybridCCS(Technology):
         :param dict tec_data: technology data
         """
         super().__init__(tec_data)
+
         self.component_options.emissions_based_on = "output"
         self.component_options.size_based_on = "output"
-        self.component_options.size_cement = "size_cement"
-        self.component_options.main_input_carrier = tec_data["Performance"][
-            "main_input_carrier"
+        self.component_options.prod_capacity_clinker = tec_data["prod_capacity_clinker"]
+        self.component_options.main_output_carrier = tec_data["Performance"][
+            "main_output_carrier"
         ]
 
     def fit_technology_performance(self, climate_data: pd.DataFrame, location: dict):
@@ -45,127 +46,37 @@ class CementHybridCCS(Technology):
         """
         super(CementHybridCCS, self).fit_technology_performance(climate_data, location)
 
-        # Number of segments
-        nr_segments = self.input_parameters.performance_data["nr_segments"]
-
-        # Read performance data from file
-        performance_data_path = Path(__file__).parent.parent.parent.parent
-        performance_data_path = (
-            performance_data_path
-            / "data/technology_data/Industrial/CementHybridCCS_data/dac_adsorption_performance.txt"
+        # fit coefficients
+        self.processed_coeff.time_independent["fit"] = (
+            self.fitting_class.fit_performance_function(
+                self.input_parameters.performance_data["performance"]
+            )
         )
 
-        performance_data = pd.read_csv(performance_data_path, sep=",")
-        performance_data = performance_data.rename(
-            columns={"T": "temp_air", "RH": "humidity"}
-        )
+        phi = {}
+        for car in self.input_parameters.performance_data["input_ratios"]:
+            phi[car] = self.input_parameters.performance_data["input_ratios"][car]
+        self.processed_coeff.time_independent["phi"] = phi
 
-        # Unit Conversion of input data
-        performance_data.E_tot = performance_data.E_tot.multiply(
-            performance_data.CO2_Out / 3600
-        )  # in MWh / h
-        performance_data.E_el = performance_data.E_el.multiply(
-            performance_data.CO2_Out / 3600
-        )  # in MWh / h
-        performance_data.E_th = performance_data.E_th.multiply(
-            performance_data.CO2_Out / 3600
-        )  # in MWh / h
-        performance_data.CO2_Out = performance_data.CO2_Out / 1000  # in t / h
+        def _calculate_bounds(self):
+            """
+            Calculates the bounds of the variables used
+            """
+            super(CementHybridCCS, self)._calculate_bounds()
 
-        # Get humidity and temperature
-        RH = copy.deepcopy(climate_data["rh"])
-        T = copy.deepcopy(climate_data["temp_air"])
+            time_steps = len(self.set_t_performance)
 
-        # Set minimum temperature
-        T.loc[T < min(performance_data.temp_air)] = min(performance_data.temp_air)
-
-        # Derive performance points for each timestep
-        def interpolate_performance_point(t, rh, point_data, var):
-            zi = griddata(
-                (point_data.temp_air, point_data.humidity),
-                point_data[var],
-                (T, RH),
-                method="linear",
+            self.bounds["input"] = self.fitting_class.calculate_input_bounds(
+                self.component_options.size_based_on, time_steps
             )
-            return zi
-
-        CO2_Out = np.empty(shape=(len(T), len(performance_data.Point.unique())))
-        E_tot = np.empty(shape=(len(T), len(performance_data.Point.unique())))
[…]
