+++ adopt_net0/components/technologies/specificTechnologies/__init__.py
@@ -3,3 +3,4 @@ from .gas_turbine import GasTurbine
 from .heat_pump import HeatPump
 from .hydro_open import HydroOpen
 from .combined_cycle import CCPP
+from .cement_hybrid_ccs import CementHybridCCS
