### 2025-02-10 13b3697
"added test for simple connection"
