### 2025-02-11 c98e18e
Include DOI in Tiggeloven et al.
Journal and DOI are now added to the citation, as requested
