+++ joss_submission/paper.bib
@@ -32,9 +32,9 @@
 @article{tiggeloven2024chemicalcluster,
   title={Optimizing emissions reduction in ammonia-ethylene chemical clusters: synergistic integration of electrification, carbon capture, and hydrogen},
   author={Tiggeloven, Julia L and Faaij, Andre P.C. and Kramer, Gert Jan and Gazzani, Matteo},
-  year={forthcoming},
-  journal={},
-  doi={}
+  year={2025},
+  journal={Industrial and Engineering Chemistry Research},
+  doi={https://doi.org/10.1021/acs.iecr.4c03817}
 }
 
 @article{morales2017hidden,
