### 2025-02-11 #7402 @julia1071
Hi @kyleniemeyer,
Thank you for your patience. We added the preprint information for the first paper and the DOI and journal details for the second. We’re all set to move forward with the publication.
