### 2025-02-12 b333f74
"simplified the connection network test"
