+++ tests/test_networks.py
@@ -40,6 +40,33 @@ def define_network(
     return netw_data
 
 
+def define_network_simple(
+    load_path: Path,
+    bidirectional_network: bool = False,
+):
+    """
+    reads TestNetworkSimple from path and creates network object
+
+    :param Path load_path:
+    :param bool bidirectional_network:
+    :return: Network object
+    """
+    with open(load_path / ("TestNetworkSimple.json")) as json_file:
+        netw_data_simple = json.load(json_file)
+
+    netw_data_simple["name"] = "TestNetwork"
+
+    if bidirectional_network:
+        netw_data_simple["Performance"]["bidirectional_network"] = 1
+        netw_data_simple["Performance"]["bidirectional_network_precise"] = 1
+    else:
+        netw_data_simple["Performance"]["bidirectional_network"] = 0
+
+    netw_data_simple = network_factory(netw_data_simple)
+
+    return netw_data_simple
+
+
 def construct_netw_model(
     netw,
     nr_timesteps: int,
@@ -197,41 +224,9 @@ def test_network_energyconsumption(request):
     assert m.var_consumption[1, "electricity", "node2"].value > 0
 
 
-def define_network_simple(
-    load_path: Path,
-    bidirectional_network: bool = False,
-    energyconsumption: bool = False,
-):
-    """
-    reads TestNetwork from path and creates network object
-
-    :param Path load_path:
-    :param bool bidirectional_network:
-    :param bool energyconsumption:
-    :return: Network object
-    """
-    with open(load_path / ("TestNetworkSimple.json")) as json_file:
-        netw_data_simple = json.load(json_file)
-
-    netw_data_simple["name"] = "TestNetwork"
-
-    if bidirectional_network:
-        netw_data_simple["Performance"]["bidirectional_network"] = 1
-        netw_data_simple["Performance"]["bidirectional_network_precise"] = 1
-    else:
-        netw_data_simple["Performance"]["bidirectional_network"] = 0
-
-    if not energyconsumption:
-        netw_data_simple["Performance"]["energyconsumption"] = {}
-
-    netw_data_simple = network_factory(netw_data_simple)
-
-    return netw_data_simple
-
-
 def test_network_connection(request):
     """
-    Tests a network that can only transport in one direction and its a simple connection
+    Tests a network that can only transport in one direction and it is a simple connection
 
     INFEASIBILITY CASES
     1) flow in both directions is constraint to 1
@@ -244,16 +239,15 @@ def test_network_connection(request):
     netw = define_network_simple(
         request.config.network_data_folder_path,
         bidirectional_network=True,
-        energyconsumption=False,
     )
 
     # INFEASIBILITY CASE
     m = construct_netw_model(netw, nr_timesteps)
     m.test_const_outflow1 = pyo.Constraint(
-        expr=m.var_inflow[1, "hydrogen", "node1"] == 1
+        expr=m.var_inflow[1, "electricity", "node1"] == 1
     )
     m.test_const_outflow2 = pyo.Constraint(
-        expr=m.var_inflow[1, "hydrogen", "node2"] == 1
+        expr=m.var_inflow[1, "electricity", "node2"] == 1
     )
     termination = run_model(m, request.config.solver, objective="capex")
     assert termination in [
@@ -264,7 +258,7 @@ def test_network_connection(request):
     # FEASIBILITY CASE
     m = construct_netw_model(netw, nr_timesteps)
     m.test_const_outflow1 = pyo.Constraint(
-        expr=m.var_inflow[1, "hydrogen", "node1"] == 1
+        expr=m.var_inflow[1, "electricity", "node1"] == 1
     )
 
     termination = run_model(m, request.config.solver, objective="capex")
