### 2025-02-13 7d0d494
Remove https://doi.org/ from Tiggeloven et al. ref
