+++ joss_submission/paper.bib
@@ -34,7 +34,7 @@
   author={Tiggeloven, Julia L and Faaij, Andre P.C. and Kramer, Gert Jan and Gazzani, Matteo},
   year={2025},
   journal={Industrial and Engineering Chemistry Research},
-  doi={https://doi.org/10.1021/acs.iecr.4c03817}
+  doi={10.1021/acs.iecr.4c03817}
 }
 
 @article{morales2017hidden,
