### 2025-02-13 #7402 @julia1071
@kyleniemeyer yes, the https://doi.org/ is removed.
