### 2025-02-14 #7402 @kyleniemeyer
@julia1071 can you just confirm the DOI for that upcoming article, 10.1021/acs.iecr.4c03817? I realize it may not yet be resolving because the article is still in press, but just want to check before publishing.
