### 2025-02-14 #7402 @julia1071
@kyleniemeyer yes, the DOI is correct, and it is indeed not resolving because it is still in press. Thanks!Op 14 feb 2025 om 17:40 heeft Kyle Niemeyer ***@***.***> het volgende geschreven:﻿
  @julia1071 can you just confirm the DOI for that upcoming article, 10.1021/acs.iecr.4c03817 ? I realize it may not yet be resolving because the article is still in press, but just want to check before publishing.—Reply to this email directly, view it on GitHub, or unsubscribe.You are receiving this because you were mentioned.Message ID: ***@***.***>

kyleniemeyer left a comment (openjournals/joss-reviews#7402)
@julia1071 can you just confirm the DOI for that upcoming article, 10.1021/acs.iecr.4c03817 ? I realize it may not yet be resolving because the article is still in press, but just want to check before publishing.
