### 2025-02-15 #7402 @kyleniemeyer
Congratulations @JeanWi (and @julia1071) on your article's publication in JOSS! Please consider [signing up as a reviewer](https://reviewers.joss.theoj.org/join) if you haven't already.

Many thanks to @trevorb1 and @datejada for reviewing this, and @AdamRJensen for editing it. JOSS wouldn't be possible without your efforts.
